Implemented deploy as single service changes
