+++ client-app/src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
-import { BrowserRouter as Router, Route, Routes, Navigate } from "react-router-dom";
-import DashboardComponent from "./components/OrcaDashboardComponent";
-import TempComponent from "./components/DraftOrcaDashboard";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+import DashboardComponent from "./components/OrcaDashboardComponentLegacy";
+import OrcaDashboardComponent from "./components/OrcaDashboardComponent";
 import GaussianDashboardComponent from "./components/GaussianDashboardComponent";
 import NavBarOrca from "./components/NavBar";
 import "bootstrap/dist/css/bootstrap.min.css";
@@ -12,10 +12,9 @@ const App = () => {
     <Router>
       <NavBarOrca />
       <Routes>
-        <Route path="/orca" element={<DashboardComponent />} />
-        <Route path="/temp" element={<TempComponent />} />
+        <Route path="/" element={<OrcaDashboardComponent />} />
+        <Route path="/temp" element={<DashboardComponent />} />
         <Route path="/gaussian" element={<GaussianDashboardComponent />} />
-        <Route path="/" element={<Navigate to="/orca" replace />} />
         <Route path="/about" element={<AboutComponent />} />
       </Routes>
     </Router>
