+++ client-app/src/components/NavBar.js
@@ -1,18 +1,11 @@
 import Container from "react-bootstrap/Container";
 import Nav from "react-bootstrap/Nav";
 import Navbar from "react-bootstrap/Navbar";
-//import { useState } from "react";
 import { useLocation , Link} from "react-router-dom";
 
 function NavBarOrca() {
   const location = useLocation();
-  //const [activeKey, setActiveKey] = useState(location.pathname);
-  /*
-  const handleSelect = (selectedKey) => {
-    setActiveKey(selectedKey);
-  };
-  */
-
+ 
   return (
     <Navbar
       collapseOnSelect
@@ -34,8 +27,8 @@ function NavBarOrca() {
         <Navbar.Toggle aria-controls="responsive-navbar-nav" />
         <Navbar.Collapse id="responsive-navbar-nav">
           <Nav className="me-auto"></Nav>
-          <Nav Nav activeKey={location.pathname}>
-            <Nav.Link href="/orca">ORCA Log Extraction</Nav.Link>
+          <Nav activeKey={location.pathname}>
+            <Nav.Link href="/">ORCA Log Extraction</Nav.Link>
             <Nav.Link href="/gaussian">Gaussian Log Extraction</Nav.Link>
             <Nav.Link href="/about">About</Nav.Link>
           </Nav>
