+++ client-app/src/components/OrcaDashboardComponent.js
@@ -1,21 +1,77 @@
 import React, { useState } from "react";
 import axios from "axios";
 import { saveAs } from "file-saver";
-import "../styles/DashboardComponent.css";
+import "../styles/OrcaDashboardComponent.css";
+import config from "../utils/config";
 
 const OrcaDashboardComponent = () => {
   const [selectedFile, setSelectedFile] = useState(null);
   const [filePath, setFilePath] = useState("");
-  const [searchTerms, setSearchTerms] = useState("");
-  const [specifyLines, setSpecifyLines] = useState("");
-  const [sections, setSections] = useState("");
-  const [useTotalLines, setUseTotalLines] = useState("");
-  const [totalLines, setTotalLines] = useState("");
+  const [searchTerms, setSearchTerms] = useState([]);
+  const [specifyLines, setSpecifyLines] = useState([]);
+  const [sections, setSections] = useState([]);
+  const [showCard, setShowCard] = useState(false);
+  const [uploadedFiles, setUploadedFiles] = useState([]);
+  const isUploadedFilesEmpty = uploadedFiles.length === 0;
+  const isSearchTermsEmpty = searchTerms.length === 0;
+  const isSpecifyLinesEmpty = specifyLines.length === 0;
+  const isSectionsEmpty = sections.length === 0;
+  const [sameCriteria, setSameCriteria] = useState(false);
   const [previewContent, setPreviewContent] = useState("");
+  const [showPreviewModal, setShowPreviewModal] = useState(false); 
 
   const onFileSelected = (event) => {
-    setSelectedFile(event.target.files[0]);
+    const selectedFile = event.target.files[0];
+    if (selectedFile.type !== "text/plain") {
+      alert("Invalid file type. Please upload a .txt file.");
+      return;
+    }
+    setSelectedFile(selectedFile);
   };
+
+  const isSearchQueryEnabled = () => {
+    return !isUploadedFilesEmpty && searchTerms.length && specifyLines.length && sections.length;
+  };
+
+  const handleSpecifyLineChange = (value) => {
+    setSpecifyLines([{ value, showInput: value === "FIRST" || value === "LAST" }]);
+  };
+
+  const renderSpecifyLine = () => {
+    const line = specifyLines[0] || { value: "", showInput: false };
+    return (
+      <div className="mb-2 d-flex align-items-center">
+        <select
+          className="form-select me-2"
+          value={line.value}
+          onChange={(e) => handleSpecifyLineChange(e.target.value)}>
+          <option value="SELECT">SELECT</option>
+          <option value="WHOLE">WHOLE</option>
+          <option value="FIRST">FIRST</option>
+          <option value="LAST">LAST</option>
+        </select>
+        {line.showInput && (
+          <input
+            type="text"
+            className="form-control"
+            placeholder="Enter line number"
+            value={line.lineNumber || ""}
+            onChange={(e) => {
+              const updatedLines = [{ ...line, lineNumber: e.target.value }];
+              setSpecifyLines(updatedLines);
+            }}
+          />
+        )}
+      </div>
+    );
+  };
+
+  const formatSpecifyLines = () => {
+    const line = specifyLines[0];
+    return line.value === "WHOLE" || line.value === "SELECT" 
+      ? line.value 
+      : `${line.value} ${line.lineNumber}`;
+   };
 
   const onUpload = () => {
     if (!selectedFile) {
@@ -27,14 +83,19 @@ const OrcaDashboardComponent = () => {
     formData.append("file", selectedFile);
 
     axios
-      .post("http://localhost:5001/upload", formData)
+      .post(`${config.apiBaseUrl}/upload`, formData)
       .then((response) => {
-        console.log("File uploaded successfully:", response);
+        const uploadedFileName = response.data.file_name;
         setFilePath(response.data.file_path);
+        setUploadedFiles((prevUploadedFiles) => [...prevUploadedFiles, uploadedFileName]);
       })
       .catch((error) => {
         console.error("Error uploading file:", error);
       });
+  };
+
+  const removeUploadedFile = (filePath) => {
+    setUploadedFiles((prevUploadedFiles) => prevUploadedFiles.filter((file) => file !== filePath));
   };
 
   const onSubmit = () => {
@@ -45,21 +106,13 @@ const OrcaDashboardComponent = () => {
 
     const data = {
       file_path: filePath.toString(),
-      search_terms: searchTerms.split(","),
-      sections: sections.split(","),
-      specify_lines: specifyLines.toString(),
+      search_terms: searchTerms,
+      sections: sections,
+      specify_lines: formatSpecifyLines()
     };
 
-    if (useTotalLines) {
-      data.use_total_lines = useTotalLines;
-    }
-
-    if (totalLines) {
-      data.total_lines = totalLines;
-    }
-
     axios
-      .post("http://localhost:5001/find-sections", data, {
+      .post(`${config.apiBaseUrl}/find-sections`, data, {
         responseType: "blob",
       })
       .then((response) => {
@@ -71,8 +124,59 @@ const OrcaDashboardComponent = () => {
       });
   };
 
+  const onSearchQuerySubmit = () => {
+    setShowCard(false);
+    if (!selectedFile) {
+      alert("Please select a file.");
+      return;
+    } else {
+      setShowCard(true);
+    }
+  };
+
   const downloadDocument = (blob) => {
     saveAs(blob, "output.docx");
+  };
+
+  const handleKeyPress = (e, setterFunc) => {
+    if (e.key === "Enter" || e.key === "," || e.key === "Tab") {
+      e.preventDefault();
+      const value = e.target.value.trim();
+      if (value) {
+        const values = value.split(",");
+        setterFunc((prevValue) => [...prevValue, ...values.map((val) => val.trim().toUpperCase())]);
+        e.target.value = "";
+      }
+    }
+  };
+
+  const handleBlur = (e, setterFunc) => {
+    const value = e.target.value.trim();
+    if (value) {
+      const values = value.split(",");
+      setterFunc((prevValue) => [...prevValue, ...values.map((val) => val.trim().toUpperCase())]);
+      e.target.value = "";
+    }
+  };
+
+  const removeTag = (index, setterFunc) => {
+    setterFunc((prevTerms) => {
+      const updatedTerms = [...prevTerms];
+      updatedTerms.splice(index, 1);
+      return updatedTerms;
+    });
+  };
+
+  const handleSameCriteriaChange = (e) => {
+    setSameCriteria(e.target.checked);
+  };
+
+  const handleDelete = () => {
+    setSearchTerms([]);
+    setSpecifyLines([]);
+    setSections([]);
+    setShowCard(false);
+    setSameCriteria(false);
   };
 
   const fetchDocumentPreview = () => {
@@ -83,27 +187,19 @@ const OrcaDashboardComponent = () => {
 
     const data = {
       file_path: filePath.toString(),
-      search_terms: searchTerms.split(","),
-      sections: sections.split(","),
-      specify_lines: specifyLines.toString(),
+      search_terms: searchTerms,
+      sections: sections,
+      specify_lines: formatSpecifyLines()
     };
 
-    if (useTotalLines) {
-      data.use_total_lines = useTotalLines;
-    }
-
-    if (totalLines) {
-      data.total_lines = totalLines;
-    }
-
     axios
-      .post("http://localhost:5001/preview", data)
+      .post(`${config.apiBaseUrl}/preview`, data)
       .then((response) => {
-        const documentContent = response.data.document_content;
-        setPreviewContent(documentContent);
+        setPreviewContent(response.data.document_content); 
+        setShowPreviewModal(true); 
       })
       .catch((error) => {
-        console.error("Error:", error);
+        console.error("Error fetching preview:", error);
       });
   };
 
@@ -114,7 +210,14 @@ const OrcaDashboardComponent = () => {
         <div className="mb-3 text-start">
           <span>Upload your ORCA data file</span>
           <div className="input-group">
-            <input type="file" className="form-control" onChange={onFileSelected} accept=".txt" />
+            <input
+              type="file"
+              className="form-control"
+              onChange={onFileSelected}
+              accept=".txt"
+              value=""
+              aria-label="Upload ORCA data file"
+            />
             <button className="btn btn-primary" onClick={onUpload}>
               Upload
             </button>
@@ -122,25 +225,58 @@ const OrcaDashboardComponent = () => {
         </div>
 
         <div className="mb-3 text-start">
+          <span>Uploaded Files:</span>
+          {uploadedFiles.map((file, index) => (
+            <span key={index} className="badge bg-secondary me-2 mb-2">
+              {file}
+              <button
+                type="button"
+                className="btn-close ms-1"
+                aria-label="Remove"
+                onClick={() => removeUploadedFile(file)}></button>
+            </span>
+          ))}
+        </div>
+
+        <div className="mb-3 text-start">
           <span>Enter the terms you wish to search for (txt only):</span>
-          <input
-            type="text"
-            className="form-control"
-            placeholder="E.g., CARTESIAN COORDINATES"
-            value={searchTerms}
-            onChange={(e) => setSearchTerms(e.target.value.toUpperCase())}
-          />
+          <div>
+            <input
+              type="text"
+              className="form-control"
+              placeholder="E.g., CARTESIAN COORDINATES"
+              onKeyPress={(e) => handleKeyPress(e, setSearchTerms)}
+              onBlur={(e) => handleBlur(e, setSearchTerms)}
+            />
+            {searchTerms.map((term, index) => (
+              <span
+                key={index}
+                className="badge bg-secondary me-2 mb-2"
+                onClick={() => removeTag(index, setSearchTerms)}>
+                {term}
+                <button type="button" className="btn-close ms-1" aria-label="Remove"></button>
+              </span>
+            ))}
+          </div>
+          {searchTerms.length > 1 && (
+            <div className="form-check mt-2">
+              <input
+                className="form-check-input"
+                type="checkbox"
+                id="sameCriteriaCheckbox"
+                checked={sameCriteria}
+                onChange={handleSameCriteriaChange}
+              />
+              <label className="form-check-label" htmlFor="sameCriteriaCheckbox">
+                Is the search criteria same for all search terms
+              </label>
+            </div>
+          )}
         </div>
 
         <div className="mb-3 text-start">
           <span>Enter how you want the lines specified:</span>
-          <input
-            type="text"
-            className="form-control"
-            placeholder="E.g., WHOLE, FIRST X, LAST X"
-            value={specifyLines}
-            onChange={(e) => setSpecifyLines(e.target.value.toUpperCase())}
-          />
+          {renderSpecifyLine()}
         </div>
 
         <div className="mb-3 text-start">
@@ -148,53 +284,103 @@ const OrcaDashboardComponent = () => {
           <input
             type="text"
             className="form-control"
-            placeholder="Input as number..."
-            value={sections}
-            onChange={(e) => setSections(e.target.value)}
+            placeholder="ex: 1-5 or 1,2,5"
+            value={sections.join(", ")}
+            onChange={(e) => setSections(e.target.value.split(",").map((val) => val.trim()))}
           />
         </div>
 
-        <div className="mb-3 text-start">
-          <span>Use total lines?</span>
-          <input
-            type="text"
-            className="form-control"
-            placeholder="TRUE/FALSE"
-            value={useTotalLines}
-            onChange={(e) => setUseTotalLines(e.target.value.toUpperCase())}
-          />
+        <div className="button-container">
+          <button
+            className="btn btn-primary"
+            onClick={() => onSearchQuerySubmit()}
+            disabled={
+              !isSearchQueryEnabled() ||
+              isUploadedFilesEmpty ||
+              isSearchTermsEmpty ||
+              isSpecifyLinesEmpty ||
+              isSectionsEmpty
+            }>
+            Submit Search Query
+          </button>
         </div>
 
-        <div className="mb-3 text-start">
-          <span>Total number of lines for output doc?</span>
-          <input
-            type="text"
-            className="form-control"
-            placeholder="Input as number..."
-            value={totalLines}
-            onChange={(e) => {
-              const inputValue = e.target.value;
-              setTotalLines(inputValue === "" ? "" : parseInt(inputValue));
-            }}
-          />
-        </div>
-        <button className="btn btn-primary" onClick={fetchDocumentPreview}>
-          Preview Output
-        </button>
-        <div className="buttonSpacing">
-          <button className="btn btn-primary" onClick={onSubmit}>
+        {!isUploadedFilesEmpty &&
+          !isSearchTermsEmpty &&
+          !isSpecifyLinesEmpty &&
+          !isSectionsEmpty &&
+          showCard && (
+            <div className="card mt-3">
+              <div className="card-body">
+                <h5 className="card-title">Search Query</h5>
+                <p className="card-text">Search Terms: {searchTerms.join(", ")}</p>
+                <p className="card-text">
+                  Specify Lines: {specifyLines[0].value !== "SELECT" && specifyLines[0].value}
+                  {specifyLines[0].lineNumber ? `, ${specifyLines[0].lineNumber}` : ""}
+                </p>
+                <p className="card-text">Sections: {sections.join(", ")}</p>
+                <div className="d-flex justify-content-end">
+                  <button className="btn btn-primary me-2">Edit</button>
+                  <button className="btn btn-danger" onClick={handleDelete}>
+                    Delete
+                  </button>
+                </div>
+              </div>
+            </div>
+          )}
+
+        <div className="button-container">
+          <button
+            className="btn btn-primary"
+            onClick={fetchDocumentPreview}
+              disabled={
+                !searchTerms.length ||
+                !specifyLines.length ||
+                !sections.length ||
+                !selectedFile ||
+                isUploadedFilesEmpty
+              }>
+            Preview
+          </button>
+        </div>    
+        {showPreviewModal && (
+          <div className="modal" style={{ display: "block", backgroundColor: "rgba(0, 0, 0, 0.5)" }}>
+            <div className="modal-dialog"  style={{ maxWidth: "80vw"}}>
+              <div className="modal-content">
+                <div className="modal-header">
+                  <h5 className="modal-title">Document Preview</h5>
+                  <button type="button" className="btn-close" aria-label="Close" onClick={() => setShowPreviewModal(false)}></button>
+                </div>
+                <div className="modal-body">
+                <pre>
+                    {previewContent}
+                  </pre>
+                </div>
+                <div className="modal-footer">
+                  <button className="btn btn-secondary" onClick={() => setShowPreviewModal(false)}>
+                    Close
+                  </button>
+                </div>
+              </div>
+            </div>
+          </div>
+        )}
+
+        <div className="button-container">
+          <button
+            className="btn btn-primary"
+            onClick={onSubmit}
+            disabled={
+              !searchTerms.length ||
+              !specifyLines.length ||
+              !sections.length ||
+              !selectedFile ||
+              isUploadedFilesEmpty
+            }>
             Download Output
           </button>
         </div>
 
-        {previewContent && (
-          <div className="document-preview">
-            <h3>Document Preview</h3>
-            <div className="preview-box">
-              <pre>{previewContent}</pre>
-            </div>
-          </div>
-        )}
       </div>
     </div>
   );
