+++ client-app/src/styles/OrcaDashboardComponent.css
@@ -1,0 +1,18 @@
+input::placeholder {
+  color: gray;
+  font-style: italic;
+  font-size: 12px;
+}
+span {
+  text-align: start;
+}
+
+.nav-link.active {
+  font-weight: bold;
+  color: #fff;
+  border-bottom: 1px solid #fff;
+}
+
+.button-container {
+  margin-bottom: 10px;
+}
