+++ client-app/src/styles/OrcaDashboardComponentLegacy.css
@@ -1,0 +1,196 @@
+input::placeholder {
+  color: gray;
+  font-style: italic;
+  font-size: 12px;
+}
+span {
+  text-align: start;
+}
+
+.nav-link.active {
+  font-weight: bold;
+  color: #fff;
+  border-bottom: 1px solid #fff;
+}
+
+.document-preview {
+  margin-top: 20px;
+}
+
+.preview-box {
+  background-color: #212529;
+  color: #ffffff;
+  padding: 20px;
+  border-radius: 4px;
+  width: 600px;
+  height: 800px;
+  overflow-y: auto;
+  margin: 0 auto;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.preview-box pre {
+  font-family: "Arial", sans-serif;
+  font-size: 14px;
+  line-height: 1.5;
+  white-space: pre-wrap;
+  word-wrap: break-word;
+}
+
+.buttonSpacing {
+  padding: 10px;
+}
+
+/* :host ::ng-deep .custom-ms 
+        .p-multiselect-label {
+        width: 300px !important;
+    }
+
+    input[type="file"]{
+        padding: 0.5rem;
+        cursor: pointer;
+        font-size: 1rem;
+    }
+
+    input[type="file"]::file-selector-button {
+        padding: 0.5rem 1rem;
+        border: 1px solid #ccc;
+        border-radius: 4px;
+        background-color: #f0f0f0;
+        cursor: pointer;
+        transition: background-color 0.3s ease;
+        }
+        
+        input[type="file"]::file-selector-button:hover {
+        background-color: #e0e0e0;
+        }
+
+    #whole {
+        font-family: 'Didot';
+        font-size: 20px;
+        text-align: center;
+        padding: 80px 0;
+        position: relative;
+    }
+    
+    .form-label {
+        border: 1px solid #ccc;
+        border-radius: 5px;
+        display: inline-block;
+        padding: 8px 15px;
+        cursor: pointer;
+        font-size: 18px;
+        transition-duration: 0.7s;
+    }
+
+    .form-label:hover {
+        background-color: black;
+        color:white;
+    }
+
+    #keyTerms-toSearch {
+        border: 1px solid #ccc;
+        border-radius: 5px;
+        padding: 8px 12px;
+        width: 210x;
+        box-sizing: border-box;
+        font-size: 18px;
+    }
+
+    #linesSpecified-toSearch {
+        border: 1px solid #ccc;
+        border-radius: 5px;
+        padding: 8px 12px;
+        width: 210x;
+        box-sizing: border-box;
+        font-size: 18px;
+    }
+
+    #sections-toSearch {
+        border: 1px solid #ccc;
+        border-radius: 5px;
+        padding: 8px 12px;
+        width: 210x;
+        box-sizing: border-box;
+        font-size: 18px;
+    }
+
+    #useTotalLines-toSearch {
+        border: 1px solid #ccc;
+        border-radius: 5px;
+        padding: 8px 12px;
+        width: 210x;
+        box-sizing: border-box;
+        font-size: 18px;
+    }
+
+    #totalLines-toSearch {
+        border: 1px solid #ccc;
+        border-radius: 5px;
+        padding: 8px 12px;
+        width: 210x;
+        box-sizing: border-box;
+        font-size: 18px;
+    }
+
+    input::placeholder {
+        color: gray;
+        font-family:'Courier New', Courier, monospace;
+        font-style: italic;
+        font-size: 12px;
+    }
+
+
+    #fileNameInput {
+        border: 1px solid #ccc;
+        border-radius: 5px;
+        padding: 8px 12px;
+        width: 210x;
+        box-sizing: border-box;
+        font-size: 18px;
+    }
+
+    #fileNameInput:focus {
+        border: 2px solid #555;
+    }
+
+    #downloadButton {
+        font-family: 'Calibri';
+        font-size: 18px;
+        color: black;
+        background-color: white;
+        border: 1px solid #ccc;
+        border-radius: 5px;
+        margin-top: 25px;
+        padding: 12px;
+    }
+
+    #downloadButton:hover,.Submit:hover {
+        background-color: black;
+        color:white;
+        transform: scale(1.1);
+        transition: all 0.5s ease-in;
+        cursor: pointer;
+    }
+
+    .Submit {
+        border: none;
+        border-radius: 5px;
+        padding: 0.5rem 1rem;
+    }
+
+    .form-control:focus {
+        box-shadow: none;
+        border-color: #86b7fe;
+      }
+      
+      .btn-primary {
+        background-color: #0d6efd;
+        border-color: #0d6efd;
+      }
+      
+      .btn-primary:hover {
+        background-color: #0b5ed7;
+        border-color: #0a58ca;
+      }
+       */
