--- client-app/src/styles/TempComponent.css
@@ -1,18 +1,0 @@
-input::placeholder {
-  color: gray;
-  font-style: italic;
-  font-size: 12px;
-}
-span {
-  text-align: start;
-}
-
-.nav-link.active {
-  font-weight: bold;
-  color: #fff;
-  border-bottom: 1px solid #fff;
-}
-
-.button-container {
-  margin-bottom: 10px;
-}
