+++ client-app/src/utils/config.js
@@ -1,0 +1,5 @@
+const config = {
+  apiBaseUrl: process.env.REACT_APP_API_BASE_URL || "http://127.0.0.1:5001",
+};
+
+export default config;
