+++ server/application/app.py
@@ -4,6 +4,7 @@ This module handles the creation of Flask app
 from flask import Flask
 from application.rest import upload_files
 from application.rest import search_orca_data
+from application.rest import start_application
 
 def create_app(config_name):
     '''
@@ -12,7 +13,9 @@ def create_app(config_name):
     app = Flask(__name__)
     config_module = f"application.config.{config_name.capitalize()}Config"
     app.config.from_object(config_module)
+
     app.register_blueprint(upload_files.blueprint)
     app.register_blueprint(search_orca_data.blueprint)
+    app.register_blueprint(start_application.blueprint)
 
     return app
