+++ server/application/static/asset-manifest.json
@@ -1,0 +1,15 @@
+{
+  "files": {
+    "main.css": "/static/css/main.181c7707.css",
+    "main.js": "/static/js/main.515ed76f.js",
+    "static/js/453.a5059bd1.chunk.js": "/static/js/453.a5059bd1.chunk.js",
+    "index.html": "/index.html",
+    "main.181c7707.css.map": "/static/css/main.181c7707.css.map",
+    "main.515ed76f.js.map": "/static/js/main.515ed76f.js.map",
+    "453.a5059bd1.chunk.js.map": "/static/js/453.a5059bd1.chunk.js.map"
+  },
+  "entrypoints": [
+    "static/css/main.181c7707.css",
+    "static/js/main.515ed76f.js"
+  ]
+}
