+++ server/application/static/static/js/main.515ed76f.js
@@ -1,0 +1,3 @@
+/*! For license information please see main.515ed76f.js.LICENSE.txt */
+(()=>{var e={450:function(e,t,n){var r,a,o;a=[],void 0===(o="function"===typeof(r=function(){"use strict";function t(e,t){return"undefined"==typeof t?t={autoBom:!1}:"object"!=typeof t&&(console.warn("Deprecated: Expected third argument to be a object"),t={autoBom:!t}),t.autoBom&&/^\s*(?:text\/\S*|application\/xml|\S*\/\S*\+xml)\s*;.*charset\s*=\s*utf-8/i.test(e.type)?new Blob(["\ufeff",e],{type:e.type}):e}function r(e,t,n){var r=new XMLHttpRequest;r.open("GET",e),r.responseType="blob",r.onload=function(){s(r.response,t,n)},r.onerror=function(){console.error("could not download file")},r.send()}function a(e){var t=new XMLHttpRequest;t.open("HEAD",e,!1);try{t.send()}catch(e){}return 200<=t.status&&299>=t.status}function o(e){try{e.dispatchEvent(new MouseEvent("click"))}catch(r){var t=document.createEvent("MouseEvents");t.initMouseEvent("click",!0,!0,window,0,0,0,80,20,!1,!1,!1,!1,0,null),e.dispatchEvent(t)}}var l="object"==typeof window&&window.window===window?window:"object"==typeof self&&self.self===self?self:"object"==typeof n.g&&n.g.global===n.g?n.g:void 0,i=l.navigator&&/Macintosh/.test(navigator.userAgent)&&/AppleWebKit/.test(navigator.userAgent)&&!/Safari/.test(navigator.userAgent),s=l.saveAs||("object"!=typeof window||window!==l?function(){}:"download"in HTMLAnchorElement.prototype&&!i?function(e,t,n){var i=l.URL||l.webkitURL,s=document.createElement("a");t=t||e.name||"download",s.download=t,s.rel="noopener","string"==typeof e?(s.href=e,s.origin===location.origin?o(s):a(s.href)?r(e,t,n):o(s,s.target="_blank")):(s.href=i.createObjectURL(e),setTimeout((function(){i.revokeObjectURL(s.href)}),4e4),setTimeout((function(){o(s)}),0))}:"msSaveOrOpenBlob"in navigator?function(e,n,l){if(n=n||e.name||"download","string"!=typeof e)navigator.msSaveOrOpenBlob(t(e,l),n);else if(a(e))r(e,n,l);else{var i=document.createElement("a");i.href=e,i.target="_blank",setTimeout((function(){o(i)}))}}:function(e,t,n,a){if((a=a||open("","_blank"))&&(a.document.title=a.document.body.innerText="downloading..."),"string"==typeof e)return r(e,t,n);var o="application/octet-stream"===e.type,s=/constructor/i.test(l.HTMLElement)||l.safari,u=/CriOS\/[\d]+/.test(navigator.userAgent);if((u||o&&s||i)&&"undefined"!=typeof FileReader){var c=new FileReader;c.onloadend=function(){var e=c.result;e=u?e:e.replace(/^data:[^;]*;/,"data:attachment/file;"),a?a.location.href=e:location=e,a=null},c.readAsDataURL(e)}else{var d=l.URL||l.webkitURL,f=d.createObjectURL(e);a?a.location=f:location.href=f,a=null,setTimeout((function(){d.revokeObjectURL(f)}),4e4)}});l.saveAs=s.saveAs=s,e.exports=s})?r.apply(t,a):r)||(e.exports=o)},740:e=>{"use strict";e.exports=function(e,t,n,r,a,o,l,i){if(!e){var s;if(void 0===t)s=new Error("Minified exception occurred; use the non-minified dev environment for the full error message and additional helpful warnings.");else{var u=[n,r,a,o,l,i],c=0;(s=new Error(t.replace(/%s/g,(function(){return u[c++]})))).name="Invariant Violation"}throw s.framesToPop=1,s}}},197:(e,t,n)=>{"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(){for(var e=arguments.length,t=Array(e),n=0;n<e;n++)t[n]=arguments[n];return(0,o.default)((function(){for(var e=arguments.length,n=Array(e),r=0;r<e;r++)n[r]=arguments[r];var a=null;return t.forEach((function(e){if(null==a){var t=e.apply(void 0,n);null!=t&&(a=t)}})),a}))};var r,a=n(534),o=(r=a)&&r.__esModule?r:{default:r};e.exports=t.default},534:(e,t)=>{"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(e){function t(t,n,r,a,o,l){var i=a||"<<anonymous>>",s=l||r;if(null==n[r])return t?new Error("Required "+o+" `"+s+"` was not specified in `"+i+"`."):null;for(var u=arguments.length,c=Array(u>6?u-6:0),d=6;d<u;d++)c[d-6]=arguments[d];return e.apply(void 0,[n,r,i,o,s].concat(c))}var n=t.bind(null,!1);return n.isRequired=t.bind(null,!0),n},e.exports=t.default},497:(e,t,n)=>{"use strict";var r=n(218);function a(){}function o(){}o.resetWarningCache=a,e.exports=function(){function e(e,t,n,a,o,l){if(l!==r){var i=new Error("Calling PropTypes validators directly is not supported by the `prop-types` package. Use PropTypes.checkPropTypes() to call them. Read more at http://fb.me/use-check-prop-types");throw i.name="Invariant Violation",i}}function t(){return e}e.isRequired=e;var n={array:e,bigint:e,bool:e,func:e,number:e,object:e,string:e,symbol:e,any:e,arrayOf:t,element:e,elementType:e,instanceOf:t,node:e,objectOf:t,oneOf:t,oneOfType:t,shape:t,exact:t,checkPropTypes:o,resetWarningCache:a};return n.PropTypes=n,n}},173:(e,t,n)=>{e.exports=n(497)()},218:e=>{"use strict";e.exports="SECRET_DO_NOT_PASS_THIS_OR_YOU_WILL_BE_FIRED"},730:(e,t,n)=>{"use strict";var r=n(43),a=n(853);function o(e){for(var t="https://reactjs.org/docs/error-decoder.html?invariant="+e,n=1;n<arguments.length;n++)t+="&args[]="+encodeURIComponent(arguments[n]);return"Minified React error #"+e+"; visit "+t+" for the full message or use the non-minified dev environment for full errors and additional helpful warnings."}var l=new Set,i={};function s(e,t){u(e,t),u(e+"Capture",t)}function u(e,t){for(i[e]=t,e=0;e<t.length;e++)l.add(t[e])}var c=!("undefined"===typeof window||"undefined"===typeof window.document||"undefined"===typeof window.document.createElement),d=Object.prototype.hasOwnProperty,f=/^[:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD][:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD\-.0-9\u00B7\u0300-\u036F\u203F-\u2040]*$/,p={},h={};function m(e,t,n,r,a,o,l){this.acceptsBooleans=2===t||3===t||4===t,this.attributeName=r,this.attributeNamespace=a,this.mustUseProperty=n,this.propertyName=e,this.type=t,this.sanitizeURL=o,this.removeEmptyString=l}var v={};"children dangerouslySetInnerHTML defaultValue defaultChecked innerHTML suppressContentEditableWarning suppressHydrationWarning style".split(" ").forEach((function(e){v[e]=new m(e,0,!1,e,null,!1,!1)})),[["acceptCharset","accept-charset"],["className","class"],["htmlFor","for"],["httpEquiv","http-equiv"]].forEach((function(e){var t=e[0];v[t]=new m(t,1,!1,e[1],null,!1,!1)})),["contentEditable","draggable","spellCheck","value"].forEach((function(e){v[e]=new m(e,2,!1,e.toLowerCase(),null,!1,!1)})),["autoReverse","externalResourcesRequired","focusable","preserveAlpha"].forEach((function(e){v[e]=new m(e,2,!1,e,null,!1,!1)})),"allowFullScreen async autoFocus autoPlay controls default defer disabled disablePictureInPicture disableRemotePlayback formNoValidate hidden loop noModule noValidate open playsInline readOnly required reversed scoped seamless itemScope".split(" ").forEach((function(e){v[e]=new m(e,3,!1,e.toLowerCase(),null,!1,!1)})),["checked","multiple","muted","selected"].forEach((function(e){v[e]=new m(e,3,!0,e,null,!1,!1)})),["capture","download"].forEach((function(e){v[e]=new m(e,4,!1,e,null,!1,!1)})),["cols","rows","size","span"].forEach((function(e){v[e]=new m(e,6,!1,e,null,!1,!1)})),["rowSpan","start"].forEach((function(e){v[e]=new m(e,5,!1,e.toLowerCase(),null,!1,!1)}));var g=/[\-:]([a-z])/g;function y(e){return e[1].toUpperCase()}function b(e,t,n,r){var a=v.hasOwnProperty(t)?v[t]:null;(null!==a?0!==a.type:r||!(2<t.length)||"o"!==t[0]&&"O"!==t[0]||"n"!==t[1]&&"N"!==t[1])&&(function(e,t,n,r){if(null===t||"undefined"===typeof t||function(e,t,n,r){if(null!==n&&0===n.type)return!1;switch(typeof t){case"function":case"symbol":return!0;case"boolean":return!r&&(null!==n?!n.acceptsBooleans:"data-"!==(e=e.toLowerCase().slice(0,5))&&"aria-"!==e);default:return!1}}(e,t,n,r))return!0;if(r)return!1;if(null!==n)switch(n.type){case 3:return!t;case 4:return!1===t;case 5:return isNaN(t);case 6:return isNaN(t)||1>t}return!1}(t,n,a,r)&&(n=null),r||null===a?function(e){return!!d.call(h,e)||!d.call(p,e)&&(f.test(e)?h[e]=!0:(p[e]=!0,!1))}(t)&&(null===n?e.removeAttribute(t):e.setAttribute(t,""+n)):a.mustUseProperty?e[a.propertyName]=null===n?3!==a.type&&"":n:(t=a.attributeName,r=a.attributeNamespace,null===n?e.removeAttribute(t):(n=3===(a=a.type)||4===a&&!0===n?"":""+n,r?e.setAttributeNS(r,t,n):e.setAttribute(t,n))))}"accent-height alignment-baseline arabic-form baseline-shift cap-height clip-path clip-rule color-interpolation color-interpolation-filters color-profile color-rendering dominant-baseline enable-background fill-opacity fill-rule flood-color flood-opacity font-family font-size font-size-adjust font-stretch font-style font-variant font-weight glyph-name glyph-orientation-horizontal glyph-orientation-vertical horiz-adv-x horiz-origin-x image-rendering letter-spacing lighting-color marker-end marker-mid marker-start overline-position overline-thickness paint-order panose-1 pointer-events rendering-intent shape-rendering stop-color stop-opacity strikethrough-position strikethrough-thickness stroke-dasharray stroke-dashoffset stroke-linecap stroke-linejoin stroke-miterlimit stroke-opacity stroke-width text-anchor text-decoration text-rendering underline-position underline-thickness unicode-bidi unicode-range units-per-em v-alphabetic v-hanging v-ideographic v-mathematical vector-effect vert-adv-y vert-origin-x vert-origin-y word-spacing writing-mode xmlns:xlink x-height".split(" ").forEach((function(e){var t=e.replace(g,y);v[t]=new m(t,1,!1,e,null,!1,!1)})),"xlink:actuate xlink:arcrole xlink:role xlink:show xlink:title xlink:type".split(" ").forEach((function(e){var t=e.replace(g,y);v[t]=new m(t,1,!1,e,"http://www.w3.org/1999/xlink",!1,!1)})),["xml:base","xml:lang","xml:space"].forEach((function(e){var t=e.replace(g,y);v[t]=new m(t,1,!1,e,"http://www.w3.org/XML/1998/namespace",!1,!1)})),["tabIndex","crossOrigin"].forEach((function(e){v[e]=new m(e,1,!1,e.toLowerCase(),null,!1,!1)})),v.xlinkHref=new m("xlinkHref",1,!1,"xlink:href","http://www.w3.org/1999/xlink",!0,!1),["src","href","action","formAction"].forEach((function(e){v[e]=new m(e,1,!1,e.toLowerCase(),null,!0,!0)}));var w=r.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,x=Symbol.for("react.element"),E=Symbol.for("react.portal"),S=Symbol.for("react.fragment"),k=Symbol.for("react.strict_mode"),C=Symbol.for("react.profiler"),N=Symbol.for("react.provider"),O=Symbol.for("react.context"),R=Symbol.for("react.forward_ref"),j=Symbol.for("react.suspense"),_=Symbol.for("react.suspense_list"),T=Symbol.for("react.memo"),P=Symbol.for("react.lazy");Symbol.for("react.scope"),Symbol.for("react.debug_trace_mode");var L=Symbol.for("react.offscreen");Symbol.for("react.legacy_hidden"),Symbol.for("react.cache"),Symbol.for("react.tracing_marker");var A=Symbol.iterator;function D(e){return null===e||"object"!==typeof e?null:"function"===typeof(e=A&&e[A]||e["@@iterator"])?e:null}var F,U=Object.assign;function z(e){if(void 0===F)try{throw Error()}catch(n){var t=n.stack.trim().match(/\n( *(at )?)/);F=t&&t[1]||""}return"\n"+F+e}var I=!1;function M(e,t){if(!e||I)return"";I=!0;var n=Error.prepareStackTrace;Error.prepareStackTrace=void 0;try{if(t)if(t=function(){throw Error()},Object.defineProperty(t.prototype,"props",{set:function(){throw Error()}}),"object"===typeof Reflect&&Reflect.construct){try{Reflect.construct(t,[])}catch(u){var r=u}Reflect.construct(e,[],t)}else{try{t.call()}catch(u){r=u}e.call(t.prototype)}else{try{throw Error()}catch(u){r=u}e()}}catch(u){if(u&&r&&"string"===typeof u.stack){for(var a=u.stack.split("\n"),o=r.stack.split("\n"),l=a.length-1,i=o.length-1;1<=l&&0<=i&&a[l]!==o[i];)i--;for(;1<=l&&0<=i;l--,i--)if(a[l]!==o[i]){if(1!==l||1!==i)do{if(l--,0>--i||a[l]!==o[i]){var s="\n"+a[l].replace(" at new "," at ");return e.displayName&&s.includes("<anonymous>")&&(s=s.replace("<anonymous>",e.displayName)),s}}while(1<=l&&0<=i);break}}}finally{I=!1,Error.prepareStackTrace=n}return(e=e?e.displayName||e.name:"")?z(e):""}function B(e){switch(e.tag){case 5:return z(e.type);case 16:return z("Lazy");case 13:return z("Suspense");case 19:return z("SuspenseList");case 0:case 2:case 15:return e=M(e.type,!1);case 11:return e=M(e.type.render,!1);case 1:return e=M(e.type,!0);default:return""}}function $(e){if(null==e)return null;if("function"===typeof e)return e.displayName||e.name||null;if("string"===typeof e)return e;switch(e){case S:return"Fragment";case E:return"Portal";case C:return"Profiler";case k:return"StrictMode";case j:return"Suspense";case _:return"SuspenseList"}if("object"===typeof e)switch(e.$$typeof){case O:return(e.displayName||"Context")+".Consumer";case N:return(e._context.displayName||"Context")+".Provider";case R:var t=e.render;return(e=e.displayName)||(e=""!==(e=t.displayName||t.name||"")?"ForwardRef("+e+")":"ForwardRef"),e;case T:return null!==(t=e.displayName||null)?t:$(e.type)||"Memo";case P:t=e._payload,e=e._init;try{return $(e(t))}catch(n){}}return null}function W(e){var t=e.type;switch(e.tag){case 24:return"Cache";case 9:return(t.displayName||"Context")+".Consumer";case 10:return(t._context.displayName||"Context")+".Provider";case 18:return"DehydratedFragment";case 11:return e=(e=t.render).displayName||e.name||"",t.displayName||(""!==e?"ForwardRef("+e+")":"ForwardRef");case 7:return"Fragment";case 5:return t;case 4:return"Portal";case 3:return"Root";case 6:return"Text";case 16:return $(t);case 8:return t===k?"StrictMode":"Mode";case 22:return"Offscreen";case 12:return"Profiler";case 21:return"Scope";case 13:return"Suspense";case 19:return"SuspenseList";case 25:return"TracingMarker";case 1:case 0:case 17:case 2:case 14:case 15:if("function"===typeof t)return t.displayName||t.name||null;if("string"===typeof t)return t}return null}function H(e){switch(typeof e){case"boolean":case"number":case"string":case"undefined":case"object":return e;default:return""}}function V(e){var t=e.type;return(e=e.nodeName)&&"input"===e.toLowerCase()&&("checkbox"===t||"radio"===t)}function q(e){e._valueTracker||(e._valueTracker=function(e){var t=V(e)?"checked":"value",n=Object.getOwnPropertyDescriptor(e.constructor.prototype,t),r=""+e[t];if(!e.hasOwnProperty(t)&&"undefined"!==typeof n&&"function"===typeof n.get&&"function"===typeof n.set){var a=n.get,o=n.set;return Object.defineProperty(e,t,{configurable:!0,get:function(){return a.call(this)},set:function(e){r=""+e,o.call(this,e)}}),Object.defineProperty(e,t,{enumerable:n.enumerable}),{getValue:function(){return r},setValue:function(e){r=""+e},stopTracking:function(){e._valueTracker=null,delete e[t]}}}}(e))}function K(e){if(!e)return!1;var t=e._valueTracker;if(!t)return!0;var n=t.getValue(),r="";return e&&(r=V(e)?e.checked?"true":"false":e.value),(e=r)!==n&&(t.setValue(e),!0)}function Q(e){if("undefined"===typeof(e=e||("undefined"!==typeof document?document:void 0)))return null;try{return e.activeElement||e.body}catch(t){return e.body}}function X(e,t){var n=t.checked;return U({},t,{defaultChecked:void 0,defaultValue:void 0,value:void 0,checked:null!=n?n:e._wrapperState.initialChecked})}function G(e,t){var n=null==t.defaultValue?"":t.defaultValue,r=null!=t.checked?t.checked:t.defaultChecked;n=H(null!=t.value?t.value:n),e._wrapperState={initialChecked:r,initialValue:n,controlled:"checkbox"===t.type||"radio"===t.type?null!=t.checked:null!=t.value}}function J(e,t){null!=(t=t.checked)&&b(e,"checked",t,!1)}function Y(e,t){J(e,t);var n=H(t.value),r=t.type;if(null!=n)"number"===r?(0===n&&""===e.value||e.value!=n)&&(e.value=""+n):e.value!==""+n&&(e.value=""+n);else if("submit"===r||"reset"===r)return void e.removeAttribute("value");t.hasOwnProperty("value")?ee(e,t.type,n):t.hasOwnProperty("defaultValue")&&ee(e,t.type,H(t.defaultValue)),null==t.checked&&null!=t.defaultChecked&&(e.defaultChecked=!!t.defaultChecked)}function Z(e,t,n){if(t.hasOwnProperty("value")||t.hasOwnProperty("defaultValue")){var r=t.type;if(!("submit"!==r&&"reset"!==r||void 0!==t.value&&null!==t.value))return;t=""+e._wrapperState.initialValue,n||t===e.value||(e.value=t),e.defaultValue=t}""!==(n=e.name)&&(e.name=""),e.defaultChecked=!!e._wrapperState.initialChecked,""!==n&&(e.name=n)}function ee(e,t,n){"number"===t&&Q(e.ownerDocument)===e||(null==n?e.defaultValue=""+e._wrapperState.initialValue:e.defaultValue!==""+n&&(e.defaultValue=""+n))}var te=Array.isArray;function ne(e,t,n,r){if(e=e.options,t){t={};for(var a=0;a<n.length;a++)t["$"+n[a]]=!0;for(n=0;n<e.length;n++)a=t.hasOwnProperty("$"+e[n].value),e[n].selected!==a&&(e[n].selected=a),a&&r&&(e[n].defaultSelected=!0)}else{for(n=""+H(n),t=null,a=0;a<e.length;a++){if(e[a].value===n)return e[a].selected=!0,void(r&&(e[a].defaultSelected=!0));null!==t||e[a].disabled||(t=e[a])}null!==t&&(t.selected=!0)}}function re(e,t){if(null!=t.dangerouslySetInnerHTML)throw Error(o(91));return U({},t,{value:void 0,defaultValue:void 0,children:""+e._wrapperState.initialValue})}function ae(e,t){var n=t.value;if(null==n){if(n=t.children,t=t.defaultValue,null!=n){if(null!=t)throw Error(o(92));if(te(n)){if(1<n.length)throw Error(o(93));n=n[0]}t=n}null==t&&(t=""),n=t}e._wrapperState={initialValue:H(n)}}function oe(e,t){var n=H(t.value),r=H(t.defaultValue);null!=n&&((n=""+n)!==e.value&&(e.value=n),null==t.defaultValue&&e.defaultValue!==n&&(e.defaultValue=n)),null!=r&&(e.defaultValue=""+r)}function le(e){var t=e.textContent;t===e._wrapperState.initialValue&&""!==t&&null!==t&&(e.value=t)}function ie(e){switch(e){case"svg":return"http://www.w3.org/2000/svg";case"math":return"http://www.w3.org/1998/Math/MathML";default:return"http://www.w3.org/1999/xhtml"}}function se(e,t){return null==e||"http://www.w3.org/1999/xhtml"===e?ie(t):"http://www.w3.org/2000/svg"===e&&"foreignObject"===t?"http://www.w3.org/1999/xhtml":e}var ue,ce,de=(ce=function(e,t){if("http://www.w3.org/2000/svg"!==e.namespaceURI||"innerHTML"in e)e.innerHTML=t;else{for((ue=ue||document.createElement("div")).innerHTML="<svg>"+t.valueOf().toString()+"</svg>",t=ue.firstChild;e.firstChild;)e.removeChild(e.firstChild);for(;t.firstChild;)e.appendChild(t.firstChild)}},"undefined"!==typeof MSApp&&MSApp.execUnsafeLocalFunction?function(e,t,n,r){MSApp.execUnsafeLocalFunction((function(){return ce(e,t)}))}:ce);function fe(e,t){if(t){var n=e.firstChild;if(n&&n===e.lastChild&&3===n.nodeType)return void(n.nodeValue=t)}e.textContent=t}var pe={animationIterationCount:!0,aspectRatio:!0,borderImageOutset:!0,borderImageSlice:!0,borderImageWidth:!0,boxFlex:!0,boxFlexGroup:!0,boxOrdinalGroup:!0,columnCount:!0,columns:!0,flex:!0,flexGrow:!0,flexPositive:!0,flexShrink:!0,flexNegative:!0,flexOrder:!0,gridArea:!0,gridRow:!0,gridRowEnd:!0,gridRowSpan:!0,gridRowStart:!0,gridColumn:!0,gridColumnEnd:!0,gridColumnSpan:!0,gridColumnStart:!0,fontWeight:!0,lineClamp:!0,lineHeight:!0,opacity:!0,order:!0,orphans:!0,tabSize:!0,widows:!0,zIndex:!0,zoom:!0,fillOpacity:!0,floodOpacity:!0,stopOpacity:!0,strokeDasharray:!0,strokeDashoffset:!0,strokeMiterlimit:!0,strokeOpacity:!0,strokeWidth:!0},he=["Webkit","ms","Moz","O"];function me(e,t,n){return null==t||"boolean"===typeof t||""===t?"":n||"number"!==typeof t||0===t||pe.hasOwnProperty(e)&&pe[e]?(""+t).trim():t+"px"}function ve(e,t){for(var n in e=e.style,t)if(t.hasOwnProperty(n)){var r=0===n.indexOf("--"),a=me(n,t[n],r);"float"===n&&(n="cssFloat"),r?e.setProperty(n,a):e[n]=a}}Object.keys(pe).forEach((function(e){he.forEach((function(t){t=t+e.charAt(0).toUpperCase()+e.substring(1),pe[t]=pe[e]}))}));var ge=U({menuitem:!0},{area:!0,base:!0,br:!0,col:!0,embed:!0,hr:!0,img:!0,input:!0,keygen:!0,link:!0,meta:!0,param:!0,source:!0,track:!0,wbr:!0});function ye(e,t){if(t){if(ge[e]&&(null!=t.children||null!=t.dangerouslySetInnerHTML))throw Error(o(137,e));if(null!=t.dangerouslySetInnerHTML){if(null!=t.children)throw Error(o(60));if("object"!==typeof t.dangerouslySetInnerHTML||!("__html"in t.dangerouslySetInnerHTML))throw Error(o(61))}if(null!=t.style&&"object"!==typeof t.style)throw Error(o(62))}}function be(e,t){if(-1===e.indexOf("-"))return"string"===typeof t.is;switch(e){case"annotation-xml":case"color-profile":case"font-face":case"font-face-src":case"font-face-uri":case"font-face-format":case"font-face-name":case"missing-glyph":return!1;default:return!0}}var we=null;function xe(e){return(e=e.target||e.srcElement||window).correspondingUseElement&&(e=e.correspondingUseElement),3===e.nodeType?e.parentNode:e}var Ee=null,Se=null,ke=null;function Ce(e){if(e=ba(e)){if("function"!==typeof Ee)throw Error(o(280));var t=e.stateNode;t&&(t=xa(t),Ee(e.stateNode,e.type,t))}}function Ne(e){Se?ke?ke.push(e):ke=[e]:Se=e}function Oe(){if(Se){var e=Se,t=ke;if(ke=Se=null,Ce(e),t)for(e=0;e<t.length;e++)Ce(t[e])}}function Re(e,t){return e(t)}function je(){}var _e=!1;function Te(e,t,n){if(_e)return e(t,n);_e=!0;try{return Re(e,t,n)}finally{_e=!1,(null!==Se||null!==ke)&&(je(),Oe())}}function Pe(e,t){var n=e.stateNode;if(null===n)return null;var r=xa(n);if(null===r)return null;n=r[t];e:switch(t){case"onClick":case"onClickCapture":case"onDoubleClick":case"onDoubleClickCapture":case"onMouseDown":case"onMouseDownCapture":case"onMouseMove":case"onMouseMoveCapture":case"onMouseUp":case"onMouseUpCapture":case"onMouseEnter":(r=!r.disabled)||(r=!("button"===(e=e.type)||"input"===e||"select"===e||"textarea"===e)),e=!r;break e;default:e=!1}if(e)return null;if(n&&"function"!==typeof n)throw Error(o(231,t,typeof n));return n}var Le=!1;if(c)try{var Ae={};Object.defineProperty(Ae,"passive",{get:function(){Le=!0}}),window.addEventListener("test",Ae,Ae),window.removeEventListener("test",Ae,Ae)}catch(ce){Le=!1}function De(e,t,n,r,a,o,l,i,s){var u=Array.prototype.slice.call(arguments,3);try{t.apply(n,u)}catch(c){this.onError(c)}}var Fe=!1,Ue=null,ze=!1,Ie=null,Me={onError:function(e){Fe=!0,Ue=e}};function Be(e,t,n,r,a,o,l,i,s){Fe=!1,Ue=null,De.apply(Me,arguments)}function $e(e){var t=e,n=e;if(e.alternate)for(;t.return;)t=t.return;else{e=t;do{0!==(4098&(t=e).flags)&&(n=t.return),e=t.return}while(e)}return 3===t.tag?n:null}function We(e){if(13===e.tag){var t=e.memoizedState;if(null===t&&(null!==(e=e.alternate)&&(t=e.memoizedState)),null!==t)return t.dehydrated}return null}function He(e){if($e(e)!==e)throw Error(o(188))}function Ve(e){return null!==(e=function(e){var t=e.alternate;if(!t){if(null===(t=$e(e)))throw Error(o(188));return t!==e?null:e}for(var n=e,r=t;;){var a=n.return;if(null===a)break;var l=a.alternate;if(null===l){if(null!==(r=a.return)){n=r;continue}break}if(a.child===l.child){for(l=a.child;l;){if(l===n)return He(a),e;if(l===r)return He(a),t;l=l.sibling}throw Error(o(188))}if(n.return!==r.return)n=a,r=l;else{for(var i=!1,s=a.child;s;){if(s===n){i=!0,n=a,r=l;break}if(s===r){i=!0,r=a,n=l;break}s=s.sibling}if(!i){for(s=l.child;s;){if(s===n){i=!0,n=l,r=a;break}if(s===r){i=!0,r=l,n=a;break}s=s.sibling}if(!i)throw Error(o(189))}}if(n.alternate!==r)throw Error(o(190))}if(3!==n.tag)throw Error(o(188));return n.stateNode.current===n?e:t}(e))?qe(e):null}function qe(e){if(5===e.tag||6===e.tag)return e;for(e=e.child;null!==e;){var t=qe(e);if(null!==t)return t;e=e.sibling}return null}var Ke=a.unstable_scheduleCallback,Qe=a.unstable_cancelCallback,Xe=a.unstable_shouldYield,Ge=a.unstable_requestPaint,Je=a.unstable_now,Ye=a.unstable_getCurrentPriorityLevel,Ze=a.unstable_ImmediatePriority,et=a.unstable_UserBlockingPriority,tt=a.unstable_NormalPriority,nt=a.unstable_LowPriority,rt=a.unstable_IdlePriority,at=null,ot=null;var lt=Math.clz32?Math.clz32:function(e){return e>>>=0,0===e?32:31-(it(e)/st|0)|0},it=Math.log,st=Math.LN2;var ut=64,ct=4194304;function dt(e){switch(e&-e){case 1:return 1;case 2:return 2;case 4:return 4;case 8:return 8;case 16:return 16;case 32:return 32;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return 4194240&e;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return 130023424&e;case 134217728:return 134217728;case 268435456:return 268435456;case 536870912:return 536870912;case 1073741824:return 1073741824;default:return e}}function ft(e,t){var n=e.pendingLanes;if(0===n)return 0;var r=0,a=e.suspendedLanes,o=e.pingedLanes,l=268435455&n;if(0!==l){var i=l&~a;0!==i?r=dt(i):0!==(o&=l)&&(r=dt(o))}else 0!==(l=n&~a)?r=dt(l):0!==o&&(r=dt(o));if(0===r)return 0;if(0!==t&&t!==r&&0===(t&a)&&((a=r&-r)>=(o=t&-t)||16===a&&0!==(4194240&o)))return t;if(0!==(4&r)&&(r|=16&n),0!==(t=e.entangledLanes))for(e=e.entanglements,t&=r;0<t;)a=1<<(n=31-lt(t)),r|=e[n],t&=~a;return r}function pt(e,t){switch(e){case 1:case 2:case 4:return t+250;case 8:case 16:case 32:case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return t+5e3;default:return-1}}function ht(e){return 0!==(e=-1073741825&e.pendingLanes)?e:1073741824&e?1073741824:0}function mt(){var e=ut;return 0===(4194240&(ut<<=1))&&(ut=64),e}function vt(e){for(var t=[],n=0;31>n;n++)t.push(e);return t}function gt(e,t,n){e.pendingLanes|=t,536870912!==t&&(e.suspendedLanes=0,e.pingedLanes=0),(e=e.eventTimes)[t=31-lt(t)]=n}function yt(e,t){var n=e.entangledLanes|=t;for(e=e.entanglements;n;){var r=31-lt(n),a=1<<r;a&t|e[r]&t&&(e[r]|=t),n&=~a}}var bt=0;function wt(e){return 1<(e&=-e)?4<e?0!==(268435455&e)?16:536870912:4:1}var xt,Et,St,kt,Ct,Nt=!1,Ot=[],Rt=null,jt=null,_t=null,Tt=new Map,Pt=new Map,Lt=[],At="mousedown mouseup touchcancel touchend touchstart auxclick dblclick pointercancel pointerdown pointerup dragend dragstart drop compositionend compositionstart keydown keypress keyup input textInput copy cut paste click change contextmenu reset submit".split(" ");function Dt(e,t){switch(e){case"focusin":case"focusout":Rt=null;break;case"dragenter":case"dragleave":jt=null;break;case"mouseover":case"mouseout":_t=null;break;case"pointerover":case"pointerout":Tt.delete(t.pointerId);break;case"gotpointercapture":case"lostpointercapture":Pt.delete(t.pointerId)}}function Ft(e,t,n,r,a,o){return null===e||e.nativeEvent!==o?(e={blockedOn:t,domEventName:n,eventSystemFlags:r,nativeEvent:o,targetContainers:[a]},null!==t&&(null!==(t=ba(t))&&Et(t)),e):(e.eventSystemFlags|=r,t=e.targetContainers,null!==a&&-1===t.indexOf(a)&&t.push(a),e)}function Ut(e){var t=ya(e.target);if(null!==t){var n=$e(t);if(null!==n)if(13===(t=n.tag)){if(null!==(t=We(n)))return e.blockedOn=t,void Ct(e.priority,(function(){St(n)}))}else if(3===t&&n.stateNode.current.memoizedState.isDehydrated)return void(e.blockedOn=3===n.tag?n.stateNode.containerInfo:null)}e.blockedOn=null}function zt(e){if(null!==e.blockedOn)return!1;for(var t=e.targetContainers;0<t.length;){var n=Xt(e.domEventName,e.eventSystemFlags,t[0],e.nativeEvent);if(null!==n)return null!==(t=ba(n))&&Et(t),e.blockedOn=n,!1;var r=new(n=e.nativeEvent).constructor(n.type,n);we=r,n.target.dispatchEvent(r),we=null,t.shift()}return!0}function It(e,t,n){zt(e)&&n.delete(t)}function Mt(){Nt=!1,null!==Rt&&zt(Rt)&&(Rt=null),null!==jt&&zt(jt)&&(jt=null),null!==_t&&zt(_t)&&(_t=null),Tt.forEach(It),Pt.forEach(It)}function Bt(e,t){e.blockedOn===t&&(e.blockedOn=null,Nt||(Nt=!0,a.unstable_scheduleCallback(a.unstable_NormalPriority,Mt)))}function $t(e){function t(t){return Bt(t,e)}if(0<Ot.length){Bt(Ot[0],e);for(var n=1;n<Ot.length;n++){var r=Ot[n];r.blockedOn===e&&(r.blockedOn=null)}}for(null!==Rt&&Bt(Rt,e),null!==jt&&Bt(jt,e),null!==_t&&Bt(_t,e),Tt.forEach(t),Pt.forEach(t),n=0;n<Lt.length;n++)(r=Lt[n]).blockedOn===e&&(r.blockedOn=null);for(;0<Lt.length&&null===(n=Lt[0]).blockedOn;)Ut(n),null===n.blockedOn&&Lt.shift()}var Wt=w.ReactCurrentBatchConfig,Ht=!0;function Vt(e,t,n,r){var a=bt,o=Wt.transition;Wt.transition=null;try{bt=1,Kt(e,t,n,r)}finally{bt=a,Wt.transition=o}}function qt(e,t,n,r){var a=bt,o=Wt.transition;Wt.transition=null;try{bt=4,Kt(e,t,n,r)}finally{bt=a,Wt.transition=o}}function Kt(e,t,n,r){if(Ht){var a=Xt(e,t,n,r);if(null===a)Hr(e,t,r,Qt,n),Dt(e,r);else if(function(e,t,n,r,a){switch(t){case"focusin":return Rt=Ft(Rt,e,t,n,r,a),!0;case"dragenter":return jt=Ft(jt,e,t,n,r,a),!0;case"mouseover":return _t=Ft(_t,e,t,n,r,a),!0;case"pointerover":var o=a.pointerId;return Tt.set(o,Ft(Tt.get(o)||null,e,t,n,r,a)),!0;case"gotpointercapture":return o=a.pointerId,Pt.set(o,Ft(Pt.get(o)||null,e,t,n,r,a)),!0}return!1}(a,e,t,n,r))r.stopPropagation();else if(Dt(e,r),4&t&&-1<At.indexOf(e)){for(;null!==a;){var o=ba(a);if(null!==o&&xt(o),null===(o=Xt(e,t,n,r))&&Hr(e,t,r,Qt,n),o===a)break;a=o}null!==a&&r.stopPropagation()}else Hr(e,t,r,null,n)}}var Qt=null;function Xt(e,t,n,r){if(Qt=null,null!==(e=ya(e=xe(r))))if(null===(t=$e(e)))e=null;else if(13===(n=t.tag)){if(null!==(e=We(t)))return e;e=null}else if(3===n){if(t.stateNode.current.memoizedState.isDehydrated)return 3===t.tag?t.stateNode.containerInfo:null;e=null}else t!==e&&(e=null);return Qt=e,null}function Gt(e){switch(e){case"cancel":case"click":case"close":case"contextmenu":case"copy":case"cut":case"auxclick":case"dblclick":case"dragend":case"dragstart":case"drop":case"focusin":case"focusout":case"input":case"invalid":case"keydown":case"keypress":case"keyup":case"mousedown":case"mouseup":case"paste":case"pause":case"play":case"pointercancel":case"pointerdown":case"pointerup":case"ratechange":case"reset":case"resize":case"seeked":case"submit":case"touchcancel":case"touchend":case"touchstart":case"volumechange":case"change":case"selectionchange":case"textInput":case"compositionstart":case"compositionend":case"compositionupdate":case"beforeblur":case"afterblur":case"beforeinput":case"blur":case"fullscreenchange":case"focus":case"hashchange":case"popstate":case"select":case"selectstart":return 1;case"drag":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"mousemove":case"mouseout":case"mouseover":case"pointermove":case"pointerout":case"pointerover":case"scroll":case"toggle":case"touchmove":case"wheel":case"mouseenter":case"mouseleave":case"pointerenter":case"pointerleave":return 4;case"message":switch(Ye()){case Ze:return 1;case et:return 4;case tt:case nt:return 16;case rt:return 536870912;default:return 16}default:return 16}}var Jt=null,Yt=null,Zt=null;function en(){if(Zt)return Zt;var e,t,n=Yt,r=n.length,a="value"in Jt?Jt.value:Jt.textContent,o=a.length;for(e=0;e<r&&n[e]===a[e];e++);var l=r-e;for(t=1;t<=l&&n[r-t]===a[o-t];t++);return Zt=a.slice(e,1<t?1-t:void 0)}function tn(e){var t=e.keyCode;return"charCode"in e?0===(e=e.charCode)&&13===t&&(e=13):e=t,10===e&&(e=13),32<=e||13===e?e:0}function nn(){return!0}function rn(){return!1}function an(e){function t(t,n,r,a,o){for(var l in this._reactName=t,this._targetInst=r,this.type=n,this.nativeEvent=a,this.target=o,this.currentTarget=null,e)e.hasOwnProperty(l)&&(t=e[l],this[l]=t?t(a):a[l]);return this.isDefaultPrevented=(null!=a.defaultPrevented?a.defaultPrevented:!1===a.returnValue)?nn:rn,this.isPropagationStopped=rn,this}return U(t.prototype,{preventDefault:function(){this.defaultPrevented=!0;var e=this.nativeEvent;e&&(e.preventDefault?e.preventDefault():"unknown"!==typeof e.returnValue&&(e.returnValue=!1),this.isDefaultPrevented=nn)},stopPropagation:function(){var e=this.nativeEvent;e&&(e.stopPropagation?e.stopPropagation():"unknown"!==typeof e.cancelBubble&&(e.cancelBubble=!0),this.isPropagationStopped=nn)},persist:function(){},isPersistent:nn}),t}var on,ln,sn,un={eventPhase:0,bubbles:0,cancelable:0,timeStamp:function(e){return e.timeStamp||Date.now()},defaultPrevented:0,isTrusted:0},cn=an(un),dn=U({},un,{view:0,detail:0}),fn=an(dn),pn=U({},dn,{screenX:0,screenY:0,clientX:0,clientY:0,pageX:0,pageY:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,getModifierState:Cn,button:0,buttons:0,relatedTarget:function(e){return void 0===e.relatedTarget?e.fromElement===e.srcElement?e.toElement:e.fromElement:e.relatedTarget},movementX:function(e){return"movementX"in e?e.movementX:(e!==sn&&(sn&&"mousemove"===e.type?(on=e.screenX-sn.screenX,ln=e.screenY-sn.screenY):ln=on=0,sn=e),on)},movementY:function(e){return"movementY"in e?e.movementY:ln}}),hn=an(pn),mn=an(U({},pn,{dataTransfer:0})),vn=an(U({},dn,{relatedTarget:0})),gn=an(U({},un,{animationName:0,elapsedTime:0,pseudoElement:0})),yn=U({},un,{clipboardData:function(e){return"clipboardData"in e?e.clipboardData:window.clipboardData}}),bn=an(yn),wn=an(U({},un,{data:0})),xn={Esc:"Escape",Spacebar:" ",Left:"ArrowLeft",Up:"ArrowUp",Right:"ArrowRight",Down:"ArrowDown",Del:"Delete",Win:"OS",Menu:"ContextMenu",Apps:"ContextMenu",Scroll:"ScrollLock",MozPrintableKey:"Unidentified"},En={8:"Backspace",9:"Tab",12:"Clear",13:"Enter",16:"Shift",17:"Control",18:"Alt",19:"Pause",20:"CapsLock",27:"Escape",32:" ",33:"PageUp",34:"PageDown",35:"End",36:"Home",37:"ArrowLeft",38:"ArrowUp",39:"ArrowRight",40:"ArrowDown",45:"Insert",46:"Delete",112:"F1",113:"F2",114:"F3",115:"F4",116:"F5",117:"F6",118:"F7",119:"F8",120:"F9",121:"F10",122:"F11",123:"F12",144:"NumLock",145:"ScrollLock",224:"Meta"},Sn={Alt:"altKey",Control:"ctrlKey",Meta:"metaKey",Shift:"shiftKey"};function kn(e){var t=this.nativeEvent;return t.getModifierState?t.getModifierState(e):!!(e=Sn[e])&&!!t[e]}function Cn(){return kn}var Nn=U({},dn,{key:function(e){if(e.key){var t=xn[e.key]||e.key;if("Unidentified"!==t)return t}return"keypress"===e.type?13===(e=tn(e))?"Enter":String.fromCharCode(e):"keydown"===e.type||"keyup"===e.type?En[e.keyCode]||"Unidentified":""},code:0,location:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,repeat:0,locale:0,getModifierState:Cn,charCode:function(e){return"keypress"===e.type?tn(e):0},keyCode:function(e){return"keydown"===e.type||"keyup"===e.type?e.keyCode:0},which:function(e){return"keypress"===e.type?tn(e):"keydown"===e.type||"keyup"===e.type?e.keyCode:0}}),On=an(Nn),Rn=an(U({},pn,{pointerId:0,width:0,height:0,pressure:0,tangentialPressure:0,tiltX:0,tiltY:0,twist:0,pointerType:0,isPrimary:0})),jn=an(U({},dn,{touches:0,targetTouches:0,changedTouches:0,altKey:0,metaKey:0,ctrlKey:0,shiftKey:0,getModifierState:Cn})),_n=an(U({},un,{propertyName:0,elapsedTime:0,pseudoElement:0})),Tn=U({},pn,{deltaX:function(e){return"deltaX"in e?e.deltaX:"wheelDeltaX"in e?-e.wheelDeltaX:0},deltaY:function(e){return"deltaY"in e?e.deltaY:"wheelDeltaY"in e?-e.wheelDeltaY:"wheelDelta"in e?-e.wheelDelta:0},deltaZ:0,deltaMode:0}),Pn=an(Tn),Ln=[9,13,27,32],An=c&&"CompositionEvent"in window,Dn=null;c&&"documentMode"in document&&(Dn=document.documentMode);var Fn=c&&"TextEvent"in window&&!Dn,Un=c&&(!An||Dn&&8<Dn&&11>=Dn),zn=String.fromCharCode(32),In=!1;function Mn(e,t){switch(e){case"keyup":return-1!==Ln.indexOf(t.keyCode);case"keydown":return 229!==t.keyCode;case"keypress":case"mousedown":case"focusout":return!0;default:return!1}}function Bn(e){return"object"===typeof(e=e.detail)&&"data"in e?e.data:null}var $n=!1;var Wn={color:!0,date:!0,datetime:!0,"datetime-local":!0,email:!0,month:!0,number:!0,password:!0,range:!0,search:!0,tel:!0,text:!0,time:!0,url:!0,week:!0};function Hn(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return"input"===t?!!Wn[e.type]:"textarea"===t}function Vn(e,t,n,r){Ne(r),0<(t=qr(t,"onChange")).length&&(n=new cn("onChange","change",null,n,r),e.push({event:n,listeners:t}))}var qn=null,Kn=null;function Qn(e){zr(e,0)}function Xn(e){if(K(wa(e)))return e}function Gn(e,t){if("change"===e)return t}var Jn=!1;if(c){var Yn;if(c){var Zn="oninput"in document;if(!Zn){var er=document.createElement("div");er.setAttribute("oninput","return;"),Zn="function"===typeof er.oninput}Yn=Zn}else Yn=!1;Jn=Yn&&(!document.documentMode||9<document.documentMode)}function tr(){qn&&(qn.detachEvent("onpropertychange",nr),Kn=qn=null)}function nr(e){if("value"===e.propertyName&&Xn(Kn)){var t=[];Vn(t,Kn,e,xe(e)),Te(Qn,t)}}function rr(e,t,n){"focusin"===e?(tr(),Kn=n,(qn=t).attachEvent("onpropertychange",nr)):"focusout"===e&&tr()}function ar(e){if("selectionchange"===e||"keyup"===e||"keydown"===e)return Xn(Kn)}function or(e,t){if("click"===e)return Xn(t)}function lr(e,t){if("input"===e||"change"===e)return Xn(t)}var ir="function"===typeof Object.is?Object.is:function(e,t){return e===t&&(0!==e||1/e===1/t)||e!==e&&t!==t};function sr(e,t){if(ir(e,t))return!0;if("object"!==typeof e||null===e||"object"!==typeof t||null===t)return!1;var n=Object.keys(e),r=Object.keys(t);if(n.length!==r.length)return!1;for(r=0;r<n.length;r++){var a=n[r];if(!d.call(t,a)||!ir(e[a],t[a]))return!1}return!0}function ur(e){for(;e&&e.firstChild;)e=e.firstChild;return e}function cr(e,t){var n,r=ur(e);for(e=0;r;){if(3===r.nodeType){if(n=e+r.textContent.length,e<=t&&n>=t)return{node:r,offset:t-e};e=n}e:{for(;r;){if(r.nextSibling){r=r.nextSibling;break e}r=r.parentNode}r=void 0}r=ur(r)}}function dr(e,t){return!(!e||!t)&&(e===t||(!e||3!==e.nodeType)&&(t&&3===t.nodeType?dr(e,t.parentNode):"contains"in e?e.contains(t):!!e.compareDocumentPosition&&!!(16&e.compareDocumentPosition(t))))}function fr(){for(var e=window,t=Q();t instanceof e.HTMLIFrameElement;){try{var n="string"===typeof t.contentWindow.location.href}catch(r){n=!1}if(!n)break;t=Q((e=t.contentWindow).document)}return t}function pr(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t&&("input"===t&&("text"===e.type||"search"===e.type||"tel"===e.type||"url"===e.type||"password"===e.type)||"textarea"===t||"true"===e.contentEditable)}function hr(e){var t=fr(),n=e.focusedElem,r=e.selectionRange;if(t!==n&&n&&n.ownerDocument&&dr(n.ownerDocument.documentElement,n)){if(null!==r&&pr(n))if(t=r.start,void 0===(e=r.end)&&(e=t),"selectionStart"in n)n.selectionStart=t,n.selectionEnd=Math.min(e,n.value.length);else if((e=(t=n.ownerDocument||document)&&t.defaultView||window).getSelection){e=e.getSelection();var a=n.textContent.length,o=Math.min(r.start,a);r=void 0===r.end?o:Math.min(r.end,a),!e.extend&&o>r&&(a=r,r=o,o=a),a=cr(n,o);var l=cr(n,r);a&&l&&(1!==e.rangeCount||e.anchorNode!==a.node||e.anchorOffset!==a.offset||e.focusNode!==l.node||e.focusOffset!==l.offset)&&((t=t.createRange()).setStart(a.node,a.offset),e.removeAllRanges(),o>r?(e.addRange(t),e.extend(l.node,l.offset)):(t.setEnd(l.node,l.offset),e.addRange(t)))}for(t=[],e=n;e=e.parentNode;)1===e.nodeType&&t.push({element:e,left:e.scrollLeft,top:e.scrollTop});for("function"===typeof n.focus&&n.focus(),n=0;n<t.length;n++)(e=t[n]).element.scrollLeft=e.left,e.element.scrollTop=e.top}}var mr=c&&"documentMode"in document&&11>=document.documentMode,vr=null,gr=null,yr=null,br=!1;function wr(e,t,n){var r=n.window===n?n.document:9===n.nodeType?n:n.ownerDocument;br||null==vr||vr!==Q(r)||("selectionStart"in(r=vr)&&pr(r)?r={start:r.selectionStart,end:r.selectionEnd}:r={anchorNode:(r=(r.ownerDocument&&r.ownerDocument.defaultView||window).getSelection()).anchorNode,anchorOffset:r.anchorOffset,focusNode:r.focusNode,focusOffset:r.focusOffset},yr&&sr(yr,r)||(yr=r,0<(r=qr(gr,"onSelect")).length&&(t=new cn("onSelect","select",null,t,n),e.push({event:t,listeners:r}),t.target=vr)))}function xr(e,t){var n={};return n[e.toLowerCase()]=t.toLowerCase(),n["Webkit"+e]="webkit"+t,n["Moz"+e]="moz"+t,n}var Er={animationend:xr("Animation","AnimationEnd"),animationiteration:xr("Animation","AnimationIteration"),animationstart:xr("Animation","AnimationStart"),transitionend:xr("Transition","TransitionEnd")},Sr={},kr={};function Cr(e){if(Sr[e])return Sr[e];if(!Er[e])return e;var t,n=Er[e];for(t in n)if(n.hasOwnProperty(t)&&t in kr)return Sr[e]=n[t];return e}c&&(kr=document.createElement("div").style,"AnimationEvent"in window||(delete Er.animationend.animation,delete Er.animationiteration.animation,delete Er.animationstart.animation),"TransitionEvent"in window||delete Er.transitionend.transition);var Nr=Cr("animationend"),Or=Cr("animationiteration"),Rr=Cr("animationstart"),jr=Cr("transitionend"),_r=new Map,Tr="abort auxClick cancel canPlay canPlayThrough click close contextMenu copy cut drag dragEnd dragEnter dragExit dragLeave dragOver dragStart drop durationChange emptied encrypted ended error gotPointerCapture input invalid keyDown keyPress keyUp load loadedData loadedMetadata loadStart lostPointerCapture mouseDown mouseMove mouseOut mouseOver mouseUp paste pause play playing pointerCancel pointerDown pointerMove pointerOut pointerOver pointerUp progress rateChange reset resize seeked seeking stalled submit suspend timeUpdate touchCancel touchEnd touchStart volumeChange scroll toggle touchMove waiting wheel".split(" ");function Pr(e,t){_r.set(e,t),s(t,[e])}for(var Lr=0;Lr<Tr.length;Lr++){var Ar=Tr[Lr];Pr(Ar.toLowerCase(),"on"+(Ar[0].toUpperCase()+Ar.slice(1)))}Pr(Nr,"onAnimationEnd"),Pr(Or,"onAnimationIteration"),Pr(Rr,"onAnimationStart"),Pr("dblclick","onDoubleClick"),Pr("focusin","onFocus"),Pr("focusout","onBlur"),Pr(jr,"onTransitionEnd"),u("onMouseEnter",["mouseout","mouseover"]),u("onMouseLeave",["mouseout","mouseover"]),u("onPointerEnter",["pointerout","pointerover"]),u("onPointerLeave",["pointerout","pointerover"]),s("onChange","change click focusin focusout input keydown keyup selectionchange".split(" ")),s("onSelect","focusout contextmenu dragend focusin keydown keyup mousedown mouseup selectionchange".split(" ")),s("onBeforeInput",["compositionend","keypress","textInput","paste"]),s("onCompositionEnd","compositionend focusout keydown keypress keyup mousedown".split(" ")),s("onCompositionStart","compositionstart focusout keydown keypress keyup mousedown".split(" ")),s("onCompositionUpdate","compositionupdate focusout keydown keypress keyup mousedown".split(" "));var Dr="abort canplay canplaythrough durationchange emptied encrypted ended error loadeddata loadedmetadata loadstart pause play playing progress ratechange resize seeked seeking stalled suspend timeupdate volumechange waiting".split(" "),Fr=new Set("cancel close invalid load scroll toggle".split(" ").concat(Dr));function Ur(e,t,n){var r=e.type||"unknown-event";e.currentTarget=n,function(e,t,n,r,a,l,i,s,u){if(Be.apply(this,arguments),Fe){if(!Fe)throw Error(o(198));var c=Ue;Fe=!1,Ue=null,ze||(ze=!0,Ie=c)}}(r,t,void 0,e),e.currentTarget=null}function zr(e,t){t=0!==(4&t);for(var n=0;n<e.length;n++){var r=e[n],a=r.event;r=r.listeners;e:{var o=void 0;if(t)for(var l=r.length-1;0<=l;l--){var i=r[l],s=i.instance,u=i.currentTarget;if(i=i.listener,s!==o&&a.isPropagationStopped())break e;Ur(a,i,u),o=s}else for(l=0;l<r.length;l++){if(s=(i=r[l]).instance,u=i.currentTarget,i=i.listener,s!==o&&a.isPropagationStopped())break e;Ur(a,i,u),o=s}}}if(ze)throw e=Ie,ze=!1,Ie=null,e}function Ir(e,t){var n=t[ma];void 0===n&&(n=t[ma]=new Set);var r=e+"__bubble";n.has(r)||(Wr(t,e,2,!1),n.add(r))}function Mr(e,t,n){var r=0;t&&(r|=4),Wr(n,e,r,t)}var Br="_reactListening"+Math.random().toString(36).slice(2);function $r(e){if(!e[Br]){e[Br]=!0,l.forEach((function(t){"selectionchange"!==t&&(Fr.has(t)||Mr(t,!1,e),Mr(t,!0,e))}));var t=9===e.nodeType?e:e.ownerDocument;null===t||t[Br]||(t[Br]=!0,Mr("selectionchange",!1,t))}}function Wr(e,t,n,r){switch(Gt(t)){case 1:var a=Vt;break;case 4:a=qt;break;default:a=Kt}n=a.bind(null,t,n,e),a=void 0,!Le||"touchstart"!==t&&"touchmove"!==t&&"wheel"!==t||(a=!0),r?void 0!==a?e.addEventListener(t,n,{capture:!0,passive:a}):e.addEventListener(t,n,!0):void 0!==a?e.addEventListener(t,n,{passive:a}):e.addEventListener(t,n,!1)}function Hr(e,t,n,r,a){var o=r;if(0===(1&t)&&0===(2&t)&&null!==r)e:for(;;){if(null===r)return;var l=r.tag;if(3===l||4===l){var i=r.stateNode.containerInfo;if(i===a||8===i.nodeType&&i.parentNode===a)break;if(4===l)for(l=r.return;null!==l;){var s=l.tag;if((3===s||4===s)&&((s=l.stateNode.containerInfo)===a||8===s.nodeType&&s.parentNode===a))return;l=l.return}for(;null!==i;){if(null===(l=ya(i)))return;if(5===(s=l.tag)||6===s){r=o=l;continue e}i=i.parentNode}}r=r.return}Te((function(){var r=o,a=xe(n),l=[];e:{var i=_r.get(e);if(void 0!==i){var s=cn,u=e;switch(e){case"keypress":if(0===tn(n))break e;case"keydown":case"keyup":s=On;break;case"focusin":u="focus",s=vn;break;case"focusout":u="blur",s=vn;break;case"beforeblur":case"afterblur":s=vn;break;case"click":if(2===n.button)break e;case"auxclick":case"dblclick":case"mousedown":case"mousemove":case"mouseup":case"mouseout":case"mouseover":case"contextmenu":s=hn;break;case"drag":case"dragend":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"dragstart":case"drop":s=mn;break;case"touchcancel":case"touchend":case"touchmove":case"touchstart":s=jn;break;case Nr:case Or:case Rr:s=gn;break;case jr:s=_n;break;case"scroll":s=fn;break;case"wheel":s=Pn;break;case"copy":case"cut":case"paste":s=bn;break;case"gotpointercapture":case"lostpointercapture":case"pointercancel":case"pointerdown":case"pointermove":case"pointerout":case"pointerover":case"pointerup":s=Rn}var c=0!==(4&t),d=!c&&"scroll"===e,f=c?null!==i?i+"Capture":null:i;c=[];for(var p,h=r;null!==h;){var m=(p=h).stateNode;if(5===p.tag&&null!==m&&(p=m,null!==f&&(null!=(m=Pe(h,f))&&c.push(Vr(h,m,p)))),d)break;h=h.return}0<c.length&&(i=new s(i,u,null,n,a),l.push({event:i,listeners:c}))}}if(0===(7&t)){if(s="mouseout"===e||"pointerout"===e,(!(i="mouseover"===e||"pointerover"===e)||n===we||!(u=n.relatedTarget||n.fromElement)||!ya(u)&&!u[ha])&&(s||i)&&(i=a.window===a?a:(i=a.ownerDocument)?i.defaultView||i.parentWindow:window,s?(s=r,null!==(u=(u=n.relatedTarget||n.toElement)?ya(u):null)&&(u!==(d=$e(u))||5!==u.tag&&6!==u.tag)&&(u=null)):(s=null,u=r),s!==u)){if(c=hn,m="onMouseLeave",f="onMouseEnter",h="mouse","pointerout"!==e&&"pointerover"!==e||(c=Rn,m="onPointerLeave",f="onPointerEnter",h="pointer"),d=null==s?i:wa(s),p=null==u?i:wa(u),(i=new c(m,h+"leave",s,n,a)).target=d,i.relatedTarget=p,m=null,ya(a)===r&&((c=new c(f,h+"enter",u,n,a)).target=p,c.relatedTarget=d,m=c),d=m,s&&u)e:{for(f=u,h=0,p=c=s;p;p=Kr(p))h++;for(p=0,m=f;m;m=Kr(m))p++;for(;0<h-p;)c=Kr(c),h--;for(;0<p-h;)f=Kr(f),p--;for(;h--;){if(c===f||null!==f&&c===f.alternate)break e;c=Kr(c),f=Kr(f)}c=null}else c=null;null!==s&&Qr(l,i,s,c,!1),null!==u&&null!==d&&Qr(l,d,u,c,!0)}if("select"===(s=(i=r?wa(r):window).nodeName&&i.nodeName.toLowerCase())||"input"===s&&"file"===i.type)var v=Gn;else if(Hn(i))if(Jn)v=lr;else{v=ar;var g=rr}else(s=i.nodeName)&&"input"===s.toLowerCase()&&("checkbox"===i.type||"radio"===i.type)&&(v=or);switch(v&&(v=v(e,r))?Vn(l,v,n,a):(g&&g(e,i,r),"focusout"===e&&(g=i._wrapperState)&&g.controlled&&"number"===i.type&&ee(i,"number",i.value)),g=r?wa(r):window,e){case"focusin":(Hn(g)||"true"===g.contentEditable)&&(vr=g,gr=r,yr=null);break;case"focusout":yr=gr=vr=null;break;case"mousedown":br=!0;break;case"contextmenu":case"mouseup":case"dragend":br=!1,wr(l,n,a);break;case"selectionchange":if(mr)break;case"keydown":case"keyup":wr(l,n,a)}var y;if(An)e:{switch(e){case"compositionstart":var b="onCompositionStart";break e;case"compositionend":b="onCompositionEnd";break e;case"compositionupdate":b="onCompositionUpdate";break e}b=void 0}else $n?Mn(e,n)&&(b="onCompositionEnd"):"keydown"===e&&229===n.keyCode&&(b="onCompositionStart");b&&(Un&&"ko"!==n.locale&&($n||"onCompositionStart"!==b?"onCompositionEnd"===b&&$n&&(y=en()):(Yt="value"in(Jt=a)?Jt.value:Jt.textContent,$n=!0)),0<(g=qr(r,b)).length&&(b=new wn(b,e,null,n,a),l.push({event:b,listeners:g}),y?b.data=y:null!==(y=Bn(n))&&(b.data=y))),(y=Fn?function(e,t){switch(e){case"compositionend":return Bn(t);case"keypress":return 32!==t.which?null:(In=!0,zn);case"textInput":return(e=t.data)===zn&&In?null:e;default:return null}}(e,n):function(e,t){if($n)return"compositionend"===e||!An&&Mn(e,t)?(e=en(),Zt=Yt=Jt=null,$n=!1,e):null;switch(e){case"paste":default:return null;case"keypress":if(!(t.ctrlKey||t.altKey||t.metaKey)||t.ctrlKey&&t.altKey){if(t.char&&1<t.char.length)return t.char;if(t.which)return String.fromCharCode(t.which)}return null;case"compositionend":return Un&&"ko"!==t.locale?null:t.data}}(e,n))&&(0<(r=qr(r,"onBeforeInput")).length&&(a=new wn("onBeforeInput","beforeinput",null,n,a),l.push({event:a,listeners:r}),a.data=y))}zr(l,t)}))}function Vr(e,t,n){return{instance:e,listener:t,currentTarget:n}}function qr(e,t){for(var n=t+"Capture",r=[];null!==e;){var a=e,o=a.stateNode;5===a.tag&&null!==o&&(a=o,null!=(o=Pe(e,n))&&r.unshift(Vr(e,o,a)),null!=(o=Pe(e,t))&&r.push(Vr(e,o,a))),e=e.return}return r}function Kr(e){if(null===e)return null;do{e=e.return}while(e&&5!==e.tag);return e||null}function Qr(e,t,n,r,a){for(var o=t._reactName,l=[];null!==n&&n!==r;){var i=n,s=i.alternate,u=i.stateNode;if(null!==s&&s===r)break;5===i.tag&&null!==u&&(i=u,a?null!=(s=Pe(n,o))&&l.unshift(Vr(n,s,i)):a||null!=(s=Pe(n,o))&&l.push(Vr(n,s,i))),n=n.return}0!==l.length&&e.push({event:t,listeners:l})}var Xr=/\r\n?/g,Gr=/\u0000|\uFFFD/g;function Jr(e){return("string"===typeof e?e:""+e).replace(Xr,"\n").replace(Gr,"")}function Yr(e,t,n){if(t=Jr(t),Jr(e)!==t&&n)throw Error(o(425))}function Zr(){}var ea=null,ta=null;function na(e,t){return"textarea"===e||"noscript"===e||"string"===typeof t.children||"number"===typeof t.children||"object"===typeof t.dangerouslySetInnerHTML&&null!==t.dangerouslySetInnerHTML&&null!=t.dangerouslySetInnerHTML.__html}var ra="function"===typeof setTimeout?setTimeout:void 0,aa="function"===typeof clearTimeout?clearTimeout:void 0,oa="function"===typeof Promise?Promise:void 0,la="function"===typeof queueMicrotask?queueMicrotask:"undefined"!==typeof oa?function(e){return oa.resolve(null).then(e).catch(ia)}:ra;function ia(e){setTimeout((function(){throw e}))}function sa(e,t){var n=t,r=0;do{var a=n.nextSibling;if(e.removeChild(n),a&&8===a.nodeType)if("/$"===(n=a.data)){if(0===r)return e.removeChild(a),void $t(t);r--}else"$"!==n&&"$?"!==n&&"$!"!==n||r++;n=a}while(n);$t(t)}function ua(e){for(;null!=e;e=e.nextSibling){var t=e.nodeType;if(1===t||3===t)break;if(8===t){if("$"===(t=e.data)||"$!"===t||"$?"===t)break;if("/$"===t)return null}}return e}function ca(e){e=e.previousSibling;for(var t=0;e;){if(8===e.nodeType){var n=e.data;if("$"===n||"$!"===n||"$?"===n){if(0===t)return e;t--}else"/$"===n&&t++}e=e.previousSibling}return null}var da=Math.random().toString(36).slice(2),fa="__reactFiber$"+da,pa="__reactProps$"+da,ha="__reactContainer$"+da,ma="__reactEvents$"+da,va="__reactListeners$"+da,ga="__reactHandles$"+da;function ya(e){var t=e[fa];if(t)return t;for(var n=e.parentNode;n;){if(t=n[ha]||n[fa]){if(n=t.alternate,null!==t.child||null!==n&&null!==n.child)for(e=ca(e);null!==e;){if(n=e[fa])return n;e=ca(e)}return t}n=(e=n).parentNode}return null}function ba(e){return!(e=e[fa]||e[ha])||5!==e.tag&&6!==e.tag&&13!==e.tag&&3!==e.tag?null:e}function wa(e){if(5===e.tag||6===e.tag)return e.stateNode;throw Error(o(33))}function xa(e){return e[pa]||null}var Ea=[],Sa=-1;function ka(e){return{current:e}}function Ca(e){0>Sa||(e.current=Ea[Sa],Ea[Sa]=null,Sa--)}function Na(e,t){Sa++,Ea[Sa]=e.current,e.current=t}var Oa={},Ra=ka(Oa),ja=ka(!1),_a=Oa;function Ta(e,t){var n=e.type.contextTypes;if(!n)return Oa;var r=e.stateNode;if(r&&r.__reactInternalMemoizedUnmaskedChildContext===t)return r.__reactInternalMemoizedMaskedChildContext;var a,o={};for(a in n)o[a]=t[a];return r&&((e=e.stateNode).__reactInternalMemoizedUnmaskedChildContext=t,e.__reactInternalMemoizedMaskedChildContext=o),o}function Pa(e){return null!==(e=e.childContextTypes)&&void 0!==e}function La(){Ca(ja),Ca(Ra)}function Aa(e,t,n){if(Ra.current!==Oa)throw Error(o(168));Na(Ra,t),Na(ja,n)}function Da(e,t,n){var r=e.stateNode;if(t=t.childContextTypes,"function"!==typeof r.getChildContext)return n;for(var a in r=r.getChildContext())if(!(a in t))throw Error(o(108,W(e)||"Unknown",a));return U({},n,r)}function Fa(e){return e=(e=e.stateNode)&&e.__reactInternalMemoizedMergedChildContext||Oa,_a=Ra.current,Na(Ra,e),Na(ja,ja.current),!0}function Ua(e,t,n){var r=e.stateNode;if(!r)throw Error(o(169));n?(e=Da(e,t,_a),r.__reactInternalMemoizedMergedChildContext=e,Ca(ja),Ca(Ra),Na(Ra,e)):Ca(ja),Na(ja,n)}var za=null,Ia=!1,Ma=!1;function Ba(e){null===za?za=[e]:za.push(e)}function $a(){if(!Ma&&null!==za){Ma=!0;var e=0,t=bt;try{var n=za;for(bt=1;e<n.length;e++){var r=n[e];do{r=r(!0)}while(null!==r)}za=null,Ia=!1}catch(a){throw null!==za&&(za=za.slice(e+1)),Ke(Ze,$a),a}finally{bt=t,Ma=!1}}return null}var Wa=[],Ha=0,Va=null,qa=0,Ka=[],Qa=0,Xa=null,Ga=1,Ja="";function Ya(e,t){Wa[Ha++]=qa,Wa[Ha++]=Va,Va=e,qa=t}function Za(e,t,n){Ka[Qa++]=Ga,Ka[Qa++]=Ja,Ka[Qa++]=Xa,Xa=e;var r=Ga;e=Ja;var a=32-lt(r)-1;r&=~(1<<a),n+=1;var o=32-lt(t)+a;if(30<o){var l=a-a%5;o=(r&(1<<l)-1).toString(32),r>>=l,a-=l,Ga=1<<32-lt(t)+a|n<<a|r,Ja=o+e}else Ga=1<<o|n<<a|r,Ja=e}function eo(e){null!==e.return&&(Ya(e,1),Za(e,1,0))}function to(e){for(;e===Va;)Va=Wa[--Ha],Wa[Ha]=null,qa=Wa[--Ha],Wa[Ha]=null;for(;e===Xa;)Xa=Ka[--Qa],Ka[Qa]=null,Ja=Ka[--Qa],Ka[Qa]=null,Ga=Ka[--Qa],Ka[Qa]=null}var no=null,ro=null,ao=!1,oo=null;function lo(e,t){var n=Tu(5,null,null,0);n.elementType="DELETED",n.stateNode=t,n.return=e,null===(t=e.deletions)?(e.deletions=[n],e.flags|=16):t.push(n)}function io(e,t){switch(e.tag){case 5:var n=e.type;return null!==(t=1!==t.nodeType||n.toLowerCase()!==t.nodeName.toLowerCase()?null:t)&&(e.stateNode=t,no=e,ro=ua(t.firstChild),!0);case 6:return null!==(t=""===e.pendingProps||3!==t.nodeType?null:t)&&(e.stateNode=t,no=e,ro=null,!0);case 13:return null!==(t=8!==t.nodeType?null:t)&&(n=null!==Xa?{id:Ga,overflow:Ja}:null,e.memoizedState={dehydrated:t,treeContext:n,retryLane:1073741824},(n=Tu(18,null,null,0)).stateNode=t,n.return=e,e.child=n,no=e,ro=null,!0);default:return!1}}function so(e){return 0!==(1&e.mode)&&0===(128&e.flags)}function uo(e){if(ao){var t=ro;if(t){var n=t;if(!io(e,t)){if(so(e))throw Error(o(418));t=ua(n.nextSibling);var r=no;t&&io(e,t)?lo(r,n):(e.flags=-4097&e.flags|2,ao=!1,no=e)}}else{if(so(e))throw Error(o(418));e.flags=-4097&e.flags|2,ao=!1,no=e}}}function co(e){for(e=e.return;null!==e&&5!==e.tag&&3!==e.tag&&13!==e.tag;)e=e.return;no=e}function fo(e){if(e!==no)return!1;if(!ao)return co(e),ao=!0,!1;var t;if((t=3!==e.tag)&&!(t=5!==e.tag)&&(t="head"!==(t=e.type)&&"body"!==t&&!na(e.type,e.memoizedProps)),t&&(t=ro)){if(so(e))throw po(),Error(o(418));for(;t;)lo(e,t),t=ua(t.nextSibling)}if(co(e),13===e.tag){if(!(e=null!==(e=e.memoizedState)?e.dehydrated:null))throw Error(o(317));e:{for(e=e.nextSibling,t=0;e;){if(8===e.nodeType){var n=e.data;if("/$"===n){if(0===t){ro=ua(e.nextSibling);break e}t--}else"$"!==n&&"$!"!==n&&"$?"!==n||t++}e=e.nextSibling}ro=null}}else ro=no?ua(e.stateNode.nextSibling):null;return!0}function po(){for(var e=ro;e;)e=ua(e.nextSibling)}function ho(){ro=no=null,ao=!1}function mo(e){null===oo?oo=[e]:oo.push(e)}var vo=w.ReactCurrentBatchConfig;function go(e,t,n){if(null!==(e=n.ref)&&"function"!==typeof e&&"object"!==typeof e){if(n._owner){if(n=n._owner){if(1!==n.tag)throw Error(o(309));var r=n.stateNode}if(!r)throw Error(o(147,e));var a=r,l=""+e;return null!==t&&null!==t.ref&&"function"===typeof t.ref&&t.ref._stringRef===l?t.ref:(t=function(e){var t=a.refs;null===e?delete t[l]:t[l]=e},t._stringRef=l,t)}if("string"!==typeof e)throw Error(o(284));if(!n._owner)throw Error(o(290,e))}return e}function yo(e,t){throw e=Object.prototype.toString.call(t),Error(o(31,"[object Object]"===e?"object with keys {"+Object.keys(t).join(", ")+"}":e))}function bo(e){return(0,e._init)(e._payload)}function wo(e){function t(t,n){if(e){var r=t.deletions;null===r?(t.deletions=[n],t.flags|=16):r.push(n)}}function n(n,r){if(!e)return null;for(;null!==r;)t(n,r),r=r.sibling;return null}function r(e,t){for(e=new Map;null!==t;)null!==t.key?e.set(t.key,t):e.set(t.index,t),t=t.sibling;return e}function a(e,t){return(e=Lu(e,t)).index=0,e.sibling=null,e}function l(t,n,r){return t.index=r,e?null!==(r=t.alternate)?(r=r.index)<n?(t.flags|=2,n):r:(t.flags|=2,n):(t.flags|=1048576,n)}function i(t){return e&&null===t.alternate&&(t.flags|=2),t}function s(e,t,n,r){return null===t||6!==t.tag?((t=Uu(n,e.mode,r)).return=e,t):((t=a(t,n)).return=e,t)}function u(e,t,n,r){var o=n.type;return o===S?d(e,t,n.props.children,r,n.key):null!==t&&(t.elementType===o||"object"===typeof o&&null!==o&&o.$$typeof===P&&bo(o)===t.type)?((r=a(t,n.props)).ref=go(e,t,n),r.return=e,r):((r=Au(n.type,n.key,n.props,null,e.mode,r)).ref=go(e,t,n),r.return=e,r)}function c(e,t,n,r){return null===t||4!==t.tag||t.stateNode.containerInfo!==n.containerInfo||t.stateNode.implementation!==n.implementation?((t=zu(n,e.mode,r)).return=e,t):((t=a(t,n.children||[])).return=e,t)}function d(e,t,n,r,o){return null===t||7!==t.tag?((t=Du(n,e.mode,r,o)).return=e,t):((t=a(t,n)).return=e,t)}function f(e,t,n){if("string"===typeof t&&""!==t||"number"===typeof t)return(t=Uu(""+t,e.mode,n)).return=e,t;if("object"===typeof t&&null!==t){switch(t.$$typeof){case x:return(n=Au(t.type,t.key,t.props,null,e.mode,n)).ref=go(e,null,t),n.return=e,n;case E:return(t=zu(t,e.mode,n)).return=e,t;case P:return f(e,(0,t._init)(t._payload),n)}if(te(t)||D(t))return(t=Du(t,e.mode,n,null)).return=e,t;yo(e,t)}return null}function p(e,t,n,r){var a=null!==t?t.key:null;if("string"===typeof n&&""!==n||"number"===typeof n)return null!==a?null:s(e,t,""+n,r);if("object"===typeof n&&null!==n){switch(n.$$typeof){case x:return n.key===a?u(e,t,n,r):null;case E:return n.key===a?c(e,t,n,r):null;case P:return p(e,t,(a=n._init)(n._payload),r)}if(te(n)||D(n))return null!==a?null:d(e,t,n,r,null);yo(e,n)}return null}function h(e,t,n,r,a){if("string"===typeof r&&""!==r||"number"===typeof r)return s(t,e=e.get(n)||null,""+r,a);if("object"===typeof r&&null!==r){switch(r.$$typeof){case x:return u(t,e=e.get(null===r.key?n:r.key)||null,r,a);case E:return c(t,e=e.get(null===r.key?n:r.key)||null,r,a);case P:return h(e,t,n,(0,r._init)(r._payload),a)}if(te(r)||D(r))return d(t,e=e.get(n)||null,r,a,null);yo(t,r)}return null}function m(a,o,i,s){for(var u=null,c=null,d=o,m=o=0,v=null;null!==d&&m<i.length;m++){d.index>m?(v=d,d=null):v=d.sibling;var g=p(a,d,i[m],s);if(null===g){null===d&&(d=v);break}e&&d&&null===g.alternate&&t(a,d),o=l(g,o,m),null===c?u=g:c.sibling=g,c=g,d=v}if(m===i.length)return n(a,d),ao&&Ya(a,m),u;if(null===d){for(;m<i.length;m++)null!==(d=f(a,i[m],s))&&(o=l(d,o,m),null===c?u=d:c.sibling=d,c=d);return ao&&Ya(a,m),u}for(d=r(a,d);m<i.length;m++)null!==(v=h(d,a,m,i[m],s))&&(e&&null!==v.alternate&&d.delete(null===v.key?m:v.key),o=l(v,o,m),null===c?u=v:c.sibling=v,c=v);return e&&d.forEach((function(e){return t(a,e)})),ao&&Ya(a,m),u}function v(a,i,s,u){var c=D(s);if("function"!==typeof c)throw Error(o(150));if(null==(s=c.call(s)))throw Error(o(151));for(var d=c=null,m=i,v=i=0,g=null,y=s.next();null!==m&&!y.done;v++,y=s.next()){m.index>v?(g=m,m=null):g=m.sibling;var b=p(a,m,y.value,u);if(null===b){null===m&&(m=g);break}e&&m&&null===b.alternate&&t(a,m),i=l(b,i,v),null===d?c=b:d.sibling=b,d=b,m=g}if(y.done)return n(a,m),ao&&Ya(a,v),c;if(null===m){for(;!y.done;v++,y=s.next())null!==(y=f(a,y.value,u))&&(i=l(y,i,v),null===d?c=y:d.sibling=y,d=y);return ao&&Ya(a,v),c}for(m=r(a,m);!y.done;v++,y=s.next())null!==(y=h(m,a,v,y.value,u))&&(e&&null!==y.alternate&&m.delete(null===y.key?v:y.key),i=l(y,i,v),null===d?c=y:d.sibling=y,d=y);return e&&m.forEach((function(e){return t(a,e)})),ao&&Ya(a,v),c}return function e(r,o,l,s){if("object"===typeof l&&null!==l&&l.type===S&&null===l.key&&(l=l.props.children),"object"===typeof l&&null!==l){switch(l.$$typeof){case x:e:{for(var u=l.key,c=o;null!==c;){if(c.key===u){if((u=l.type)===S){if(7===c.tag){n(r,c.sibling),(o=a(c,l.props.children)).return=r,r=o;break e}}else if(c.elementType===u||"object"===typeof u&&null!==u&&u.$$typeof===P&&bo(u)===c.type){n(r,c.sibling),(o=a(c,l.props)).ref=go(r,c,l),o.return=r,r=o;break e}n(r,c);break}t(r,c),c=c.sibling}l.type===S?((o=Du(l.props.children,r.mode,s,l.key)).return=r,r=o):((s=Au(l.type,l.key,l.props,null,r.mode,s)).ref=go(r,o,l),s.return=r,r=s)}return i(r);case E:e:{for(c=l.key;null!==o;){if(o.key===c){if(4===o.tag&&o.stateNode.containerInfo===l.containerInfo&&o.stateNode.implementation===l.implementation){n(r,o.sibling),(o=a(o,l.children||[])).return=r,r=o;break e}n(r,o);break}t(r,o),o=o.sibling}(o=zu(l,r.mode,s)).return=r,r=o}return i(r);case P:return e(r,o,(c=l._init)(l._payload),s)}if(te(l))return m(r,o,l,s);if(D(l))return v(r,o,l,s);yo(r,l)}return"string"===typeof l&&""!==l||"number"===typeof l?(l=""+l,null!==o&&6===o.tag?(n(r,o.sibling),(o=a(o,l)).return=r,r=o):(n(r,o),(o=Uu(l,r.mode,s)).return=r,r=o),i(r)):n(r,o)}}var xo=wo(!0),Eo=wo(!1),So=ka(null),ko=null,Co=null,No=null;function Oo(){No=Co=ko=null}function Ro(e){var t=So.current;Ca(So),e._currentValue=t}function jo(e,t,n){for(;null!==e;){var r=e.alternate;if((e.childLanes&t)!==t?(e.childLanes|=t,null!==r&&(r.childLanes|=t)):null!==r&&(r.childLanes&t)!==t&&(r.childLanes|=t),e===n)break;e=e.return}}function _o(e,t){ko=e,No=Co=null,null!==(e=e.dependencies)&&null!==e.firstContext&&(0!==(e.lanes&t)&&(bi=!0),e.firstContext=null)}function To(e){var t=e._currentValue;if(No!==e)if(e={context:e,memoizedValue:t,next:null},null===Co){if(null===ko)throw Error(o(308));Co=e,ko.dependencies={lanes:0,firstContext:e}}else Co=Co.next=e;return t}var Po=null;function Lo(e){null===Po?Po=[e]:Po.push(e)}function Ao(e,t,n,r){var a=t.interleaved;return null===a?(n.next=n,Lo(t)):(n.next=a.next,a.next=n),t.interleaved=n,Do(e,r)}function Do(e,t){e.lanes|=t;var n=e.alternate;for(null!==n&&(n.lanes|=t),n=e,e=e.return;null!==e;)e.childLanes|=t,null!==(n=e.alternate)&&(n.childLanes|=t),n=e,e=e.return;return 3===n.tag?n.stateNode:null}var Fo=!1;function Uo(e){e.updateQueue={baseState:e.memoizedState,firstBaseUpdate:null,lastBaseUpdate:null,shared:{pending:null,interleaved:null,lanes:0},effects:null}}function zo(e,t){e=e.updateQueue,t.updateQueue===e&&(t.updateQueue={baseState:e.baseState,firstBaseUpdate:e.firstBaseUpdate,lastBaseUpdate:e.lastBaseUpdate,shared:e.shared,effects:e.effects})}function Io(e,t){return{eventTime:e,lane:t,tag:0,payload:null,callback:null,next:null}}function Mo(e,t,n){var r=e.updateQueue;if(null===r)return null;if(r=r.shared,0!==(2&Rs)){var a=r.pending;return null===a?t.next=t:(t.next=a.next,a.next=t),r.pending=t,Do(e,n)}return null===(a=r.interleaved)?(t.next=t,Lo(r)):(t.next=a.next,a.next=t),r.interleaved=t,Do(e,n)}function Bo(e,t,n){if(null!==(t=t.updateQueue)&&(t=t.shared,0!==(4194240&n))){var r=t.lanes;n|=r&=e.pendingLanes,t.lanes=n,yt(e,n)}}function $o(e,t){var n=e.updateQueue,r=e.alternate;if(null!==r&&n===(r=r.updateQueue)){var a=null,o=null;if(null!==(n=n.firstBaseUpdate)){do{var l={eventTime:n.eventTime,lane:n.lane,tag:n.tag,payload:n.payload,callback:n.callback,next:null};null===o?a=o=l:o=o.next=l,n=n.next}while(null!==n);null===o?a=o=t:o=o.next=t}else a=o=t;return n={baseState:r.baseState,firstBaseUpdate:a,lastBaseUpdate:o,shared:r.shared,effects:r.effects},void(e.updateQueue=n)}null===(e=n.lastBaseUpdate)?n.firstBaseUpdate=t:e.next=t,n.lastBaseUpdate=t}function Wo(e,t,n,r){var a=e.updateQueue;Fo=!1;var o=a.firstBaseUpdate,l=a.lastBaseUpdate,i=a.shared.pending;if(null!==i){a.shared.pending=null;var s=i,u=s.next;s.next=null,null===l?o=u:l.next=u,l=s;var c=e.alternate;null!==c&&((i=(c=c.updateQueue).lastBaseUpdate)!==l&&(null===i?c.firstBaseUpdate=u:i.next=u,c.lastBaseUpdate=s))}if(null!==o){var d=a.baseState;for(l=0,c=u=s=null,i=o;;){var f=i.lane,p=i.eventTime;if((r&f)===f){null!==c&&(c=c.next={eventTime:p,lane:0,tag:i.tag,payload:i.payload,callback:i.callback,next:null});e:{var h=e,m=i;switch(f=t,p=n,m.tag){case 1:if("function"===typeof(h=m.payload)){d=h.call(p,d,f);break e}d=h;break e;case 3:h.flags=-65537&h.flags|128;case 0:if(null===(f="function"===typeof(h=m.payload)?h.call(p,d,f):h)||void 0===f)break e;d=U({},d,f);break e;case 2:Fo=!0}}null!==i.callback&&0!==i.lane&&(e.flags|=64,null===(f=a.effects)?a.effects=[i]:f.push(i))}else p={eventTime:p,lane:f,tag:i.tag,payload:i.payload,callback:i.callback,next:null},null===c?(u=c=p,s=d):c=c.next=p,l|=f;if(null===(i=i.next)){if(null===(i=a.shared.pending))break;i=(f=i).next,f.next=null,a.lastBaseUpdate=f,a.shared.pending=null}}if(null===c&&(s=d),a.baseState=s,a.firstBaseUpdate=u,a.lastBaseUpdate=c,null!==(t=a.shared.interleaved)){a=t;do{l|=a.lane,a=a.next}while(a!==t)}else null===o&&(a.shared.lanes=0);Fs|=l,e.lanes=l,e.memoizedState=d}}function Ho(e,t,n){if(e=t.effects,t.effects=null,null!==e)for(t=0;t<e.length;t++){var r=e[t],a=r.callback;if(null!==a){if(r.callback=null,r=n,"function"!==typeof a)throw Error(o(191,a));a.call(r)}}}var Vo={},qo=ka(Vo),Ko=ka(Vo),Qo=ka(Vo);function Xo(e){if(e===Vo)throw Error(o(174));return e}function Go(e,t){switch(Na(Qo,t),Na(Ko,e),Na(qo,Vo),e=t.nodeType){case 9:case 11:t=(t=t.documentElement)?t.namespaceURI:se(null,"");break;default:t=se(t=(e=8===e?t.parentNode:t).namespaceURI||null,e=e.tagName)}Ca(qo),Na(qo,t)}function Jo(){Ca(qo),Ca(Ko),Ca(Qo)}function Yo(e){Xo(Qo.current);var t=Xo(qo.current),n=se(t,e.type);t!==n&&(Na(Ko,e),Na(qo,n))}function Zo(e){Ko.current===e&&(Ca(qo),Ca(Ko))}var el=ka(0);function tl(e){for(var t=e;null!==t;){if(13===t.tag){var n=t.memoizedState;if(null!==n&&(null===(n=n.dehydrated)||"$?"===n.data||"$!"===n.data))return t}else if(19===t.tag&&void 0!==t.memoizedProps.revealOrder){if(0!==(128&t.flags))return t}else if(null!==t.child){t.child.return=t,t=t.child;continue}if(t===e)break;for(;null===t.sibling;){if(null===t.return||t.return===e)return null;t=t.return}t.sibling.return=t.return,t=t.sibling}return null}var nl=[];function rl(){for(var e=0;e<nl.length;e++)nl[e]._workInProgressVersionPrimary=null;nl.length=0}var al=w.ReactCurrentDispatcher,ol=w.ReactCurrentBatchConfig,ll=0,il=null,sl=null,ul=null,cl=!1,dl=!1,fl=0,pl=0;function hl(){throw Error(o(321))}function ml(e,t){if(null===t)return!1;for(var n=0;n<t.length&&n<e.length;n++)if(!ir(e[n],t[n]))return!1;return!0}function vl(e,t,n,r,a,l){if(ll=l,il=t,t.memoizedState=null,t.updateQueue=null,t.lanes=0,al.current=null===e||null===e.memoizedState?Zl:ei,e=n(r,a),dl){l=0;do{if(dl=!1,fl=0,25<=l)throw Error(o(301));l+=1,ul=sl=null,t.updateQueue=null,al.current=ti,e=n(r,a)}while(dl)}if(al.current=Yl,t=null!==sl&&null!==sl.next,ll=0,ul=sl=il=null,cl=!1,t)throw Error(o(300));return e}function gl(){var e=0!==fl;return fl=0,e}function yl(){var e={memoizedState:null,baseState:null,baseQueue:null,queue:null,next:null};return null===ul?il.memoizedState=ul=e:ul=ul.next=e,ul}function bl(){if(null===sl){var e=il.alternate;e=null!==e?e.memoizedState:null}else e=sl.next;var t=null===ul?il.memoizedState:ul.next;if(null!==t)ul=t,sl=e;else{if(null===e)throw Error(o(310));e={memoizedState:(sl=e).memoizedState,baseState:sl.baseState,baseQueue:sl.baseQueue,queue:sl.queue,next:null},null===ul?il.memoizedState=ul=e:ul=ul.next=e}return ul}function wl(e,t){return"function"===typeof t?t(e):t}function xl(e){var t=bl(),n=t.queue;if(null===n)throw Error(o(311));n.lastRenderedReducer=e;var r=sl,a=r.baseQueue,l=n.pending;if(null!==l){if(null!==a){var i=a.next;a.next=l.next,l.next=i}r.baseQueue=a=l,n.pending=null}if(null!==a){l=a.next,r=r.baseState;var s=i=null,u=null,c=l;do{var d=c.lane;if((ll&d)===d)null!==u&&(u=u.next={lane:0,action:c.action,hasEagerState:c.hasEagerState,eagerState:c.eagerState,next:null}),r=c.hasEagerState?c.eagerState:e(r,c.action);else{var f={lane:d,action:c.action,hasEagerState:c.hasEagerState,eagerState:c.eagerState,next:null};null===u?(s=u=f,i=r):u=u.next=f,il.lanes|=d,Fs|=d}c=c.next}while(null!==c&&c!==l);null===u?i=r:u.next=s,ir(r,t.memoizedState)||(bi=!0),t.memoizedState=r,t.baseState=i,t.baseQueue=u,n.lastRenderedState=r}if(null!==(e=n.interleaved)){a=e;do{l=a.lane,il.lanes|=l,Fs|=l,a=a.next}while(a!==e)}else null===a&&(n.lanes=0);return[t.memoizedState,n.dispatch]}function El(e){var t=bl(),n=t.queue;if(null===n)throw Error(o(311));n.lastRenderedReducer=e;var r=n.dispatch,a=n.pending,l=t.memoizedState;if(null!==a){n.pending=null;var i=a=a.next;do{l=e(l,i.action),i=i.next}while(i!==a);ir(l,t.memoizedState)||(bi=!0),t.memoizedState=l,null===t.baseQueue&&(t.baseState=l),n.lastRenderedState=l}return[l,r]}function Sl(){}function kl(e,t){var n=il,r=bl(),a=t(),l=!ir(r.memoizedState,a);if(l&&(r.memoizedState=a,bi=!0),r=r.queue,Fl(Ol.bind(null,n,r,e),[e]),r.getSnapshot!==t||l||null!==ul&&1&ul.memoizedState.tag){if(n.flags|=2048,Tl(9,Nl.bind(null,n,r,a,t),void 0,null),null===js)throw Error(o(349));0!==(30&ll)||Cl(n,t,a)}return a}function Cl(e,t,n){e.flags|=16384,e={getSnapshot:t,value:n},null===(t=il.updateQueue)?(t={lastEffect:null,stores:null},il.updateQueue=t,t.stores=[e]):null===(n=t.stores)?t.stores=[e]:n.push(e)}function Nl(e,t,n,r){t.value=n,t.getSnapshot=r,Rl(t)&&jl(e)}function Ol(e,t,n){return n((function(){Rl(t)&&jl(e)}))}function Rl(e){var t=e.getSnapshot;e=e.value;try{var n=t();return!ir(e,n)}catch(r){return!0}}function jl(e){var t=Do(e,1);null!==t&&nu(t,e,1,-1)}function _l(e){var t=yl();return"function"===typeof e&&(e=e()),t.memoizedState=t.baseState=e,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:wl,lastRenderedState:e},t.queue=e,e=e.dispatch=Ql.bind(null,il,e),[t.memoizedState,e]}function Tl(e,t,n,r){return e={tag:e,create:t,destroy:n,deps:r,next:null},null===(t=il.updateQueue)?(t={lastEffect:null,stores:null},il.updateQueue=t,t.lastEffect=e.next=e):null===(n=t.lastEffect)?t.lastEffect=e.next=e:(r=n.next,n.next=e,e.next=r,t.lastEffect=e),e}function Pl(){return bl().memoizedState}function Ll(e,t,n,r){var a=yl();il.flags|=e,a.memoizedState=Tl(1|t,n,void 0,void 0===r?null:r)}function Al(e,t,n,r){var a=bl();r=void 0===r?null:r;var o=void 0;if(null!==sl){var l=sl.memoizedState;if(o=l.destroy,null!==r&&ml(r,l.deps))return void(a.memoizedState=Tl(t,n,o,r))}il.flags|=e,a.memoizedState=Tl(1|t,n,o,r)}function Dl(e,t){return Ll(8390656,8,e,t)}function Fl(e,t){return Al(2048,8,e,t)}function Ul(e,t){return Al(4,2,e,t)}function zl(e,t){return Al(4,4,e,t)}function Il(e,t){return"function"===typeof t?(e=e(),t(e),function(){t(null)}):null!==t&&void 0!==t?(e=e(),t.current=e,function(){t.current=null}):void 0}function Ml(e,t,n){return n=null!==n&&void 0!==n?n.concat([e]):null,Al(4,4,Il.bind(null,t,e),n)}function Bl(){}function $l(e,t){var n=bl();t=void 0===t?null:t;var r=n.memoizedState;return null!==r&&null!==t&&ml(t,r[1])?r[0]:(n.memoizedState=[e,t],e)}function Wl(e,t){var n=bl();t=void 0===t?null:t;var r=n.memoizedState;return null!==r&&null!==t&&ml(t,r[1])?r[0]:(e=e(),n.memoizedState=[e,t],e)}function Hl(e,t,n){return 0===(21&ll)?(e.baseState&&(e.baseState=!1,bi=!0),e.memoizedState=n):(ir(n,t)||(n=mt(),il.lanes|=n,Fs|=n,e.baseState=!0),t)}function Vl(e,t){var n=bt;bt=0!==n&&4>n?n:4,e(!0);var r=ol.transition;ol.transition={};try{e(!1),t()}finally{bt=n,ol.transition=r}}function ql(){return bl().memoizedState}function Kl(e,t,n){var r=tu(e);if(n={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null},Xl(e))Gl(t,n);else if(null!==(n=Ao(e,t,n,r))){nu(n,e,r,eu()),Jl(n,t,r)}}function Ql(e,t,n){var r=tu(e),a={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null};if(Xl(e))Gl(t,a);else{var o=e.alternate;if(0===e.lanes&&(null===o||0===o.lanes)&&null!==(o=t.lastRenderedReducer))try{var l=t.lastRenderedState,i=o(l,n);if(a.hasEagerState=!0,a.eagerState=i,ir(i,l)){var s=t.interleaved;return null===s?(a.next=a,Lo(t)):(a.next=s.next,s.next=a),void(t.interleaved=a)}}catch(u){}null!==(n=Ao(e,t,a,r))&&(nu(n,e,r,a=eu()),Jl(n,t,r))}}function Xl(e){var t=e.alternate;return e===il||null!==t&&t===il}function Gl(e,t){dl=cl=!0;var n=e.pending;null===n?t.next=t:(t.next=n.next,n.next=t),e.pending=t}function Jl(e,t,n){if(0!==(4194240&n)){var r=t.lanes;n|=r&=e.pendingLanes,t.lanes=n,yt(e,n)}}var Yl={readContext:To,useCallback:hl,useContext:hl,useEffect:hl,useImperativeHandle:hl,useInsertionEffect:hl,useLayoutEffect:hl,useMemo:hl,useReducer:hl,useRef:hl,useState:hl,useDebugValue:hl,useDeferredValue:hl,useTransition:hl,useMutableSource:hl,useSyncExternalStore:hl,useId:hl,unstable_isNewReconciler:!1},Zl={readContext:To,useCallback:function(e,t){return yl().memoizedState=[e,void 0===t?null:t],e},useContext:To,useEffect:Dl,useImperativeHandle:function(e,t,n){return n=null!==n&&void 0!==n?n.concat([e]):null,Ll(4194308,4,Il.bind(null,t,e),n)},useLayoutEffect:function(e,t){return Ll(4194308,4,e,t)},useInsertionEffect:function(e,t){return Ll(4,2,e,t)},useMemo:function(e,t){var n=yl();return t=void 0===t?null:t,e=e(),n.memoizedState=[e,t],e},useReducer:function(e,t,n){var r=yl();return t=void 0!==n?n(t):t,r.memoizedState=r.baseState=t,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:e,lastRenderedState:t},r.queue=e,e=e.dispatch=Kl.bind(null,il,e),[r.memoizedState,e]},useRef:function(e){return e={current:e},yl().memoizedState=e},useState:_l,useDebugValue:Bl,useDeferredValue:function(e){return yl().memoizedState=e},useTransition:function(){var e=_l(!1),t=e[0];return e=Vl.bind(null,e[1]),yl().memoizedState=e,[t,e]},useMutableSource:function(){},useSyncExternalStore:function(e,t,n){var r=il,a=yl();if(ao){if(void 0===n)throw Error(o(407));n=n()}else{if(n=t(),null===js)throw Error(o(349));0!==(30&ll)||Cl(r,t,n)}a.memoizedState=n;var l={value:n,getSnapshot:t};return a.queue=l,Dl(Ol.bind(null,r,l,e),[e]),r.flags|=2048,Tl(9,Nl.bind(null,r,l,n,t),void 0,null),n},useId:function(){var e=yl(),t=js.identifierPrefix;if(ao){var n=Ja;t=":"+t+"R"+(n=(Ga&~(1<<32-lt(Ga)-1)).toString(32)+n),0<(n=fl++)&&(t+="H"+n.toString(32)),t+=":"}else t=":"+t+"r"+(n=pl++).toString(32)+":";return e.memoizedState=t},unstable_isNewReconciler:!1},ei={readContext:To,useCallback:$l,useContext:To,useEffect:Fl,useImperativeHandle:Ml,useInsertionEffect:Ul,useLayoutEffect:zl,useMemo:Wl,useReducer:xl,useRef:Pl,useState:function(){return xl(wl)},useDebugValue:Bl,useDeferredValue:function(e){return Hl(bl(),sl.memoizedState,e)},useTransition:function(){return[xl(wl)[0],bl().memoizedState]},useMutableSource:Sl,useSyncExternalStore:kl,useId:ql,unstable_isNewReconciler:!1},ti={readContext:To,useCallback:$l,useContext:To,useEffect:Fl,useImperativeHandle:Ml,useInsertionEffect:Ul,useLayoutEffect:zl,useMemo:Wl,useReducer:El,useRef:Pl,useState:function(){return El(wl)},useDebugValue:Bl,useDeferredValue:function(e){var t=bl();return null===sl?t.memoizedState=e:Hl(t,sl.memoizedState,e)},useTransition:function(){return[El(wl)[0],bl().memoizedState]},useMutableSource:Sl,useSyncExternalStore:kl,useId:ql,unstable_isNewReconciler:!1};function ni(e,t){if(e&&e.defaultProps){for(var n in t=U({},t),e=e.defaultProps)void 0===t[n]&&(t[n]=e[n]);return t}return t}function ri(e,t,n,r){n=null===(n=n(r,t=e.memoizedState))||void 0===n?t:U({},t,n),e.memoizedState=n,0===e.lanes&&(e.updateQueue.baseState=n)}var ai={isMounted:function(e){return!!(e=e._reactInternals)&&$e(e)===e},enqueueSetState:function(e,t,n){e=e._reactInternals;var r=eu(),a=tu(e),o=Io(r,a);o.payload=t,void 0!==n&&null!==n&&(o.callback=n),null!==(t=Mo(e,o,a))&&(nu(t,e,a,r),Bo(t,e,a))},enqueueReplaceState:function(e,t,n){e=e._reactInternals;var r=eu(),a=tu(e),o=Io(r,a);o.tag=1,o.payload=t,void 0!==n&&null!==n&&(o.callback=n),null!==(t=Mo(e,o,a))&&(nu(t,e,a,r),Bo(t,e,a))},enqueueForceUpdate:function(e,t){e=e._reactInternals;var n=eu(),r=tu(e),a=Io(n,r);a.tag=2,void 0!==t&&null!==t&&(a.callback=t),null!==(t=Mo(e,a,r))&&(nu(t,e,r,n),Bo(t,e,r))}};function oi(e,t,n,r,a,o,l){return"function"===typeof(e=e.stateNode).shouldComponentUpdate?e.shouldComponentUpdate(r,o,l):!t.prototype||!t.prototype.isPureReactComponent||(!sr(n,r)||!sr(a,o))}function li(e,t,n){var r=!1,a=Oa,o=t.contextType;return"object"===typeof o&&null!==o?o=To(o):(a=Pa(t)?_a:Ra.current,o=(r=null!==(r=t.contextTypes)&&void 0!==r)?Ta(e,a):Oa),t=new t(n,o),e.memoizedState=null!==t.state&&void 0!==t.state?t.state:null,t.updater=ai,e.stateNode=t,t._reactInternals=e,r&&((e=e.stateNode).__reactInternalMemoizedUnmaskedChildContext=a,e.__reactInternalMemoizedMaskedChildContext=o),t}function ii(e,t,n,r){e=t.state,"function"===typeof t.componentWillReceiveProps&&t.componentWillReceiveProps(n,r),"function"===typeof t.UNSAFE_componentWillReceiveProps&&t.UNSAFE_componentWillReceiveProps(n,r),t.state!==e&&ai.enqueueReplaceState(t,t.state,null)}function si(e,t,n,r){var a=e.stateNode;a.props=n,a.state=e.memoizedState,a.refs={},Uo(e);var o=t.contextType;"object"===typeof o&&null!==o?a.context=To(o):(o=Pa(t)?_a:Ra.current,a.context=Ta(e,o)),a.state=e.memoizedState,"function"===typeof(o=t.getDerivedStateFromProps)&&(ri(e,t,o,n),a.state=e.memoizedState),"function"===typeof t.getDerivedStateFromProps||"function"===typeof a.getSnapshotBeforeUpdate||"function"!==typeof a.UNSAFE_componentWillMount&&"function"!==typeof a.componentWillMount||(t=a.state,"function"===typeof a.componentWillMount&&a.componentWillMount(),"function"===typeof a.UNSAFE_componentWillMount&&a.UNSAFE_componentWillMount(),t!==a.state&&ai.enqueueReplaceState(a,a.state,null),Wo(e,n,a,r),a.state=e.memoizedState),"function"===typeof a.componentDidMount&&(e.flags|=4194308)}function ui(e,t){try{var n="",r=t;do{n+=B(r),r=r.return}while(r);var a=n}catch(o){a="\nError generating stack: "+o.message+"\n"+o.stack}return{value:e,source:t,stack:a,digest:null}}function ci(e,t,n){return{value:e,source:null,stack:null!=n?n:null,digest:null!=t?t:null}}function di(e,t){try{console.error(t.value)}catch(n){setTimeout((function(){throw n}))}}var fi="function"===typeof WeakMap?WeakMap:Map;function pi(e,t,n){(n=Io(-1,n)).tag=3,n.payload={element:null};var r=t.value;return n.callback=function(){Hs||(Hs=!0,Vs=r),di(0,t)},n}function hi(e,t,n){(n=Io(-1,n)).tag=3;var r=e.type.getDerivedStateFromError;if("function"===typeof r){var a=t.value;n.payload=function(){return r(a)},n.callback=function(){di(0,t)}}var o=e.stateNode;return null!==o&&"function"===typeof o.componentDidCatch&&(n.callback=function(){di(0,t),"function"!==typeof r&&(null===qs?qs=new Set([this]):qs.add(this));var e=t.stack;this.componentDidCatch(t.value,{componentStack:null!==e?e:""})}),n}function mi(e,t,n){var r=e.pingCache;if(null===r){r=e.pingCache=new fi;var a=new Set;r.set(t,a)}else void 0===(a=r.get(t))&&(a=new Set,r.set(t,a));a.has(n)||(a.add(n),e=Cu.bind(null,e,t,n),t.then(e,e))}function vi(e){do{var t;if((t=13===e.tag)&&(t=null===(t=e.memoizedState)||null!==t.dehydrated),t)return e;e=e.return}while(null!==e);return null}function gi(e,t,n,r,a){return 0===(1&e.mode)?(e===t?e.flags|=65536:(e.flags|=128,n.flags|=131072,n.flags&=-52805,1===n.tag&&(null===n.alternate?n.tag=17:((t=Io(-1,1)).tag=2,Mo(n,t,1))),n.lanes|=1),e):(e.flags|=65536,e.lanes=a,e)}var yi=w.ReactCurrentOwner,bi=!1;function wi(e,t,n,r){t.child=null===e?Eo(t,null,n,r):xo(t,e.child,n,r)}function xi(e,t,n,r,a){n=n.render;var o=t.ref;return _o(t,a),r=vl(e,t,n,r,o,a),n=gl(),null===e||bi?(ao&&n&&eo(t),t.flags|=1,wi(e,t,r,a),t.child):(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~a,Hi(e,t,a))}function Ei(e,t,n,r,a){if(null===e){var o=n.type;return"function"!==typeof o||Pu(o)||void 0!==o.defaultProps||null!==n.compare||void 0!==n.defaultProps?((e=Au(n.type,null,r,t,t.mode,a)).ref=t.ref,e.return=t,t.child=e):(t.tag=15,t.type=o,Si(e,t,o,r,a))}if(o=e.child,0===(e.lanes&a)){var l=o.memoizedProps;if((n=null!==(n=n.compare)?n:sr)(l,r)&&e.ref===t.ref)return Hi(e,t,a)}return t.flags|=1,(e=Lu(o,r)).ref=t.ref,e.return=t,t.child=e}function Si(e,t,n,r,a){if(null!==e){var o=e.memoizedProps;if(sr(o,r)&&e.ref===t.ref){if(bi=!1,t.pendingProps=r=o,0===(e.lanes&a))return t.lanes=e.lanes,Hi(e,t,a);0!==(131072&e.flags)&&(bi=!0)}}return Ni(e,t,n,r,a)}function ki(e,t,n){var r=t.pendingProps,a=r.children,o=null!==e?e.memoizedState:null;if("hidden"===r.mode)if(0===(1&t.mode))t.memoizedState={baseLanes:0,cachePool:null,transitions:null},Na(Ls,Ps),Ps|=n;else{if(0===(1073741824&n))return e=null!==o?o.baseLanes|n:n,t.lanes=t.childLanes=1073741824,t.memoizedState={baseLanes:e,cachePool:null,transitions:null},t.updateQueue=null,Na(Ls,Ps),Ps|=e,null;t.memoizedState={baseLanes:0,cachePool:null,transitions:null},r=null!==o?o.baseLanes:n,Na(Ls,Ps),Ps|=r}else null!==o?(r=o.baseLanes|n,t.memoizedState=null):r=n,Na(Ls,Ps),Ps|=r;return wi(e,t,a,n),t.child}function Ci(e,t){var n=t.ref;(null===e&&null!==n||null!==e&&e.ref!==n)&&(t.flags|=512,t.flags|=2097152)}function Ni(e,t,n,r,a){var o=Pa(n)?_a:Ra.current;return o=Ta(t,o),_o(t,a),n=vl(e,t,n,r,o,a),r=gl(),null===e||bi?(ao&&r&&eo(t),t.flags|=1,wi(e,t,n,a),t.child):(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~a,Hi(e,t,a))}function Oi(e,t,n,r,a){if(Pa(n)){var o=!0;Fa(t)}else o=!1;if(_o(t,a),null===t.stateNode)Wi(e,t),li(t,n,r),si(t,n,r,a),r=!0;else if(null===e){var l=t.stateNode,i=t.memoizedProps;l.props=i;var s=l.context,u=n.contextType;"object"===typeof u&&null!==u?u=To(u):u=Ta(t,u=Pa(n)?_a:Ra.current);var c=n.getDerivedStateFromProps,d="function"===typeof c||"function"===typeof l.getSnapshotBeforeUpdate;d||"function"!==typeof l.UNSAFE_componentWillReceiveProps&&"function"!==typeof l.componentWillReceiveProps||(i!==r||s!==u)&&ii(t,l,r,u),Fo=!1;var f=t.memoizedState;l.state=f,Wo(t,r,l,a),s=t.memoizedState,i!==r||f!==s||ja.current||Fo?("function"===typeof c&&(ri(t,n,c,r),s=t.memoizedState),(i=Fo||oi(t,n,i,r,f,s,u))?(d||"function"!==typeof l.UNSAFE_componentWillMount&&"function"!==typeof l.componentWillMount||("function"===typeof l.componentWillMount&&l.componentWillMount(),"function"===typeof l.UNSAFE_componentWillMount&&l.UNSAFE_componentWillMount()),"function"===typeof l.componentDidMount&&(t.flags|=4194308)):("function"===typeof l.componentDidMount&&(t.flags|=4194308),t.memoizedProps=r,t.memoizedState=s),l.props=r,l.state=s,l.context=u,r=i):("function"===typeof l.componentDidMount&&(t.flags|=4194308),r=!1)}else{l=t.stateNode,zo(e,t),i=t.memoizedProps,u=t.type===t.elementType?i:ni(t.type,i),l.props=u,d=t.pendingProps,f=l.context,"object"===typeof(s=n.contextType)&&null!==s?s=To(s):s=Ta(t,s=Pa(n)?_a:Ra.current);var p=n.getDerivedStateFromProps;(c="function"===typeof p||"function"===typeof l.getSnapshotBeforeUpdate)||"function"!==typeof l.UNSAFE_componentWillReceiveProps&&"function"!==typeof l.componentWillReceiveProps||(i!==d||f!==s)&&ii(t,l,r,s),Fo=!1,f=t.memoizedState,l.state=f,Wo(t,r,l,a);var h=t.memoizedState;i!==d||f!==h||ja.current||Fo?("function"===typeof p&&(ri(t,n,p,r),h=t.memoizedState),(u=Fo||oi(t,n,u,r,f,h,s)||!1)?(c||"function"!==typeof l.UNSAFE_componentWillUpdate&&"function"!==typeof l.componentWillUpdate||("function"===typeof l.componentWillUpdate&&l.componentWillUpdate(r,h,s),"function"===typeof l.UNSAFE_componentWillUpdate&&l.UNSAFE_componentWillUpdate(r,h,s)),"function"===typeof l.componentDidUpdate&&(t.flags|=4),"function"===typeof l.getSnapshotBeforeUpdate&&(t.flags|=1024)):("function"!==typeof l.componentDidUpdate||i===e.memoizedProps&&f===e.memoizedState||(t.flags|=4),"function"!==typeof l.getSnapshotBeforeUpdate||i===e.memoizedProps&&f===e.memoizedState||(t.flags|=1024),t.memoizedProps=r,t.memoizedState=h),l.props=r,l.state=h,l.context=s,r=u):("function"!==typeof l.componentDidUpdate||i===e.memoizedProps&&f===e.memoizedState||(t.flags|=4),"function"!==typeof l.getSnapshotBeforeUpdate||i===e.memoizedProps&&f===e.memoizedState||(t.flags|=1024),r=!1)}return Ri(e,t,n,r,o,a)}function Ri(e,t,n,r,a,o){Ci(e,t);var l=0!==(128&t.flags);if(!r&&!l)return a&&Ua(t,n,!1),Hi(e,t,o);r=t.stateNode,yi.current=t;var i=l&&"function"!==typeof n.getDerivedStateFromError?null:r.render();return t.flags|=1,null!==e&&l?(t.child=xo(t,e.child,null,o),t.child=xo(t,null,i,o)):wi(e,t,i,o),t.memoizedState=r.state,a&&Ua(t,n,!0),t.child}function ji(e){var t=e.stateNode;t.pendingContext?Aa(0,t.pendingContext,t.pendingContext!==t.context):t.context&&Aa(0,t.context,!1),Go(e,t.containerInfo)}function _i(e,t,n,r,a){return ho(),mo(a),t.flags|=256,wi(e,t,n,r),t.child}var Ti,Pi,Li,Ai,Di={dehydrated:null,treeContext:null,retryLane:0};function Fi(e){return{baseLanes:e,cachePool:null,transitions:null}}function Ui(e,t,n){var r,a=t.pendingProps,l=el.current,i=!1,s=0!==(128&t.flags);if((r=s)||(r=(null===e||null!==e.memoizedState)&&0!==(2&l)),r?(i=!0,t.flags&=-129):null!==e&&null===e.memoizedState||(l|=1),Na(el,1&l),null===e)return uo(t),null!==(e=t.memoizedState)&&null!==(e=e.dehydrated)?(0===(1&t.mode)?t.lanes=1:"$!"===e.data?t.lanes=8:t.lanes=1073741824,null):(s=a.children,e=a.fallback,i?(a=t.mode,i=t.child,s={mode:"hidden",children:s},0===(1&a)&&null!==i?(i.childLanes=0,i.pendingProps=s):i=Fu(s,a,0,null),e=Du(e,a,n,null),i.return=t,e.return=t,i.sibling=e,t.child=i,t.child.memoizedState=Fi(n),t.memoizedState=Di,e):zi(t,s));if(null!==(l=e.memoizedState)&&null!==(r=l.dehydrated))return function(e,t,n,r,a,l,i){if(n)return 256&t.flags?(t.flags&=-257,Ii(e,t,i,r=ci(Error(o(422))))):null!==t.memoizedState?(t.child=e.child,t.flags|=128,null):(l=r.fallback,a=t.mode,r=Fu({mode:"visible",children:r.children},a,0,null),(l=Du(l,a,i,null)).flags|=2,r.return=t,l.return=t,r.sibling=l,t.child=r,0!==(1&t.mode)&&xo(t,e.child,null,i),t.child.memoizedState=Fi(i),t.memoizedState=Di,l);if(0===(1&t.mode))return Ii(e,t,i,null);if("$!"===a.data){if(r=a.nextSibling&&a.nextSibling.dataset)var s=r.dgst;return r=s,Ii(e,t,i,r=ci(l=Error(o(419)),r,void 0))}if(s=0!==(i&e.childLanes),bi||s){if(null!==(r=js)){switch(i&-i){case 4:a=2;break;case 16:a=8;break;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:a=32;break;case 536870912:a=268435456;break;default:a=0}0!==(a=0!==(a&(r.suspendedLanes|i))?0:a)&&a!==l.retryLane&&(l.retryLane=a,Do(e,a),nu(r,e,a,-1))}return mu(),Ii(e,t,i,r=ci(Error(o(421))))}return"$?"===a.data?(t.flags|=128,t.child=e.child,t=Ou.bind(null,e),a._reactRetry=t,null):(e=l.treeContext,ro=ua(a.nextSibling),no=t,ao=!0,oo=null,null!==e&&(Ka[Qa++]=Ga,Ka[Qa++]=Ja,Ka[Qa++]=Xa,Ga=e.id,Ja=e.overflow,Xa=t),t=zi(t,r.children),t.flags|=4096,t)}(e,t,s,a,r,l,n);if(i){i=a.fallback,s=t.mode,r=(l=e.child).sibling;var u={mode:"hidden",children:a.children};return 0===(1&s)&&t.child!==l?((a=t.child).childLanes=0,a.pendingProps=u,t.deletions=null):(a=Lu(l,u)).subtreeFlags=14680064&l.subtreeFlags,null!==r?i=Lu(r,i):(i=Du(i,s,n,null)).flags|=2,i.return=t,a.return=t,a.sibling=i,t.child=a,a=i,i=t.child,s=null===(s=e.child.memoizedState)?Fi(n):{baseLanes:s.baseLanes|n,cachePool:null,transitions:s.transitions},i.memoizedState=s,i.childLanes=e.childLanes&~n,t.memoizedState=Di,a}return e=(i=e.child).sibling,a=Lu(i,{mode:"visible",children:a.children}),0===(1&t.mode)&&(a.lanes=n),a.return=t,a.sibling=null,null!==e&&(null===(n=t.deletions)?(t.deletions=[e],t.flags|=16):n.push(e)),t.child=a,t.memoizedState=null,a}function zi(e,t){return(t=Fu({mode:"visible",children:t},e.mode,0,null)).return=e,e.child=t}function Ii(e,t,n,r){return null!==r&&mo(r),xo(t,e.child,null,n),(e=zi(t,t.pendingProps.children)).flags|=2,t.memoizedState=null,e}function Mi(e,t,n){e.lanes|=t;var r=e.alternate;null!==r&&(r.lanes|=t),jo(e.return,t,n)}function Bi(e,t,n,r,a){var o=e.memoizedState;null===o?e.memoizedState={isBackwards:t,rendering:null,renderingStartTime:0,last:r,tail:n,tailMode:a}:(o.isBackwards=t,o.rendering=null,o.renderingStartTime=0,o.last=r,o.tail=n,o.tailMode=a)}function $i(e,t,n){var r=t.pendingProps,a=r.revealOrder,o=r.tail;if(wi(e,t,r.children,n),0!==(2&(r=el.current)))r=1&r|2,t.flags|=128;else{if(null!==e&&0!==(128&e.flags))e:for(e=t.child;null!==e;){if(13===e.tag)null!==e.memoizedState&&Mi(e,n,t);else if(19===e.tag)Mi(e,n,t);else if(null!==e.child){e.child.return=e,e=e.child;continue}if(e===t)break e;for(;null===e.sibling;){if(null===e.return||e.return===t)break e;e=e.return}e.sibling.return=e.return,e=e.sibling}r&=1}if(Na(el,r),0===(1&t.mode))t.memoizedState=null;else switch(a){case"forwards":for(n=t.child,a=null;null!==n;)null!==(e=n.alternate)&&null===tl(e)&&(a=n),n=n.sibling;null===(n=a)?(a=t.child,t.child=null):(a=n.sibling,n.sibling=null),Bi(t,!1,a,n,o);break;case"backwards":for(n=null,a=t.child,t.child=null;null!==a;){if(null!==(e=a.alternate)&&null===tl(e)){t.child=a;break}e=a.sibling,a.sibling=n,n=a,a=e}Bi(t,!0,n,null,o);break;case"together":Bi(t,!1,null,null,void 0);break;default:t.memoizedState=null}return t.child}function Wi(e,t){0===(1&t.mode)&&null!==e&&(e.alternate=null,t.alternate=null,t.flags|=2)}function Hi(e,t,n){if(null!==e&&(t.dependencies=e.dependencies),Fs|=t.lanes,0===(n&t.childLanes))return null;if(null!==e&&t.child!==e.child)throw Error(o(153));if(null!==t.child){for(n=Lu(e=t.child,e.pendingProps),t.child=n,n.return=t;null!==e.sibling;)e=e.sibling,(n=n.sibling=Lu(e,e.pendingProps)).return=t;n.sibling=null}return t.child}function Vi(e,t){if(!ao)switch(e.tailMode){case"hidden":t=e.tail;for(var n=null;null!==t;)null!==t.alternate&&(n=t),t=t.sibling;null===n?e.tail=null:n.sibling=null;break;case"collapsed":n=e.tail;for(var r=null;null!==n;)null!==n.alternate&&(r=n),n=n.sibling;null===r?t||null===e.tail?e.tail=null:e.tail.sibling=null:r.sibling=null}}function qi(e){var t=null!==e.alternate&&e.alternate.child===e.child,n=0,r=0;if(t)for(var a=e.child;null!==a;)n|=a.lanes|a.childLanes,r|=14680064&a.subtreeFlags,r|=14680064&a.flags,a.return=e,a=a.sibling;else for(a=e.child;null!==a;)n|=a.lanes|a.childLanes,r|=a.subtreeFlags,r|=a.flags,a.return=e,a=a.sibling;return e.subtreeFlags|=r,e.childLanes=n,t}function Ki(e,t,n){var r=t.pendingProps;switch(to(t),t.tag){case 2:case 16:case 15:case 0:case 11:case 7:case 8:case 12:case 9:case 14:return qi(t),null;case 1:case 17:return Pa(t.type)&&La(),qi(t),null;case 3:return r=t.stateNode,Jo(),Ca(ja),Ca(Ra),rl(),r.pendingContext&&(r.context=r.pendingContext,r.pendingContext=null),null!==e&&null!==e.child||(fo(t)?t.flags|=4:null===e||e.memoizedState.isDehydrated&&0===(256&t.flags)||(t.flags|=1024,null!==oo&&(lu(oo),oo=null))),Pi(e,t),qi(t),null;case 5:Zo(t);var a=Xo(Qo.current);if(n=t.type,null!==e&&null!=t.stateNode)Li(e,t,n,r,a),e.ref!==t.ref&&(t.flags|=512,t.flags|=2097152);else{if(!r){if(null===t.stateNode)throw Error(o(166));return qi(t),null}if(e=Xo(qo.current),fo(t)){r=t.stateNode,n=t.type;var l=t.memoizedProps;switch(r[fa]=t,r[pa]=l,e=0!==(1&t.mode),n){case"dialog":Ir("cancel",r),Ir("close",r);break;case"iframe":case"object":case"embed":Ir("load",r);break;case"video":case"audio":for(a=0;a<Dr.length;a++)Ir(Dr[a],r);break;case"source":Ir("error",r);break;case"img":case"image":case"link":Ir("error",r),Ir("load",r);break;case"details":Ir("toggle",r);break;case"input":G(r,l),Ir("invalid",r);break;case"select":r._wrapperState={wasMultiple:!!l.multiple},Ir("invalid",r);break;case"textarea":ae(r,l),Ir("invalid",r)}for(var s in ye(n,l),a=null,l)if(l.hasOwnProperty(s)){var u=l[s];"children"===s?"string"===typeof u?r.textContent!==u&&(!0!==l.suppressHydrationWarning&&Yr(r.textContent,u,e),a=["children",u]):"number"===typeof u&&r.textContent!==""+u&&(!0!==l.suppressHydrationWarning&&Yr(r.textContent,u,e),a=["children",""+u]):i.hasOwnProperty(s)&&null!=u&&"onScroll"===s&&Ir("scroll",r)}switch(n){case"input":q(r),Z(r,l,!0);break;case"textarea":q(r),le(r);break;case"select":case"option":break;default:"function"===typeof l.onClick&&(r.onclick=Zr)}r=a,t.updateQueue=r,null!==r&&(t.flags|=4)}else{s=9===a.nodeType?a:a.ownerDocument,"http://www.w3.org/1999/xhtml"===e&&(e=ie(n)),"http://www.w3.org/1999/xhtml"===e?"script"===n?((e=s.createElement("div")).innerHTML="<script><\/script>",e=e.removeChild(e.firstChild)):"string"===typeof r.is?e=s.createElement(n,{is:r.is}):(e=s.createElement(n),"select"===n&&(s=e,r.multiple?s.multiple=!0:r.size&&(s.size=r.size))):e=s.createElementNS(e,n),e[fa]=t,e[pa]=r,Ti(e,t,!1,!1),t.stateNode=e;e:{switch(s=be(n,r),n){case"dialog":Ir("cancel",e),Ir("close",e),a=r;break;case"iframe":case"object":case"embed":Ir("load",e),a=r;break;case"video":case"audio":for(a=0;a<Dr.length;a++)Ir(Dr[a],e);a=r;break;case"source":Ir("error",e),a=r;break;case"img":case"image":case"link":Ir("error",e),Ir("load",e),a=r;break;case"details":Ir("toggle",e),a=r;break;case"input":G(e,r),a=X(e,r),Ir("invalid",e);break;case"option":default:a=r;break;case"select":e._wrapperState={wasMultiple:!!r.multiple},a=U({},r,{value:void 0}),Ir("invalid",e);break;case"textarea":ae(e,r),a=re(e,r),Ir("invalid",e)}for(l in ye(n,a),u=a)if(u.hasOwnProperty(l)){var c=u[l];"style"===l?ve(e,c):"dangerouslySetInnerHTML"===l?null!=(c=c?c.__html:void 0)&&de(e,c):"children"===l?"string"===typeof c?("textarea"!==n||""!==c)&&fe(e,c):"number"===typeof c&&fe(e,""+c):"suppressContentEditableWarning"!==l&&"suppressHydrationWarning"!==l&&"autoFocus"!==l&&(i.hasOwnProperty(l)?null!=c&&"onScroll"===l&&Ir("scroll",e):null!=c&&b(e,l,c,s))}switch(n){case"input":q(e),Z(e,r,!1);break;case"textarea":q(e),le(e);break;case"option":null!=r.value&&e.setAttribute("value",""+H(r.value));break;case"select":e.multiple=!!r.multiple,null!=(l=r.value)?ne(e,!!r.multiple,l,!1):null!=r.defaultValue&&ne(e,!!r.multiple,r.defaultValue,!0);break;default:"function"===typeof a.onClick&&(e.onclick=Zr)}switch(n){case"button":case"input":case"select":case"textarea":r=!!r.autoFocus;break e;case"img":r=!0;break e;default:r=!1}}r&&(t.flags|=4)}null!==t.ref&&(t.flags|=512,t.flags|=2097152)}return qi(t),null;case 6:if(e&&null!=t.stateNode)Ai(e,t,e.memoizedProps,r);else{if("string"!==typeof r&&null===t.stateNode)throw Error(o(166));if(n=Xo(Qo.current),Xo(qo.current),fo(t)){if(r=t.stateNode,n=t.memoizedProps,r[fa]=t,(l=r.nodeValue!==n)&&null!==(e=no))switch(e.tag){case 3:Yr(r.nodeValue,n,0!==(1&e.mode));break;case 5:!0!==e.memoizedProps.suppressHydrationWarning&&Yr(r.nodeValue,n,0!==(1&e.mode))}l&&(t.flags|=4)}else(r=(9===n.nodeType?n:n.ownerDocument).createTextNode(r))[fa]=t,t.stateNode=r}return qi(t),null;case 13:if(Ca(el),r=t.memoizedState,null===e||null!==e.memoizedState&&null!==e.memoizedState.dehydrated){if(ao&&null!==ro&&0!==(1&t.mode)&&0===(128&t.flags))po(),ho(),t.flags|=98560,l=!1;else if(l=fo(t),null!==r&&null!==r.dehydrated){if(null===e){if(!l)throw Error(o(318));if(!(l=null!==(l=t.memoizedState)?l.dehydrated:null))throw Error(o(317));l[fa]=t}else ho(),0===(128&t.flags)&&(t.memoizedState=null),t.flags|=4;qi(t),l=!1}else null!==oo&&(lu(oo),oo=null),l=!0;if(!l)return 65536&t.flags?t:null}return 0!==(128&t.flags)?(t.lanes=n,t):((r=null!==r)!==(null!==e&&null!==e.memoizedState)&&r&&(t.child.flags|=8192,0!==(1&t.mode)&&(null===e||0!==(1&el.current)?0===As&&(As=3):mu())),null!==t.updateQueue&&(t.flags|=4),qi(t),null);case 4:return Jo(),Pi(e,t),null===e&&$r(t.stateNode.containerInfo),qi(t),null;case 10:return Ro(t.type._context),qi(t),null;case 19:if(Ca(el),null===(l=t.memoizedState))return qi(t),null;if(r=0!==(128&t.flags),null===(s=l.rendering))if(r)Vi(l,!1);else{if(0!==As||null!==e&&0!==(128&e.flags))for(e=t.child;null!==e;){if(null!==(s=tl(e))){for(t.flags|=128,Vi(l,!1),null!==(r=s.updateQueue)&&(t.updateQueue=r,t.flags|=4),t.subtreeFlags=0,r=n,n=t.child;null!==n;)e=r,(l=n).flags&=14680066,null===(s=l.alternate)?(l.childLanes=0,l.lanes=e,l.child=null,l.subtreeFlags=0,l.memoizedProps=null,l.memoizedState=null,l.updateQueue=null,l.dependencies=null,l.stateNode=null):(l.childLanes=s.childLanes,l.lanes=s.lanes,l.child=s.child,l.subtreeFlags=0,l.deletions=null,l.memoizedProps=s.memoizedProps,l.memoizedState=s.memoizedState,l.updateQueue=s.updateQueue,l.type=s.type,e=s.dependencies,l.dependencies=null===e?null:{lanes:e.lanes,firstContext:e.firstContext}),n=n.sibling;return Na(el,1&el.current|2),t.child}e=e.sibling}null!==l.tail&&Je()>$s&&(t.flags|=128,r=!0,Vi(l,!1),t.lanes=4194304)}else{if(!r)if(null!==(e=tl(s))){if(t.flags|=128,r=!0,null!==(n=e.updateQueue)&&(t.updateQueue=n,t.flags|=4),Vi(l,!0),null===l.tail&&"hidden"===l.tailMode&&!s.alternate&&!ao)return qi(t),null}else 2*Je()-l.renderingStartTime>$s&&1073741824!==n&&(t.flags|=128,r=!0,Vi(l,!1),t.lanes=4194304);l.isBackwards?(s.sibling=t.child,t.child=s):(null!==(n=l.last)?n.sibling=s:t.child=s,l.last=s)}return null!==l.tail?(t=l.tail,l.rendering=t,l.tail=t.sibling,l.renderingStartTime=Je(),t.sibling=null,n=el.current,Na(el,r?1&n|2:1&n),t):(qi(t),null);case 22:case 23:return du(),r=null!==t.memoizedState,null!==e&&null!==e.memoizedState!==r&&(t.flags|=8192),r&&0!==(1&t.mode)?0!==(1073741824&Ps)&&(qi(t),6&t.subtreeFlags&&(t.flags|=8192)):qi(t),null;case 24:case 25:return null}throw Error(o(156,t.tag))}function Qi(e,t){switch(to(t),t.tag){case 1:return Pa(t.type)&&La(),65536&(e=t.flags)?(t.flags=-65537&e|128,t):null;case 3:return Jo(),Ca(ja),Ca(Ra),rl(),0!==(65536&(e=t.flags))&&0===(128&e)?(t.flags=-65537&e|128,t):null;case 5:return Zo(t),null;case 13:if(Ca(el),null!==(e=t.memoizedState)&&null!==e.dehydrated){if(null===t.alternate)throw Error(o(340));ho()}return 65536&(e=t.flags)?(t.flags=-65537&e|128,t):null;case 19:return Ca(el),null;case 4:return Jo(),null;case 10:return Ro(t.type._context),null;case 22:case 23:return du(),null;default:return null}}Ti=function(e,t){for(var n=t.child;null!==n;){if(5===n.tag||6===n.tag)e.appendChild(n.stateNode);else if(4!==n.tag&&null!==n.child){n.child.return=n,n=n.child;continue}if(n===t)break;for(;null===n.sibling;){if(null===n.return||n.return===t)return;n=n.return}n.sibling.return=n.return,n=n.sibling}},Pi=function(){},Li=function(e,t,n,r){var a=e.memoizedProps;if(a!==r){e=t.stateNode,Xo(qo.current);var o,l=null;switch(n){case"input":a=X(e,a),r=X(e,r),l=[];break;case"select":a=U({},a,{value:void 0}),r=U({},r,{value:void 0}),l=[];break;case"textarea":a=re(e,a),r=re(e,r),l=[];break;default:"function"!==typeof a.onClick&&"function"===typeof r.onClick&&(e.onclick=Zr)}for(c in ye(n,r),n=null,a)if(!r.hasOwnProperty(c)&&a.hasOwnProperty(c)&&null!=a[c])if("style"===c){var s=a[c];for(o in s)s.hasOwnProperty(o)&&(n||(n={}),n[o]="")}else"dangerouslySetInnerHTML"!==c&&"children"!==c&&"suppressContentEditableWarning"!==c&&"suppressHydrationWarning"!==c&&"autoFocus"!==c&&(i.hasOwnProperty(c)?l||(l=[]):(l=l||[]).push(c,null));for(c in r){var u=r[c];if(s=null!=a?a[c]:void 0,r.hasOwnProperty(c)&&u!==s&&(null!=u||null!=s))if("style"===c)if(s){for(o in s)!s.hasOwnProperty(o)||u&&u.hasOwnProperty(o)||(n||(n={}),n[o]="");for(o in u)u.hasOwnProperty(o)&&s[o]!==u[o]&&(n||(n={}),n[o]=u[o])}else n||(l||(l=[]),l.push(c,n)),n=u;else"dangerouslySetInnerHTML"===c?(u=u?u.__html:void 0,s=s?s.__html:void 0,null!=u&&s!==u&&(l=l||[]).push(c,u)):"children"===c?"string"!==typeof u&&"number"!==typeof u||(l=l||[]).push(c,""+u):"suppressContentEditableWarning"!==c&&"suppressHydrationWarning"!==c&&(i.hasOwnProperty(c)?(null!=u&&"onScroll"===c&&Ir("scroll",e),l||s===u||(l=[])):(l=l||[]).push(c,u))}n&&(l=l||[]).push("style",n);var c=l;(t.updateQueue=c)&&(t.flags|=4)}},Ai=function(e,t,n,r){n!==r&&(t.flags|=4)};var Xi=!1,Gi=!1,Ji="function"===typeof WeakSet?WeakSet:Set,Yi=null;function Zi(e,t){var n=e.ref;if(null!==n)if("function"===typeof n)try{n(null)}catch(r){ku(e,t,r)}else n.current=null}function es(e,t,n){try{n()}catch(r){ku(e,t,r)}}var ts=!1;function ns(e,t,n){var r=t.updateQueue;if(null!==(r=null!==r?r.lastEffect:null)){var a=r=r.next;do{if((a.tag&e)===e){var o=a.destroy;a.destroy=void 0,void 0!==o&&es(t,n,o)}a=a.next}while(a!==r)}}function rs(e,t){if(null!==(t=null!==(t=t.updateQueue)?t.lastEffect:null)){var n=t=t.next;do{if((n.tag&e)===e){var r=n.create;n.destroy=r()}n=n.next}while(n!==t)}}function as(e){var t=e.ref;if(null!==t){var n=e.stateNode;e.tag,e=n,"function"===typeof t?t(e):t.current=e}}function os(e){var t=e.alternate;null!==t&&(e.alternate=null,os(t)),e.child=null,e.deletions=null,e.sibling=null,5===e.tag&&(null!==(t=e.stateNode)&&(delete t[fa],delete t[pa],delete t[ma],delete t[va],delete t[ga])),e.stateNode=null,e.return=null,e.dependencies=null,e.memoizedProps=null,e.memoizedState=null,e.pendingProps=null,e.stateNode=null,e.updateQueue=null}function ls(e){return 5===e.tag||3===e.tag||4===e.tag}function is(e){e:for(;;){for(;null===e.sibling;){if(null===e.return||ls(e.return))return null;e=e.return}for(e.sibling.return=e.return,e=e.sibling;5!==e.tag&&6!==e.tag&&18!==e.tag;){if(2&e.flags)continue e;if(null===e.child||4===e.tag)continue e;e.child.return=e,e=e.child}if(!(2&e.flags))return e.stateNode}}function ss(e,t,n){var r=e.tag;if(5===r||6===r)e=e.stateNode,t?8===n.nodeType?n.parentNode.insertBefore(e,t):n.insertBefore(e,t):(8===n.nodeType?(t=n.parentNode).insertBefore(e,n):(t=n).appendChild(e),null!==(n=n._reactRootContainer)&&void 0!==n||null!==t.onclick||(t.onclick=Zr));else if(4!==r&&null!==(e=e.child))for(ss(e,t,n),e=e.sibling;null!==e;)ss(e,t,n),e=e.sibling}function us(e,t,n){var r=e.tag;if(5===r||6===r)e=e.stateNode,t?n.insertBefore(e,t):n.appendChild(e);else if(4!==r&&null!==(e=e.child))for(us(e,t,n),e=e.sibling;null!==e;)us(e,t,n),e=e.sibling}var cs=null,ds=!1;function fs(e,t,n){for(n=n.child;null!==n;)ps(e,t,n),n=n.sibling}function ps(e,t,n){if(ot&&"function"===typeof ot.onCommitFiberUnmount)try{ot.onCommitFiberUnmount(at,n)}catch(i){}switch(n.tag){case 5:Gi||Zi(n,t);case 6:var r=cs,a=ds;cs=null,fs(e,t,n),ds=a,null!==(cs=r)&&(ds?(e=cs,n=n.stateNode,8===e.nodeType?e.parentNode.removeChild(n):e.removeChild(n)):cs.removeChild(n.stateNode));break;case 18:null!==cs&&(ds?(e=cs,n=n.stateNode,8===e.nodeType?sa(e.parentNode,n):1===e.nodeType&&sa(e,n),$t(e)):sa(cs,n.stateNode));break;case 4:r=cs,a=ds,cs=n.stateNode.containerInfo,ds=!0,fs(e,t,n),cs=r,ds=a;break;case 0:case 11:case 14:case 15:if(!Gi&&(null!==(r=n.updateQueue)&&null!==(r=r.lastEffect))){a=r=r.next;do{var o=a,l=o.destroy;o=o.tag,void 0!==l&&(0!==(2&o)||0!==(4&o))&&es(n,t,l),a=a.next}while(a!==r)}fs(e,t,n);break;case 1:if(!Gi&&(Zi(n,t),"function"===typeof(r=n.stateNode).componentWillUnmount))try{r.props=n.memoizedProps,r.state=n.memoizedState,r.componentWillUnmount()}catch(i){ku(n,t,i)}fs(e,t,n);break;case 21:fs(e,t,n);break;case 22:1&n.mode?(Gi=(r=Gi)||null!==n.memoizedState,fs(e,t,n),Gi=r):fs(e,t,n);break;default:fs(e,t,n)}}function hs(e){var t=e.updateQueue;if(null!==t){e.updateQueue=null;var n=e.stateNode;null===n&&(n=e.stateNode=new Ji),t.forEach((function(t){var r=Ru.bind(null,e,t);n.has(t)||(n.add(t),t.then(r,r))}))}}function ms(e,t){var n=t.deletions;if(null!==n)for(var r=0;r<n.length;r++){var a=n[r];try{var l=e,i=t,s=i;e:for(;null!==s;){switch(s.tag){case 5:cs=s.stateNode,ds=!1;break e;case 3:case 4:cs=s.stateNode.containerInfo,ds=!0;break e}s=s.return}if(null===cs)throw Error(o(160));ps(l,i,a),cs=null,ds=!1;var u=a.alternate;null!==u&&(u.return=null),a.return=null}catch(c){ku(a,t,c)}}if(12854&t.subtreeFlags)for(t=t.child;null!==t;)vs(t,e),t=t.sibling}function vs(e,t){var n=e.alternate,r=e.flags;switch(e.tag){case 0:case 11:case 14:case 15:if(ms(t,e),gs(e),4&r){try{ns(3,e,e.return),rs(3,e)}catch(v){ku(e,e.return,v)}try{ns(5,e,e.return)}catch(v){ku(e,e.return,v)}}break;case 1:ms(t,e),gs(e),512&r&&null!==n&&Zi(n,n.return);break;case 5:if(ms(t,e),gs(e),512&r&&null!==n&&Zi(n,n.return),32&e.flags){var a=e.stateNode;try{fe(a,"")}catch(v){ku(e,e.return,v)}}if(4&r&&null!=(a=e.stateNode)){var l=e.memoizedProps,i=null!==n?n.memoizedProps:l,s=e.type,u=e.updateQueue;if(e.updateQueue=null,null!==u)try{"input"===s&&"radio"===l.type&&null!=l.name&&J(a,l),be(s,i);var c=be(s,l);for(i=0;i<u.length;i+=2){var d=u[i],f=u[i+1];"style"===d?ve(a,f):"dangerouslySetInnerHTML"===d?de(a,f):"children"===d?fe(a,f):b(a,d,f,c)}switch(s){case"input":Y(a,l);break;case"textarea":oe(a,l);break;case"select":var p=a._wrapperState.wasMultiple;a._wrapperState.wasMultiple=!!l.multiple;var h=l.value;null!=h?ne(a,!!l.multiple,h,!1):p!==!!l.multiple&&(null!=l.defaultValue?ne(a,!!l.multiple,l.defaultValue,!0):ne(a,!!l.multiple,l.multiple?[]:"",!1))}a[pa]=l}catch(v){ku(e,e.return,v)}}break;case 6:if(ms(t,e),gs(e),4&r){if(null===e.stateNode)throw Error(o(162));a=e.stateNode,l=e.memoizedProps;try{a.nodeValue=l}catch(v){ku(e,e.return,v)}}break;case 3:if(ms(t,e),gs(e),4&r&&null!==n&&n.memoizedState.isDehydrated)try{$t(t.containerInfo)}catch(v){ku(e,e.return,v)}break;case 4:default:ms(t,e),gs(e);break;case 13:ms(t,e),gs(e),8192&(a=e.child).flags&&(l=null!==a.memoizedState,a.stateNode.isHidden=l,!l||null!==a.alternate&&null!==a.alternate.memoizedState||(Bs=Je())),4&r&&hs(e);break;case 22:if(d=null!==n&&null!==n.memoizedState,1&e.mode?(Gi=(c=Gi)||d,ms(t,e),Gi=c):ms(t,e),gs(e),8192&r){if(c=null!==e.memoizedState,(e.stateNode.isHidden=c)&&!d&&0!==(1&e.mode))for(Yi=e,d=e.child;null!==d;){for(f=Yi=d;null!==Yi;){switch(h=(p=Yi).child,p.tag){case 0:case 11:case 14:case 15:ns(4,p,p.return);break;case 1:Zi(p,p.return);var m=p.stateNode;if("function"===typeof m.componentWillUnmount){r=p,n=p.return;try{t=r,m.props=t.memoizedProps,m.state=t.memoizedState,m.componentWillUnmount()}catch(v){ku(r,n,v)}}break;case 5:Zi(p,p.return);break;case 22:if(null!==p.memoizedState){xs(f);continue}}null!==h?(h.return=p,Yi=h):xs(f)}d=d.sibling}e:for(d=null,f=e;;){if(5===f.tag){if(null===d){d=f;try{a=f.stateNode,c?"function"===typeof(l=a.style).setProperty?l.setProperty("display","none","important"):l.display="none":(s=f.stateNode,i=void 0!==(u=f.memoizedProps.style)&&null!==u&&u.hasOwnProperty("display")?u.display:null,s.style.display=me("display",i))}catch(v){ku(e,e.return,v)}}}else if(6===f.tag){if(null===d)try{f.stateNode.nodeValue=c?"":f.memoizedProps}catch(v){ku(e,e.return,v)}}else if((22!==f.tag&&23!==f.tag||null===f.memoizedState||f===e)&&null!==f.child){f.child.return=f,f=f.child;continue}if(f===e)break e;for(;null===f.sibling;){if(null===f.return||f.return===e)break e;d===f&&(d=null),f=f.return}d===f&&(d=null),f.sibling.return=f.return,f=f.sibling}}break;case 19:ms(t,e),gs(e),4&r&&hs(e);case 21:}}function gs(e){var t=e.flags;if(2&t){try{e:{for(var n=e.return;null!==n;){if(ls(n)){var r=n;break e}n=n.return}throw Error(o(160))}switch(r.tag){case 5:var a=r.stateNode;32&r.flags&&(fe(a,""),r.flags&=-33),us(e,is(e),a);break;case 3:case 4:var l=r.stateNode.containerInfo;ss(e,is(e),l);break;default:throw Error(o(161))}}catch(i){ku(e,e.return,i)}e.flags&=-3}4096&t&&(e.flags&=-4097)}function ys(e,t,n){Yi=e,bs(e,t,n)}function bs(e,t,n){for(var r=0!==(1&e.mode);null!==Yi;){var a=Yi,o=a.child;if(22===a.tag&&r){var l=null!==a.memoizedState||Xi;if(!l){var i=a.alternate,s=null!==i&&null!==i.memoizedState||Gi;i=Xi;var u=Gi;if(Xi=l,(Gi=s)&&!u)for(Yi=a;null!==Yi;)s=(l=Yi).child,22===l.tag&&null!==l.memoizedState?Es(a):null!==s?(s.return=l,Yi=s):Es(a);for(;null!==o;)Yi=o,bs(o,t,n),o=o.sibling;Yi=a,Xi=i,Gi=u}ws(e)}else 0!==(8772&a.subtreeFlags)&&null!==o?(o.return=a,Yi=o):ws(e)}}function ws(e){for(;null!==Yi;){var t=Yi;if(0!==(8772&t.flags)){var n=t.alternate;try{if(0!==(8772&t.flags))switch(t.tag){case 0:case 11:case 15:Gi||rs(5,t);break;case 1:var r=t.stateNode;if(4&t.flags&&!Gi)if(null===n)r.componentDidMount();else{var a=t.elementType===t.type?n.memoizedProps:ni(t.type,n.memoizedProps);r.componentDidUpdate(a,n.memoizedState,r.__reactInternalSnapshotBeforeUpdate)}var l=t.updateQueue;null!==l&&Ho(t,l,r);break;case 3:var i=t.updateQueue;if(null!==i){if(n=null,null!==t.child)switch(t.child.tag){case 5:case 1:n=t.child.stateNode}Ho(t,i,n)}break;case 5:var s=t.stateNode;if(null===n&&4&t.flags){n=s;var u=t.memoizedProps;switch(t.type){case"button":case"input":case"select":case"textarea":u.autoFocus&&n.focus();break;case"img":u.src&&(n.src=u.src)}}break;case 6:case 4:case 12:case 19:case 17:case 21:case 22:case 23:case 25:break;case 13:if(null===t.memoizedState){var c=t.alternate;if(null!==c){var d=c.memoizedState;if(null!==d){var f=d.dehydrated;null!==f&&$t(f)}}}break;default:throw Error(o(163))}Gi||512&t.flags&&as(t)}catch(p){ku(t,t.return,p)}}if(t===e){Yi=null;break}if(null!==(n=t.sibling)){n.return=t.return,Yi=n;break}Yi=t.return}}function xs(e){for(;null!==Yi;){var t=Yi;if(t===e){Yi=null;break}var n=t.sibling;if(null!==n){n.return=t.return,Yi=n;break}Yi=t.return}}function Es(e){for(;null!==Yi;){var t=Yi;try{switch(t.tag){case 0:case 11:case 15:var n=t.return;try{rs(4,t)}catch(s){ku(t,n,s)}break;case 1:var r=t.stateNode;if("function"===typeof r.componentDidMount){var a=t.return;try{r.componentDidMount()}catch(s){ku(t,a,s)}}var o=t.return;try{as(t)}catch(s){ku(t,o,s)}break;case 5:var l=t.return;try{as(t)}catch(s){ku(t,l,s)}}}catch(s){ku(t,t.return,s)}if(t===e){Yi=null;break}var i=t.sibling;if(null!==i){i.return=t.return,Yi=i;break}Yi=t.return}}var Ss,ks=Math.ceil,Cs=w.ReactCurrentDispatcher,Ns=w.ReactCurrentOwner,Os=w.ReactCurrentBatchConfig,Rs=0,js=null,_s=null,Ts=0,Ps=0,Ls=ka(0),As=0,Ds=null,Fs=0,Us=0,zs=0,Is=null,Ms=null,Bs=0,$s=1/0,Ws=null,Hs=!1,Vs=null,qs=null,Ks=!1,Qs=null,Xs=0,Gs=0,Js=null,Ys=-1,Zs=0;function eu(){return 0!==(6&Rs)?Je():-1!==Ys?Ys:Ys=Je()}function tu(e){return 0===(1&e.mode)?1:0!==(2&Rs)&&0!==Ts?Ts&-Ts:null!==vo.transition?(0===Zs&&(Zs=mt()),Zs):0!==(e=bt)?e:e=void 0===(e=window.event)?16:Gt(e.type)}function nu(e,t,n,r){if(50<Gs)throw Gs=0,Js=null,Error(o(185));gt(e,n,r),0!==(2&Rs)&&e===js||(e===js&&(0===(2&Rs)&&(Us|=n),4===As&&iu(e,Ts)),ru(e,r),1===n&&0===Rs&&0===(1&t.mode)&&($s=Je()+500,Ia&&$a()))}function ru(e,t){var n=e.callbackNode;!function(e,t){for(var n=e.suspendedLanes,r=e.pingedLanes,a=e.expirationTimes,o=e.pendingLanes;0<o;){var l=31-lt(o),i=1<<l,s=a[l];-1===s?0!==(i&n)&&0===(i&r)||(a[l]=pt(i,t)):s<=t&&(e.expiredLanes|=i),o&=~i}}(e,t);var r=ft(e,e===js?Ts:0);if(0===r)null!==n&&Qe(n),e.callbackNode=null,e.callbackPriority=0;else if(t=r&-r,e.callbackPriority!==t){if(null!=n&&Qe(n),1===t)0===e.tag?function(e){Ia=!0,Ba(e)}(su.bind(null,e)):Ba(su.bind(null,e)),la((function(){0===(6&Rs)&&$a()})),n=null;else{switch(wt(r)){case 1:n=Ze;break;case 4:n=et;break;case 16:default:n=tt;break;case 536870912:n=rt}n=ju(n,au.bind(null,e))}e.callbackPriority=t,e.callbackNode=n}}function au(e,t){if(Ys=-1,Zs=0,0!==(6&Rs))throw Error(o(327));var n=e.callbackNode;if(Eu()&&e.callbackNode!==n)return null;var r=ft(e,e===js?Ts:0);if(0===r)return null;if(0!==(30&r)||0!==(r&e.expiredLanes)||t)t=vu(e,r);else{t=r;var a=Rs;Rs|=2;var l=hu();for(js===e&&Ts===t||(Ws=null,$s=Je()+500,fu(e,t));;)try{yu();break}catch(s){pu(e,s)}Oo(),Cs.current=l,Rs=a,null!==_s?t=0:(js=null,Ts=0,t=As)}if(0!==t){if(2===t&&(0!==(a=ht(e))&&(r=a,t=ou(e,a))),1===t)throw n=Ds,fu(e,0),iu(e,r),ru(e,Je()),n;if(6===t)iu(e,r);else{if(a=e.current.alternate,0===(30&r)&&!function(e){for(var t=e;;){if(16384&t.flags){var n=t.updateQueue;if(null!==n&&null!==(n=n.stores))for(var r=0;r<n.length;r++){var a=n[r],o=a.getSnapshot;a=a.value;try{if(!ir(o(),a))return!1}catch(i){return!1}}}if(n=t.child,16384&t.subtreeFlags&&null!==n)n.return=t,t=n;else{if(t===e)break;for(;null===t.sibling;){if(null===t.return||t.return===e)return!0;t=t.return}t.sibling.return=t.return,t=t.sibling}}return!0}(a)&&(2===(t=vu(e,r))&&(0!==(l=ht(e))&&(r=l,t=ou(e,l))),1===t))throw n=Ds,fu(e,0),iu(e,r),ru(e,Je()),n;switch(e.finishedWork=a,e.finishedLanes=r,t){case 0:case 1:throw Error(o(345));case 2:case 5:xu(e,Ms,Ws);break;case 3:if(iu(e,r),(130023424&r)===r&&10<(t=Bs+500-Je())){if(0!==ft(e,0))break;if(((a=e.suspendedLanes)&r)!==r){eu(),e.pingedLanes|=e.suspendedLanes&a;break}e.timeoutHandle=ra(xu.bind(null,e,Ms,Ws),t);break}xu(e,Ms,Ws);break;case 4:if(iu(e,r),(4194240&r)===r)break;for(t=e.eventTimes,a=-1;0<r;){var i=31-lt(r);l=1<<i,(i=t[i])>a&&(a=i),r&=~l}if(r=a,10<(r=(120>(r=Je()-r)?120:480>r?480:1080>r?1080:1920>r?1920:3e3>r?3e3:4320>r?4320:1960*ks(r/1960))-r)){e.timeoutHandle=ra(xu.bind(null,e,Ms,Ws),r);break}xu(e,Ms,Ws);break;default:throw Error(o(329))}}}return ru(e,Je()),e.callbackNode===n?au.bind(null,e):null}function ou(e,t){var n=Is;return e.current.memoizedState.isDehydrated&&(fu(e,t).flags|=256),2!==(e=vu(e,t))&&(t=Ms,Ms=n,null!==t&&lu(t)),e}function lu(e){null===Ms?Ms=e:Ms.push.apply(Ms,e)}function iu(e,t){for(t&=~zs,t&=~Us,e.suspendedLanes|=t,e.pingedLanes&=~t,e=e.expirationTimes;0<t;){var n=31-lt(t),r=1<<n;e[n]=-1,t&=~r}}function su(e){if(0!==(6&Rs))throw Error(o(327));Eu();var t=ft(e,0);if(0===(1&t))return ru(e,Je()),null;var n=vu(e,t);if(0!==e.tag&&2===n){var r=ht(e);0!==r&&(t=r,n=ou(e,r))}if(1===n)throw n=Ds,fu(e,0),iu(e,t),ru(e,Je()),n;if(6===n)throw Error(o(345));return e.finishedWork=e.current.alternate,e.finishedLanes=t,xu(e,Ms,Ws),ru(e,Je()),null}function uu(e,t){var n=Rs;Rs|=1;try{return e(t)}finally{0===(Rs=n)&&($s=Je()+500,Ia&&$a())}}function cu(e){null!==Qs&&0===Qs.tag&&0===(6&Rs)&&Eu();var t=Rs;Rs|=1;var n=Os.transition,r=bt;try{if(Os.transition=null,bt=1,e)return e()}finally{bt=r,Os.transition=n,0===(6&(Rs=t))&&$a()}}function du(){Ps=Ls.current,Ca(Ls)}function fu(e,t){e.finishedWork=null,e.finishedLanes=0;var n=e.timeoutHandle;if(-1!==n&&(e.timeoutHandle=-1,aa(n)),null!==_s)for(n=_s.return;null!==n;){var r=n;switch(to(r),r.tag){case 1:null!==(r=r.type.childContextTypes)&&void 0!==r&&La();break;case 3:Jo(),Ca(ja),Ca(Ra),rl();break;case 5:Zo(r);break;case 4:Jo();break;case 13:case 19:Ca(el);break;case 10:Ro(r.type._context);break;case 22:case 23:du()}n=n.return}if(js=e,_s=e=Lu(e.current,null),Ts=Ps=t,As=0,Ds=null,zs=Us=Fs=0,Ms=Is=null,null!==Po){for(t=0;t<Po.length;t++)if(null!==(r=(n=Po[t]).interleaved)){n.interleaved=null;var a=r.next,o=n.pending;if(null!==o){var l=o.next;o.next=a,r.next=l}n.pending=r}Po=null}return e}function pu(e,t){for(;;){var n=_s;try{if(Oo(),al.current=Yl,cl){for(var r=il.memoizedState;null!==r;){var a=r.queue;null!==a&&(a.pending=null),r=r.next}cl=!1}if(ll=0,ul=sl=il=null,dl=!1,fl=0,Ns.current=null,null===n||null===n.return){As=1,Ds=t,_s=null;break}e:{var l=e,i=n.return,s=n,u=t;if(t=Ts,s.flags|=32768,null!==u&&"object"===typeof u&&"function"===typeof u.then){var c=u,d=s,f=d.tag;if(0===(1&d.mode)&&(0===f||11===f||15===f)){var p=d.alternate;p?(d.updateQueue=p.updateQueue,d.memoizedState=p.memoizedState,d.lanes=p.lanes):(d.updateQueue=null,d.memoizedState=null)}var h=vi(i);if(null!==h){h.flags&=-257,gi(h,i,s,0,t),1&h.mode&&mi(l,c,t),u=c;var m=(t=h).updateQueue;if(null===m){var v=new Set;v.add(u),t.updateQueue=v}else m.add(u);break e}if(0===(1&t)){mi(l,c,t),mu();break e}u=Error(o(426))}else if(ao&&1&s.mode){var g=vi(i);if(null!==g){0===(65536&g.flags)&&(g.flags|=256),gi(g,i,s,0,t),mo(ui(u,s));break e}}l=u=ui(u,s),4!==As&&(As=2),null===Is?Is=[l]:Is.push(l),l=i;do{switch(l.tag){case 3:l.flags|=65536,t&=-t,l.lanes|=t,$o(l,pi(0,u,t));break e;case 1:s=u;var y=l.type,b=l.stateNode;if(0===(128&l.flags)&&("function"===typeof y.getDerivedStateFromError||null!==b&&"function"===typeof b.componentDidCatch&&(null===qs||!qs.has(b)))){l.flags|=65536,t&=-t,l.lanes|=t,$o(l,hi(l,s,t));break e}}l=l.return}while(null!==l)}wu(n)}catch(w){t=w,_s===n&&null!==n&&(_s=n=n.return);continue}break}}function hu(){var e=Cs.current;return Cs.current=Yl,null===e?Yl:e}function mu(){0!==As&&3!==As&&2!==As||(As=4),null===js||0===(268435455&Fs)&&0===(268435455&Us)||iu(js,Ts)}function vu(e,t){var n=Rs;Rs|=2;var r=hu();for(js===e&&Ts===t||(Ws=null,fu(e,t));;)try{gu();break}catch(a){pu(e,a)}if(Oo(),Rs=n,Cs.current=r,null!==_s)throw Error(o(261));return js=null,Ts=0,As}function gu(){for(;null!==_s;)bu(_s)}function yu(){for(;null!==_s&&!Xe();)bu(_s)}function bu(e){var t=Ss(e.alternate,e,Ps);e.memoizedProps=e.pendingProps,null===t?wu(e):_s=t,Ns.current=null}function wu(e){var t=e;do{var n=t.alternate;if(e=t.return,0===(32768&t.flags)){if(null!==(n=Ki(n,t,Ps)))return void(_s=n)}else{if(null!==(n=Qi(n,t)))return n.flags&=32767,void(_s=n);if(null===e)return As=6,void(_s=null);e.flags|=32768,e.subtreeFlags=0,e.deletions=null}if(null!==(t=t.sibling))return void(_s=t);_s=t=e}while(null!==t);0===As&&(As=5)}function xu(e,t,n){var r=bt,a=Os.transition;try{Os.transition=null,bt=1,function(e,t,n,r){do{Eu()}while(null!==Qs);if(0!==(6&Rs))throw Error(o(327));n=e.finishedWork;var a=e.finishedLanes;if(null===n)return null;if(e.finishedWork=null,e.finishedLanes=0,n===e.current)throw Error(o(177));e.callbackNode=null,e.callbackPriority=0;var l=n.lanes|n.childLanes;if(function(e,t){var n=e.pendingLanes&~t;e.pendingLanes=t,e.suspendedLanes=0,e.pingedLanes=0,e.expiredLanes&=t,e.mutableReadLanes&=t,e.entangledLanes&=t,t=e.entanglements;var r=e.eventTimes;for(e=e.expirationTimes;0<n;){var a=31-lt(n),o=1<<a;t[a]=0,r[a]=-1,e[a]=-1,n&=~o}}(e,l),e===js&&(_s=js=null,Ts=0),0===(2064&n.subtreeFlags)&&0===(2064&n.flags)||Ks||(Ks=!0,ju(tt,(function(){return Eu(),null}))),l=0!==(15990&n.flags),0!==(15990&n.subtreeFlags)||l){l=Os.transition,Os.transition=null;var i=bt;bt=1;var s=Rs;Rs|=4,Ns.current=null,function(e,t){if(ea=Ht,pr(e=fr())){if("selectionStart"in e)var n={start:e.selectionStart,end:e.selectionEnd};else e:{var r=(n=(n=e.ownerDocument)&&n.defaultView||window).getSelection&&n.getSelection();if(r&&0!==r.rangeCount){n=r.anchorNode;var a=r.anchorOffset,l=r.focusNode;r=r.focusOffset;try{n.nodeType,l.nodeType}catch(x){n=null;break e}var i=0,s=-1,u=-1,c=0,d=0,f=e,p=null;t:for(;;){for(var h;f!==n||0!==a&&3!==f.nodeType||(s=i+a),f!==l||0!==r&&3!==f.nodeType||(u=i+r),3===f.nodeType&&(i+=f.nodeValue.length),null!==(h=f.firstChild);)p=f,f=h;for(;;){if(f===e)break t;if(p===n&&++c===a&&(s=i),p===l&&++d===r&&(u=i),null!==(h=f.nextSibling))break;p=(f=p).parentNode}f=h}n=-1===s||-1===u?null:{start:s,end:u}}else n=null}n=n||{start:0,end:0}}else n=null;for(ta={focusedElem:e,selectionRange:n},Ht=!1,Yi=t;null!==Yi;)if(e=(t=Yi).child,0!==(1028&t.subtreeFlags)&&null!==e)e.return=t,Yi=e;else for(;null!==Yi;){t=Yi;try{var m=t.alternate;if(0!==(1024&t.flags))switch(t.tag){case 0:case 11:case 15:case 5:case 6:case 4:case 17:break;case 1:if(null!==m){var v=m.memoizedProps,g=m.memoizedState,y=t.stateNode,b=y.getSnapshotBeforeUpdate(t.elementType===t.type?v:ni(t.type,v),g);y.__reactInternalSnapshotBeforeUpdate=b}break;case 3:var w=t.stateNode.containerInfo;1===w.nodeType?w.textContent="":9===w.nodeType&&w.documentElement&&w.removeChild(w.documentElement);break;default:throw Error(o(163))}}catch(x){ku(t,t.return,x)}if(null!==(e=t.sibling)){e.return=t.return,Yi=e;break}Yi=t.return}m=ts,ts=!1}(e,n),vs(n,e),hr(ta),Ht=!!ea,ta=ea=null,e.current=n,ys(n,e,a),Ge(),Rs=s,bt=i,Os.transition=l}else e.current=n;if(Ks&&(Ks=!1,Qs=e,Xs=a),l=e.pendingLanes,0===l&&(qs=null),function(e){if(ot&&"function"===typeof ot.onCommitFiberRoot)try{ot.onCommitFiberRoot(at,e,void 0,128===(128&e.current.flags))}catch(t){}}(n.stateNode),ru(e,Je()),null!==t)for(r=e.onRecoverableError,n=0;n<t.length;n++)a=t[n],r(a.value,{componentStack:a.stack,digest:a.digest});if(Hs)throw Hs=!1,e=Vs,Vs=null,e;0!==(1&Xs)&&0!==e.tag&&Eu(),l=e.pendingLanes,0!==(1&l)?e===Js?Gs++:(Gs=0,Js=e):Gs=0,$a()}(e,t,n,r)}finally{Os.transition=a,bt=r}return null}function Eu(){if(null!==Qs){var e=wt(Xs),t=Os.transition,n=bt;try{if(Os.transition=null,bt=16>e?16:e,null===Qs)var r=!1;else{if(e=Qs,Qs=null,Xs=0,0!==(6&Rs))throw Error(o(331));var a=Rs;for(Rs|=4,Yi=e.current;null!==Yi;){var l=Yi,i=l.child;if(0!==(16&Yi.flags)){var s=l.deletions;if(null!==s){for(var u=0;u<s.length;u++){var c=s[u];for(Yi=c;null!==Yi;){var d=Yi;switch(d.tag){case 0:case 11:case 15:ns(8,d,l)}var f=d.child;if(null!==f)f.return=d,Yi=f;else for(;null!==Yi;){var p=(d=Yi).sibling,h=d.return;if(os(d),d===c){Yi=null;break}if(null!==p){p.return=h,Yi=p;break}Yi=h}}}var m=l.alternate;if(null!==m){var v=m.child;if(null!==v){m.child=null;do{var g=v.sibling;v.sibling=null,v=g}while(null!==v)}}Yi=l}}if(0!==(2064&l.subtreeFlags)&&null!==i)i.return=l,Yi=i;else e:for(;null!==Yi;){if(0!==(2048&(l=Yi).flags))switch(l.tag){case 0:case 11:case 15:ns(9,l,l.return)}var y=l.sibling;if(null!==y){y.return=l.return,Yi=y;break e}Yi=l.return}}var b=e.current;for(Yi=b;null!==Yi;){var w=(i=Yi).child;if(0!==(2064&i.subtreeFlags)&&null!==w)w.return=i,Yi=w;else e:for(i=b;null!==Yi;){if(0!==(2048&(s=Yi).flags))try{switch(s.tag){case 0:case 11:case 15:rs(9,s)}}catch(E){ku(s,s.return,E)}if(s===i){Yi=null;break e}var x=s.sibling;if(null!==x){x.return=s.return,Yi=x;break e}Yi=s.return}}if(Rs=a,$a(),ot&&"function"===typeof ot.onPostCommitFiberRoot)try{ot.onPostCommitFiberRoot(at,e)}catch(E){}r=!0}return r}finally{bt=n,Os.transition=t}}return!1}function Su(e,t,n){e=Mo(e,t=pi(0,t=ui(n,t),1),1),t=eu(),null!==e&&(gt(e,1,t),ru(e,t))}function ku(e,t,n){if(3===e.tag)Su(e,e,n);else for(;null!==t;){if(3===t.tag){Su(t,e,n);break}if(1===t.tag){var r=t.stateNode;if("function"===typeof t.type.getDerivedStateFromError||"function"===typeof r.componentDidCatch&&(null===qs||!qs.has(r))){t=Mo(t,e=hi(t,e=ui(n,e),1),1),e=eu(),null!==t&&(gt(t,1,e),ru(t,e));break}}t=t.return}}function Cu(e,t,n){var r=e.pingCache;null!==r&&r.delete(t),t=eu(),e.pingedLanes|=e.suspendedLanes&n,js===e&&(Ts&n)===n&&(4===As||3===As&&(130023424&Ts)===Ts&&500>Je()-Bs?fu(e,0):zs|=n),ru(e,t)}function Nu(e,t){0===t&&(0===(1&e.mode)?t=1:(t=ct,0===(130023424&(ct<<=1))&&(ct=4194304)));var n=eu();null!==(e=Do(e,t))&&(gt(e,t,n),ru(e,n))}function Ou(e){var t=e.memoizedState,n=0;null!==t&&(n=t.retryLane),Nu(e,n)}function Ru(e,t){var n=0;switch(e.tag){case 13:var r=e.stateNode,a=e.memoizedState;null!==a&&(n=a.retryLane);break;case 19:r=e.stateNode;break;default:throw Error(o(314))}null!==r&&r.delete(t),Nu(e,n)}function ju(e,t){return Ke(e,t)}function _u(e,t,n,r){this.tag=e,this.key=n,this.sibling=this.child=this.return=this.stateNode=this.type=this.elementType=null,this.index=0,this.ref=null,this.pendingProps=t,this.dependencies=this.memoizedState=this.updateQueue=this.memoizedProps=null,this.mode=r,this.subtreeFlags=this.flags=0,this.deletions=null,this.childLanes=this.lanes=0,this.alternate=null}function Tu(e,t,n,r){return new _u(e,t,n,r)}function Pu(e){return!(!(e=e.prototype)||!e.isReactComponent)}function Lu(e,t){var n=e.alternate;return null===n?((n=Tu(e.tag,t,e.key,e.mode)).elementType=e.elementType,n.type=e.type,n.stateNode=e.stateNode,n.alternate=e,e.alternate=n):(n.pendingProps=t,n.type=e.type,n.flags=0,n.subtreeFlags=0,n.deletions=null),n.flags=14680064&e.flags,n.childLanes=e.childLanes,n.lanes=e.lanes,n.child=e.child,n.memoizedProps=e.memoizedProps,n.memoizedState=e.memoizedState,n.updateQueue=e.updateQueue,t=e.dependencies,n.dependencies=null===t?null:{lanes:t.lanes,firstContext:t.firstContext},n.sibling=e.sibling,n.index=e.index,n.ref=e.ref,n}function Au(e,t,n,r,a,l){var i=2;if(r=e,"function"===typeof e)Pu(e)&&(i=1);else if("string"===typeof e)i=5;else e:switch(e){case S:return Du(n.children,a,l,t);case k:i=8,a|=8;break;case C:return(e=Tu(12,n,t,2|a)).elementType=C,e.lanes=l,e;case j:return(e=Tu(13,n,t,a)).elementType=j,e.lanes=l,e;case _:return(e=Tu(19,n,t,a)).elementType=_,e.lanes=l,e;case L:return Fu(n,a,l,t);default:if("object"===typeof e&&null!==e)switch(e.$$typeof){case N:i=10;break e;case O:i=9;break e;case R:i=11;break e;case T:i=14;break e;case P:i=16,r=null;break e}throw Error(o(130,null==e?e:typeof e,""))}return(t=Tu(i,n,t,a)).elementType=e,t.type=r,t.lanes=l,t}function Du(e,t,n,r){return(e=Tu(7,e,r,t)).lanes=n,e}function Fu(e,t,n,r){return(e=Tu(22,e,r,t)).elementType=L,e.lanes=n,e.stateNode={isHidden:!1},e}function Uu(e,t,n){return(e=Tu(6,e,null,t)).lanes=n,e}function zu(e,t,n){return(t=Tu(4,null!==e.children?e.children:[],e.key,t)).lanes=n,t.stateNode={containerInfo:e.containerInfo,pendingChildren:null,implementation:e.implementation},t}function Iu(e,t,n,r,a){this.tag=t,this.containerInfo=e,this.finishedWork=this.pingCache=this.current=this.pendingChildren=null,this.timeoutHandle=-1,this.callbackNode=this.pendingContext=this.context=null,this.callbackPriority=0,this.eventTimes=vt(0),this.expirationTimes=vt(-1),this.entangledLanes=this.finishedLanes=this.mutableReadLanes=this.expiredLanes=this.pingedLanes=this.suspendedLanes=this.pendingLanes=0,this.entanglements=vt(0),this.identifierPrefix=r,this.onRecoverableError=a,this.mutableSourceEagerHydrationData=null}function Mu(e,t,n,r,a,o,l,i,s){return e=new Iu(e,t,n,i,s),1===t?(t=1,!0===o&&(t|=8)):t=0,o=Tu(3,null,null,t),e.current=o,o.stateNode=e,o.memoizedState={element:r,isDehydrated:n,cache:null,transitions:null,pendingSuspenseBoundaries:null},Uo(o),e}function Bu(e){if(!e)return Oa;e:{if($e(e=e._reactInternals)!==e||1!==e.tag)throw Error(o(170));var t=e;do{switch(t.tag){case 3:t=t.stateNode.context;break e;case 1:if(Pa(t.type)){t=t.stateNode.__reactInternalMemoizedMergedChildContext;break e}}t=t.return}while(null!==t);throw Error(o(171))}if(1===e.tag){var n=e.type;if(Pa(n))return Da(e,n,t)}return t}function $u(e,t,n,r,a,o,l,i,s){return(e=Mu(n,r,!0,e,0,o,0,i,s)).context=Bu(null),n=e.current,(o=Io(r=eu(),a=tu(n))).callback=void 0!==t&&null!==t?t:null,Mo(n,o,a),e.current.lanes=a,gt(e,a,r),ru(e,r),e}function Wu(e,t,n,r){var a=t.current,o=eu(),l=tu(a);return n=Bu(n),null===t.context?t.context=n:t.pendingContext=n,(t=Io(o,l)).payload={element:e},null!==(r=void 0===r?null:r)&&(t.callback=r),null!==(e=Mo(a,t,l))&&(nu(e,a,l,o),Bo(e,a,l)),l}function Hu(e){return(e=e.current).child?(e.child.tag,e.child.stateNode):null}function Vu(e,t){if(null!==(e=e.memoizedState)&&null!==e.dehydrated){var n=e.retryLane;e.retryLane=0!==n&&n<t?n:t}}function qu(e,t){Vu(e,t),(e=e.alternate)&&Vu(e,t)}Ss=function(e,t,n){if(null!==e)if(e.memoizedProps!==t.pendingProps||ja.current)bi=!0;else{if(0===(e.lanes&n)&&0===(128&t.flags))return bi=!1,function(e,t,n){switch(t.tag){case 3:ji(t),ho();break;case 5:Yo(t);break;case 1:Pa(t.type)&&Fa(t);break;case 4:Go(t,t.stateNode.containerInfo);break;case 10:var r=t.type._context,a=t.memoizedProps.value;Na(So,r._currentValue),r._currentValue=a;break;case 13:if(null!==(r=t.memoizedState))return null!==r.dehydrated?(Na(el,1&el.current),t.flags|=128,null):0!==(n&t.child.childLanes)?Ui(e,t,n):(Na(el,1&el.current),null!==(e=Hi(e,t,n))?e.sibling:null);Na(el,1&el.current);break;case 19:if(r=0!==(n&t.childLanes),0!==(128&e.flags)){if(r)return $i(e,t,n);t.flags|=128}if(null!==(a=t.memoizedState)&&(a.rendering=null,a.tail=null,a.lastEffect=null),Na(el,el.current),r)break;return null;case 22:case 23:return t.lanes=0,ki(e,t,n)}return Hi(e,t,n)}(e,t,n);bi=0!==(131072&e.flags)}else bi=!1,ao&&0!==(1048576&t.flags)&&Za(t,qa,t.index);switch(t.lanes=0,t.tag){case 2:var r=t.type;Wi(e,t),e=t.pendingProps;var a=Ta(t,Ra.current);_o(t,n),a=vl(null,t,r,e,a,n);var l=gl();return t.flags|=1,"object"===typeof a&&null!==a&&"function"===typeof a.render&&void 0===a.$$typeof?(t.tag=1,t.memoizedState=null,t.updateQueue=null,Pa(r)?(l=!0,Fa(t)):l=!1,t.memoizedState=null!==a.state&&void 0!==a.state?a.state:null,Uo(t),a.updater=ai,t.stateNode=a,a._reactInternals=t,si(t,r,e,n),t=Ri(null,t,r,!0,l,n)):(t.tag=0,ao&&l&&eo(t),wi(null,t,a,n),t=t.child),t;case 16:r=t.elementType;e:{switch(Wi(e,t),e=t.pendingProps,r=(a=r._init)(r._payload),t.type=r,a=t.tag=function(e){if("function"===typeof e)return Pu(e)?1:0;if(void 0!==e&&null!==e){if((e=e.$$typeof)===R)return 11;if(e===T)return 14}return 2}(r),e=ni(r,e),a){case 0:t=Ni(null,t,r,e,n);break e;case 1:t=Oi(null,t,r,e,n);break e;case 11:t=xi(null,t,r,e,n);break e;case 14:t=Ei(null,t,r,ni(r.type,e),n);break e}throw Error(o(306,r,""))}return t;case 0:return r=t.type,a=t.pendingProps,Ni(e,t,r,a=t.elementType===r?a:ni(r,a),n);case 1:return r=t.type,a=t.pendingProps,Oi(e,t,r,a=t.elementType===r?a:ni(r,a),n);case 3:e:{if(ji(t),null===e)throw Error(o(387));r=t.pendingProps,a=(l=t.memoizedState).element,zo(e,t),Wo(t,r,null,n);var i=t.memoizedState;if(r=i.element,l.isDehydrated){if(l={element:r,isDehydrated:!1,cache:i.cache,pendingSuspenseBoundaries:i.pendingSuspenseBoundaries,transitions:i.transitions},t.updateQueue.baseState=l,t.memoizedState=l,256&t.flags){t=_i(e,t,r,n,a=ui(Error(o(423)),t));break e}if(r!==a){t=_i(e,t,r,n,a=ui(Error(o(424)),t));break e}for(ro=ua(t.stateNode.containerInfo.firstChild),no=t,ao=!0,oo=null,n=Eo(t,null,r,n),t.child=n;n;)n.flags=-3&n.flags|4096,n=n.sibling}else{if(ho(),r===a){t=Hi(e,t,n);break e}wi(e,t,r,n)}t=t.child}return t;case 5:return Yo(t),null===e&&uo(t),r=t.type,a=t.pendingProps,l=null!==e?e.memoizedProps:null,i=a.children,na(r,a)?i=null:null!==l&&na(r,l)&&(t.flags|=32),Ci(e,t),wi(e,t,i,n),t.child;case 6:return null===e&&uo(t),null;case 13:return Ui(e,t,n);case 4:return Go(t,t.stateNode.containerInfo),r=t.pendingProps,null===e?t.child=xo(t,null,r,n):wi(e,t,r,n),t.child;case 11:return r=t.type,a=t.pendingProps,xi(e,t,r,a=t.elementType===r?a:ni(r,a),n);case 7:return wi(e,t,t.pendingProps,n),t.child;case 8:case 12:return wi(e,t,t.pendingProps.children,n),t.child;case 10:e:{if(r=t.type._context,a=t.pendingProps,l=t.memoizedProps,i=a.value,Na(So,r._currentValue),r._currentValue=i,null!==l)if(ir(l.value,i)){if(l.children===a.children&&!ja.current){t=Hi(e,t,n);break e}}else for(null!==(l=t.child)&&(l.return=t);null!==l;){var s=l.dependencies;if(null!==s){i=l.child;for(var u=s.firstContext;null!==u;){if(u.context===r){if(1===l.tag){(u=Io(-1,n&-n)).tag=2;var c=l.updateQueue;if(null!==c){var d=(c=c.shared).pending;null===d?u.next=u:(u.next=d.next,d.next=u),c.pending=u}}l.lanes|=n,null!==(u=l.alternate)&&(u.lanes|=n),jo(l.return,n,t),s.lanes|=n;break}u=u.next}}else if(10===l.tag)i=l.type===t.type?null:l.child;else if(18===l.tag){if(null===(i=l.return))throw Error(o(341));i.lanes|=n,null!==(s=i.alternate)&&(s.lanes|=n),jo(i,n,t),i=l.sibling}else i=l.child;if(null!==i)i.return=l;else for(i=l;null!==i;){if(i===t){i=null;break}if(null!==(l=i.sibling)){l.return=i.return,i=l;break}i=i.return}l=i}wi(e,t,a.children,n),t=t.child}return t;case 9:return a=t.type,r=t.pendingProps.children,_o(t,n),r=r(a=To(a)),t.flags|=1,wi(e,t,r,n),t.child;case 14:return a=ni(r=t.type,t.pendingProps),Ei(e,t,r,a=ni(r.type,a),n);case 15:return Si(e,t,t.type,t.pendingProps,n);case 17:return r=t.type,a=t.pendingProps,a=t.elementType===r?a:ni(r,a),Wi(e,t),t.tag=1,Pa(r)?(e=!0,Fa(t)):e=!1,_o(t,n),li(t,r,a),si(t,r,a,n),Ri(null,t,r,!0,e,n);case 19:return $i(e,t,n);case 22:return ki(e,t,n)}throw Error(o(156,t.tag))};var Ku="function"===typeof reportError?reportError:function(e){console.error(e)};function Qu(e){this._internalRoot=e}function Xu(e){this._internalRoot=e}function Gu(e){return!(!e||1!==e.nodeType&&9!==e.nodeType&&11!==e.nodeType)}function Ju(e){return!(!e||1!==e.nodeType&&9!==e.nodeType&&11!==e.nodeType&&(8!==e.nodeType||" react-mount-point-unstable "!==e.nodeValue))}function Yu(){}function Zu(e,t,n,r,a){var o=n._reactRootContainer;if(o){var l=o;if("function"===typeof a){var i=a;a=function(){var e=Hu(l);i.call(e)}}Wu(t,l,e,a)}else l=function(e,t,n,r,a){if(a){if("function"===typeof r){var o=r;r=function(){var e=Hu(l);o.call(e)}}var l=$u(t,r,e,0,null,!1,0,"",Yu);return e._reactRootContainer=l,e[ha]=l.current,$r(8===e.nodeType?e.parentNode:e),cu(),l}for(;a=e.lastChild;)e.removeChild(a);if("function"===typeof r){var i=r;r=function(){var e=Hu(s);i.call(e)}}var s=Mu(e,0,!1,null,0,!1,0,"",Yu);return e._reactRootContainer=s,e[ha]=s.current,$r(8===e.nodeType?e.parentNode:e),cu((function(){Wu(t,s,n,r)})),s}(n,t,e,a,r);return Hu(l)}Xu.prototype.render=Qu.prototype.render=function(e){var t=this._internalRoot;if(null===t)throw Error(o(409));Wu(e,t,null,null)},Xu.prototype.unmount=Qu.prototype.unmount=function(){var e=this._internalRoot;if(null!==e){this._internalRoot=null;var t=e.containerInfo;cu((function(){Wu(null,e,null,null)})),t[ha]=null}},Xu.prototype.unstable_scheduleHydration=function(e){if(e){var t=kt();e={blockedOn:null,target:e,priority:t};for(var n=0;n<Lt.length&&0!==t&&t<Lt[n].priority;n++);Lt.splice(n,0,e),0===n&&Ut(e)}},xt=function(e){switch(e.tag){case 3:var t=e.stateNode;if(t.current.memoizedState.isDehydrated){var n=dt(t.pendingLanes);0!==n&&(yt(t,1|n),ru(t,Je()),0===(6&Rs)&&($s=Je()+500,$a()))}break;case 13:cu((function(){var t=Do(e,1);if(null!==t){var n=eu();nu(t,e,1,n)}})),qu(e,1)}},Et=function(e){if(13===e.tag){var t=Do(e,134217728);if(null!==t)nu(t,e,134217728,eu());qu(e,134217728)}},St=function(e){if(13===e.tag){var t=tu(e),n=Do(e,t);if(null!==n)nu(n,e,t,eu());qu(e,t)}},kt=function(){return bt},Ct=function(e,t){var n=bt;try{return bt=e,t()}finally{bt=n}},Ee=function(e,t,n){switch(t){case"input":if(Y(e,n),t=n.name,"radio"===n.type&&null!=t){for(n=e;n.parentNode;)n=n.parentNode;for(n=n.querySelectorAll("input[name="+JSON.stringify(""+t)+'][type="radio"]'),t=0;t<n.length;t++){var r=n[t];if(r!==e&&r.form===e.form){var a=xa(r);if(!a)throw Error(o(90));K(r),Y(r,a)}}}break;case"textarea":oe(e,n);break;case"select":null!=(t=n.value)&&ne(e,!!n.multiple,t,!1)}},Re=uu,je=cu;var ec={usingClientEntryPoint:!1,Events:[ba,wa,xa,Ne,Oe,uu]},tc={findFiberByHostInstance:ya,bundleType:0,version:"18.3.1",rendererPackageName:"react-dom"},nc={bundleType:tc.bundleType,version:tc.version,rendererPackageName:tc.rendererPackageName,rendererConfig:tc.rendererConfig,overrideHookState:null,overrideHookStateDeletePath:null,overrideHookStateRenamePath:null,overrideProps:null,overridePropsDeletePath:null,overridePropsRenamePath:null,setErrorHandler:null,setSuspenseHandler:null,scheduleUpdate:null,currentDispatcherRef:w.ReactCurrentDispatcher,findHostInstanceByFiber:function(e){return null===(e=Ve(e))?null:e.stateNode},findFiberByHostInstance:tc.findFiberByHostInstance||function(){return null},findHostInstancesForRefresh:null,scheduleRefresh:null,scheduleRoot:null,setRefreshHandler:null,getCurrentFiber:null,reconcilerVersion:"18.3.1-next-f1338f8080-20240426"};if("undefined"!==typeof __REACT_DEVTOOLS_GLOBAL_HOOK__){var rc=__REACT_DEVTOOLS_GLOBAL_HOOK__;if(!rc.isDisabled&&rc.supportsFiber)try{at=rc.inject(nc),ot=rc}catch(ce){}}t.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=ec,t.createPortal=function(e,t){var n=2<arguments.length&&void 0!==arguments[2]?arguments[2]:null;if(!Gu(t))throw Error(o(200));return function(e,t,n){var r=3<arguments.length&&void 0!==arguments[3]?arguments[3]:null;return{$$typeof:E,key:null==r?null:""+r,children:e,containerInfo:t,implementation:n}}(e,t,null,n)},t.createRoot=function(e,t){if(!Gu(e))throw Error(o(299));var n=!1,r="",a=Ku;return null!==t&&void 0!==t&&(!0===t.unstable_strictMode&&(n=!0),void 0!==t.identifierPrefix&&(r=t.identifierPrefix),void 0!==t.onRecoverableError&&(a=t.onRecoverableError)),t=Mu(e,1,!1,null,0,n,0,r,a),e[ha]=t.current,$r(8===e.nodeType?e.parentNode:e),new Qu(t)},t.findDOMNode=function(e){if(null==e)return null;if(1===e.nodeType)return e;var t=e._reactInternals;if(void 0===t){if("function"===typeof e.render)throw Error(o(188));throw e=Object.keys(e).join(","),Error(o(268,e))}return e=null===(e=Ve(t))?null:e.stateNode},t.flushSync=function(e){return cu(e)},t.hydrate=function(e,t,n){if(!Ju(t))throw Error(o(200));return Zu(null,e,t,!0,n)},t.hydrateRoot=function(e,t,n){if(!Gu(e))throw Error(o(405));var r=null!=n&&n.hydratedSources||null,a=!1,l="",i=Ku;if(null!==n&&void 0!==n&&(!0===n.unstable_strictMode&&(a=!0),void 0!==n.identifierPrefix&&(l=n.identifierPrefix),void 0!==n.onRecoverableError&&(i=n.onRecoverableError)),t=$u(t,null,e,1,null!=n?n:null,a,0,l,i),e[ha]=t.current,$r(e),r)for(e=0;e<r.length;e++)a=(a=(n=r[e])._getVersion)(n._source),null==t.mutableSourceEagerHydrationData?t.mutableSourceEagerHydrationData=[n,a]:t.mutableSourceEagerHydrationData.push(n,a);return new Xu(t)},t.render=function(e,t,n){if(!Ju(t))throw Error(o(200));return Zu(null,e,t,!1,n)},t.unmountComponentAtNode=function(e){if(!Ju(e))throw Error(o(40));return!!e._reactRootContainer&&(cu((function(){Zu(null,null,e,!1,(function(){e._reactRootContainer=null,e[ha]=null}))})),!0)},t.unstable_batchedUpdates=uu,t.unstable_renderSubtreeIntoContainer=function(e,t,n,r){if(!Ju(n))throw Error(o(200));if(null==e||void 0===e._reactInternals)throw Error(o(38));return Zu(e,t,n,!1,r)},t.version="18.3.1-next-f1338f8080-20240426"},391:(e,t,n)=>{"use strict";var r=n(950);t.createRoot=r.createRoot,t.hydrateRoot=r.hydrateRoot},950:(e,t,n)=>{"use strict";!function e(){if("undefined"!==typeof __REACT_DEVTOOLS_GLOBAL_HOOK__&&"function"===typeof __REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE)try{__REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE(e)}catch(t){console.error(t)}}(),e.exports=n(730)},153:(e,t,n)=>{"use strict";var r=n(43),a=Symbol.for("react.element"),o=Symbol.for("react.fragment"),l=Object.prototype.hasOwnProperty,i=r.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.ReactCurrentOwner,s={key:!0,ref:!0,__self:!0,__source:!0};function u(e,t,n){var r,o={},u=null,c=null;for(r in void 0!==n&&(u=""+n),void 0!==t.key&&(u=""+t.key),void 0!==t.ref&&(c=t.ref),t)l.call(t,r)&&!s.hasOwnProperty(r)&&(o[r]=t[r]);if(e&&e.defaultProps)for(r in t=e.defaultProps)void 0===o[r]&&(o[r]=t[r]);return{$$typeof:a,type:e,key:u,ref:c,props:o,_owner:i.current}}t.Fragment=o,t.jsx=u,t.jsxs=u},202:(e,t)=>{"use strict";var n=Symbol.for("react.element"),r=Symbol.for("react.portal"),a=Symbol.for("react.fragment"),o=Symbol.for("react.strict_mode"),l=Symbol.for("react.profiler"),i=Symbol.for("react.provider"),s=Symbol.for("react.context"),u=Symbol.for("react.forward_ref"),c=Symbol.for("react.suspense"),d=Symbol.for("react.memo"),f=Symbol.for("react.lazy"),p=Symbol.iterator;var h={isMounted:function(){return!1},enqueueForceUpdate:function(){},enqueueReplaceState:function(){},enqueueSetState:function(){}},m=Object.assign,v={};function g(e,t,n){this.props=e,this.context=t,this.refs=v,this.updater=n||h}function y(){}function b(e,t,n){this.props=e,this.context=t,this.refs=v,this.updater=n||h}g.prototype.isReactComponent={},g.prototype.setState=function(e,t){if("object"!==typeof e&&"function"!==typeof e&&null!=e)throw Error("setState(...): takes an object of state variables to update or a function which returns an object of state variables.");this.updater.enqueueSetState(this,e,t,"setState")},g.prototype.forceUpdate=function(e){this.updater.enqueueForceUpdate(this,e,"forceUpdate")},y.prototype=g.prototype;var w=b.prototype=new y;w.constructor=b,m(w,g.prototype),w.isPureReactComponent=!0;var x=Array.isArray,E=Object.prototype.hasOwnProperty,S={current:null},k={key:!0,ref:!0,__self:!0,__source:!0};function C(e,t,r){var a,o={},l=null,i=null;if(null!=t)for(a in void 0!==t.ref&&(i=t.ref),void 0!==t.key&&(l=""+t.key),t)E.call(t,a)&&!k.hasOwnProperty(a)&&(o[a]=t[a]);var s=arguments.length-2;if(1===s)o.children=r;else if(1<s){for(var u=Array(s),c=0;c<s;c++)u[c]=arguments[c+2];o.children=u}if(e&&e.defaultProps)for(a in s=e.defaultProps)void 0===o[a]&&(o[a]=s[a]);return{$$typeof:n,type:e,key:l,ref:i,props:o,_owner:S.current}}function N(e){return"object"===typeof e&&null!==e&&e.$$typeof===n}var O=/\/+/g;function R(e,t){return"object"===typeof e&&null!==e&&null!=e.key?function(e){var t={"=":"=0",":":"=2"};return"$"+e.replace(/[=:]/g,(function(e){return t[e]}))}(""+e.key):t.toString(36)}function j(e,t,a,o,l){var i=typeof e;"undefined"!==i&&"boolean"!==i||(e=null);var s=!1;if(null===e)s=!0;else switch(i){case"string":case"number":s=!0;break;case"object":switch(e.$$typeof){case n:case r:s=!0}}if(s)return l=l(s=e),e=""===o?"."+R(s,0):o,x(l)?(a="",null!=e&&(a=e.replace(O,"$&/")+"/"),j(l,t,a,"",(function(e){return e}))):null!=l&&(N(l)&&(l=function(e,t){return{$$typeof:n,type:e.type,key:t,ref:e.ref,props:e.props,_owner:e._owner}}(l,a+(!l.key||s&&s.key===l.key?"":(""+l.key).replace(O,"$&/")+"/")+e)),t.push(l)),1;if(s=0,o=""===o?".":o+":",x(e))for(var u=0;u<e.length;u++){var c=o+R(i=e[u],u);s+=j(i,t,a,c,l)}else if(c=function(e){return null===e||"object"!==typeof e?null:"function"===typeof(e=p&&e[p]||e["@@iterator"])?e:null}(e),"function"===typeof c)for(e=c.call(e),u=0;!(i=e.next()).done;)s+=j(i=i.value,t,a,c=o+R(i,u++),l);else if("object"===i)throw t=String(e),Error("Objects are not valid as a React child (found: "+("[object Object]"===t?"object with keys {"+Object.keys(e).join(", ")+"}":t)+"). If you meant to render a collection of children, use an array instead.");return s}function _(e,t,n){if(null==e)return e;var r=[],a=0;return j(e,r,"","",(function(e){return t.call(n,e,a++)})),r}function T(e){if(-1===e._status){var t=e._result;(t=t()).then((function(t){0!==e._status&&-1!==e._status||(e._status=1,e._result=t)}),(function(t){0!==e._status&&-1!==e._status||(e._status=2,e._result=t)})),-1===e._status&&(e._status=0,e._result=t)}if(1===e._status)return e._result.default;throw e._result}var P={current:null},L={transition:null},A={ReactCurrentDispatcher:P,ReactCurrentBatchConfig:L,ReactCurrentOwner:S};function D(){throw Error("act(...) is not supported in production builds of React.")}t.Children={map:_,forEach:function(e,t,n){_(e,(function(){t.apply(this,arguments)}),n)},count:function(e){var t=0;return _(e,(function(){t++})),t},toArray:function(e){return _(e,(function(e){return e}))||[]},only:function(e){if(!N(e))throw Error("React.Children.only expected to receive a single React element child.");return e}},t.Component=g,t.Fragment=a,t.Profiler=l,t.PureComponent=b,t.StrictMode=o,t.Suspense=c,t.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=A,t.act=D,t.cloneElement=function(e,t,r){if(null===e||void 0===e)throw Error("React.cloneElement(...): The argument must be a React element, but you passed "+e+".");var a=m({},e.props),o=e.key,l=e.ref,i=e._owner;if(null!=t){if(void 0!==t.ref&&(l=t.ref,i=S.current),void 0!==t.key&&(o=""+t.key),e.type&&e.type.defaultProps)var s=e.type.defaultProps;for(u in t)E.call(t,u)&&!k.hasOwnProperty(u)&&(a[u]=void 0===t[u]&&void 0!==s?s[u]:t[u])}var u=arguments.length-2;if(1===u)a.children=r;else if(1<u){s=Array(u);for(var c=0;c<u;c++)s[c]=arguments[c+2];a.children=s}return{$$typeof:n,type:e.type,key:o,ref:l,props:a,_owner:i}},t.createContext=function(e){return(e={$$typeof:s,_currentValue:e,_currentValue2:e,_threadCount:0,Provider:null,Consumer:null,_defaultValue:null,_globalName:null}).Provider={$$typeof:i,_context:e},e.Consumer=e},t.createElement=C,t.createFactory=function(e){var t=C.bind(null,e);return t.type=e,t},t.createRef=function(){return{current:null}},t.forwardRef=function(e){return{$$typeof:u,render:e}},t.isValidElement=N,t.lazy=function(e){return{$$typeof:f,_payload:{_status:-1,_result:e},_init:T}},t.memo=function(e,t){return{$$typeof:d,type:e,compare:void 0===t?null:t}},t.startTransition=function(e){var t=L.transition;L.transition={};try{e()}finally{L.transition=t}},t.unstable_act=D,t.useCallback=function(e,t){return P.current.useCallback(e,t)},t.useContext=function(e){return P.current.useContext(e)},t.useDebugValue=function(){},t.useDeferredValue=function(e){return P.current.useDeferredValue(e)},t.useEffect=function(e,t){return P.current.useEffect(e,t)},t.useId=function(){return P.current.useId()},t.useImperativeHandle=function(e,t,n){return P.current.useImperativeHandle(e,t,n)},t.useInsertionEffect=function(e,t){return P.current.useInsertionEffect(e,t)},t.useLayoutEffect=function(e,t){return P.current.useLayoutEffect(e,t)},t.useMemo=function(e,t){return P.current.useMemo(e,t)},t.useReducer=function(e,t,n){return P.current.useReducer(e,t,n)},t.useRef=function(e){return P.current.useRef(e)},t.useState=function(e){return P.current.useState(e)},t.useSyncExternalStore=function(e,t,n){return P.current.useSyncExternalStore(e,t,n)},t.useTransition=function(){return P.current.useTransition()},t.version="18.3.1"},43:(e,t,n)=>{"use strict";e.exports=n(202)},579:(e,t,n)=>{"use strict";e.exports=n(153)},234:(e,t)=>{"use strict";function n(e,t){var n=e.length;e.push(t);e:for(;0<n;){var r=n-1>>>1,a=e[r];if(!(0<o(a,t)))break e;e[r]=t,e[n]=a,n=r}}function r(e){return 0===e.length?null:e[0]}function a(e){if(0===e.length)return null;var t=e[0],n=e.pop();if(n!==t){e[0]=n;e:for(var r=0,a=e.length,l=a>>>1;r<l;){var i=2*(r+1)-1,s=e[i],u=i+1,c=e[u];if(0>o(s,n))u<a&&0>o(c,s)?(e[r]=c,e[u]=n,r=u):(e[r]=s,e[i]=n,r=i);else{if(!(u<a&&0>o(c,n)))break e;e[r]=c,e[u]=n,r=u}}}return t}function o(e,t){var n=e.sortIndex-t.sortIndex;return 0!==n?n:e.id-t.id}if("object"===typeof performance&&"function"===typeof performance.now){var l=performance;t.unstable_now=function(){return l.now()}}else{var i=Date,s=i.now();t.unstable_now=function(){return i.now()-s}}var u=[],c=[],d=1,f=null,p=3,h=!1,m=!1,v=!1,g="function"===typeof setTimeout?setTimeout:null,y="function"===typeof clearTimeout?clearTimeout:null,b="undefined"!==typeof setImmediate?setImmediate:null;function w(e){for(var t=r(c);null!==t;){if(null===t.callback)a(c);else{if(!(t.startTime<=e))break;a(c),t.sortIndex=t.expirationTime,n(u,t)}t=r(c)}}function x(e){if(v=!1,w(e),!m)if(null!==r(u))m=!0,L(E);else{var t=r(c);null!==t&&A(x,t.startTime-e)}}function E(e,n){m=!1,v&&(v=!1,y(N),N=-1),h=!0;var o=p;try{for(w(n),f=r(u);null!==f&&(!(f.expirationTime>n)||e&&!j());){var l=f.callback;if("function"===typeof l){f.callback=null,p=f.priorityLevel;var i=l(f.expirationTime<=n);n=t.unstable_now(),"function"===typeof i?f.callback=i:f===r(u)&&a(u),w(n)}else a(u);f=r(u)}if(null!==f)var s=!0;else{var d=r(c);null!==d&&A(x,d.startTime-n),s=!1}return s}finally{f=null,p=o,h=!1}}"undefined"!==typeof navigator&&void 0!==navigator.scheduling&&void 0!==navigator.scheduling.isInputPending&&navigator.scheduling.isInputPending.bind(navigator.scheduling);var S,k=!1,C=null,N=-1,O=5,R=-1;function j(){return!(t.unstable_now()-R<O)}function _(){if(null!==C){var e=t.unstable_now();R=e;var n=!0;try{n=C(!0,e)}finally{n?S():(k=!1,C=null)}}else k=!1}if("function"===typeof b)S=function(){b(_)};else if("undefined"!==typeof MessageChannel){var T=new MessageChannel,P=T.port2;T.port1.onmessage=_,S=function(){P.postMessage(null)}}else S=function(){g(_,0)};function L(e){C=e,k||(k=!0,S())}function A(e,n){N=g((function(){e(t.unstable_now())}),n)}t.unstable_IdlePriority=5,t.unstable_ImmediatePriority=1,t.unstable_LowPriority=4,t.unstable_NormalPriority=3,t.unstable_Profiling=null,t.unstable_UserBlockingPriority=2,t.unstable_cancelCallback=function(e){e.callback=null},t.unstable_continueExecution=function(){m||h||(m=!0,L(E))},t.unstable_forceFrameRate=function(e){0>e||125<e?console.error("forceFrameRate takes a positive int between 0 and 125, forcing frame rates higher than 125 fps is not supported"):O=0<e?Math.floor(1e3/e):5},t.unstable_getCurrentPriorityLevel=function(){return p},t.unstable_getFirstCallbackNode=function(){return r(u)},t.unstable_next=function(e){switch(p){case 1:case 2:case 3:var t=3;break;default:t=p}var n=p;p=t;try{return e()}finally{p=n}},t.unstable_pauseExecution=function(){},t.unstable_requestPaint=function(){},t.unstable_runWithPriority=function(e,t){switch(e){case 1:case 2:case 3:case 4:case 5:break;default:e=3}var n=p;p=e;try{return t()}finally{p=n}},t.unstable_scheduleCallback=function(e,a,o){var l=t.unstable_now();switch("object"===typeof o&&null!==o?o="number"===typeof(o=o.delay)&&0<o?l+o:l:o=l,e){case 1:var i=-1;break;case 2:i=250;break;case 5:i=1073741823;break;case 4:i=1e4;break;default:i=5e3}return e={id:d++,callback:a,priorityLevel:e,startTime:o,expirationTime:i=o+i,sortIndex:-1},o>l?(e.sortIndex=o,n(c,e),null===r(u)&&e===r(c)&&(v?(y(N),N=-1):v=!0,A(x,o-l))):(e.sortIndex=i,n(u,e),m||h||(m=!0,L(E))),e},t.unstable_shouldYield=j,t.unstable_wrapCallback=function(e){var t=p;return function(){var n=p;p=t;try{return e.apply(this,arguments)}finally{p=n}}}},853:(e,t,n)=>{"use strict";e.exports=n(234)},139:(e,t)=>{var n;!function(){"use strict";var r={}.hasOwnProperty;function a(){for(var e="",t=0;t<arguments.length;t++){var n=arguments[t];n&&(e=l(e,o(n)))}return e}function o(e){if("string"===typeof e||"number"===typeof e)return e;if("object"!==typeof e)return"";if(Array.isArray(e))return a.apply(null,e);if(e.toString!==Object.prototype.toString&&!e.toString.toString().includes("[native code]"))return e.toString();var t="";for(var n in e)r.call(e,n)&&e[n]&&(t=l(t,n));return t}function l(e,t){return t?e?e+" "+t:e+t:e}e.exports?(a.default=a,e.exports=a):void 0===(n=function(){return a}.apply(t,[]))||(e.exports=n)}()}},t={};function n(r){var a=t[r];if(void 0!==a)return a.exports;var o=t[r]={exports:{}};return e[r].call(o.exports,o,o.exports,n),o.exports}n.m=e,n.n=e=>{var t=e&&e.__esModule?()=>e.default:()=>e;return n.d(t,{a:t}),t},(()=>{var e,t=Object.getPrototypeOf?e=>Object.getPrototypeOf(e):e=>e.__proto__;n.t=function(r,a){if(1&a&&(r=this(r)),8&a)return r;if("object"===typeof r&&r){if(4&a&&r.__esModule)return r;if(16&a&&"function"===typeof r.then)return r}var o=Object.create(null);n.r(o);var l={};e=e||[null,t({}),t([]),t(t)];for(var i=2&a&&r;"object"==typeof i&&!~e.indexOf(i);i=t(i))Object.getOwnPropertyNames(i).forEach((e=>l[e]=()=>r[e]));return l.default=()=>r,n.d(o,l),o}})(),n.d=(e,t)=>{for(var r in t)n.o(t,r)&&!n.o(e,r)&&Object.defineProperty(e,r,{enumerable:!0,get:t[r]})},n.f={},n.e=e=>Promise.all(Object.keys(n.f).reduce(((t,r)=>(n.f[r](e,t),t)),[])),n.u=e=>"static/js/"+e+".a5059bd1.chunk.js",n.miniCssF=e=>{},n.g=function(){if("object"===typeof globalThis)return globalThis;try{return this||new Function("return this")()}catch(e){if("object"===typeof window)return window}}(),n.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t),(()=>{var e={},t="client-app:";n.l=(r,a,o,l)=>{if(e[r])e[r].push(a);else{var i,s;if(void 0!==o)for(var u=document.getElementsByTagName("script"),c=0;c<u.length;c++){var d=u[c];if(d.getAttribute("src")==r||d.getAttribute("data-webpack")==t+o){i=d;break}}i||(s=!0,(i=document.createElement("script")).charset="utf-8",i.timeout=120,n.nc&&i.setAttribute("nonce",n.nc),i.setAttribute("data-webpack",t+o),i.src=r),e[r]=[a];var f=(t,n)=>{i.onerror=i.onload=null,clearTimeout(p);var a=e[r];if(delete e[r],i.parentNode&&i.parentNode.removeChild(i),a&&a.forEach((e=>e(n))),t)return t(n)},p=setTimeout(f.bind(null,void 0,{type:"timeout",target:i}),12e4);i.onerror=f.bind(null,i.onerror),i.onload=f.bind(null,i.onload),s&&document.head.appendChild(i)}}})(),n.r=e=>{"undefined"!==typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.p="/",(()=>{var e={792:0};n.f.j=(t,r)=>{var a=n.o(e,t)?e[t]:void 0;if(0!==a)if(a)r.push(a[2]);else{var o=new Promise(((n,r)=>a=e[t]=[n,r]));r.push(a[2]=o);var l=n.p+n.u(t),i=new Error;n.l(l,(r=>{if(n.o(e,t)&&(0!==(a=e[t])&&(e[t]=void 0),a)){var o=r&&("load"===r.type?"missing":r.type),l=r&&r.target&&r.target.src;i.message="Loading chunk "+t+" failed.\n("+o+": "+l+")",i.name="ChunkLoadError",i.type=o,i.request=l,a[1](i)}}),"chunk-"+t,t)}};var t=(t,r)=>{var a,o,l=r[0],i=r[1],s=r[2],u=0;if(l.some((t=>0!==e[t]))){for(a in i)n.o(i,a)&&(n.m[a]=i[a]);if(s)s(n)}for(t&&t(r);u<l.length;u++)o=l[u],n.o(e,o)&&e[o]&&e[o][0](),e[o]=0},r=self.webpackChunkclient_app=self.webpackChunkclient_app||[];r.forEach(t.bind(null,0)),r.push=t.bind(null,r.push.bind(r))})(),(()=>{"use strict";var e={};n.r(e),n.d(e,{hasBrowserEnv:()=>At,hasStandardBrowserEnv:()=>Ft,hasStandardBrowserWebWorkerEnv:()=>Ut,navigator:()=>Dt,origin:()=>zt});var t,r=n(43),a=n.t(r,2),o=n(391),l=n(950),i=n.t(l,2);function s(){return s=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},s.apply(this,arguments)}!function(e){e.Pop="POP",e.Push="PUSH",e.Replace="REPLACE"}(t||(t={}));const u="popstate";function c(e,t){if(!1===e||null===e||"undefined"===typeof e)throw new Error(t)}function d(e,t){if(!e){"undefined"!==typeof console&&console.warn(t);try{throw new Error(t)}catch(n){}}}function f(e,t){return{usr:e.state,key:e.key,idx:t}}function p(e,t,n,r){return void 0===n&&(n=null),s({pathname:"string"===typeof e?e:e.pathname,search:"",hash:""},"string"===typeof t?m(t):t,{state:n,key:t&&t.key||r||Math.random().toString(36).substr(2,8)})}function h(e){let{pathname:t="/",search:n="",hash:r=""}=e;return n&&"?"!==n&&(t+="?"===n.charAt(0)?n:"?"+n),r&&"#"!==r&&(t+="#"===r.charAt(0)?r:"#"+r),t}function m(e){let t={};if(e){let n=e.indexOf("#");n>=0&&(t.hash=e.substr(n),e=e.substr(0,n));let r=e.indexOf("?");r>=0&&(t.search=e.substr(r),e=e.substr(0,r)),e&&(t.pathname=e)}return t}function v(e,n,r,a){void 0===a&&(a={});let{window:o=document.defaultView,v5Compat:l=!1}=a,i=o.history,d=t.Pop,m=null,v=g();function g(){return(i.state||{idx:null}).idx}function y(){d=t.Pop;let e=g(),n=null==e?null:e-v;v=e,m&&m({action:d,location:w.location,delta:n})}function b(e){let t="null"!==o.location.origin?o.location.origin:o.location.href,n="string"===typeof e?e:h(e);return n=n.replace(/ $/,"%20"),c(t,"No window.location.(origin|href) available to create URL for href: "+n),new URL(n,t)}null==v&&(v=0,i.replaceState(s({},i.state,{idx:v}),""));let w={get action(){return d},get location(){return e(o,i)},listen(e){if(m)throw new Error("A history only accepts one active listener");return o.addEventListener(u,y),m=e,()=>{o.removeEventListener(u,y),m=null}},createHref:e=>n(o,e),createURL:b,encodeLocation(e){let t=b(e);return{pathname:t.pathname,search:t.search,hash:t.hash}},push:function(e,n){d=t.Push;let a=p(w.location,e,n);r&&r(a,e),v=g()+1;let s=f(a,v),u=w.createHref(a);try{i.pushState(s,"",u)}catch(c){if(c instanceof DOMException&&"DataCloneError"===c.name)throw c;o.location.assign(u)}l&&m&&m({action:d,location:w.location,delta:1})},replace:function(e,n){d=t.Replace;let a=p(w.location,e,n);r&&r(a,e),v=g();let o=f(a,v),s=w.createHref(a);i.replaceState(o,"",s),l&&m&&m({action:d,location:w.location,delta:0})},go:e=>i.go(e)};return w}var g;!function(e){e.data="data",e.deferred="deferred",e.redirect="redirect",e.error="error"}(g||(g={}));new Set(["lazy","caseSensitive","path","id","index","children"]);function y(e,t,n){return void 0===n&&(n="/"),b(e,t,n,!1)}function b(e,t,n,r){let a=L(("string"===typeof t?m(t):t).pathname||"/",n);if(null==a)return null;let o=w(e);!function(e){e.sort(((e,t)=>e.score!==t.score?t.score-e.score:function(e,t){let n=e.length===t.length&&e.slice(0,-1).every(((e,n)=>e===t[n]));return n?e[e.length-1]-t[t.length-1]:0}(e.routesMeta.map((e=>e.childrenIndex)),t.routesMeta.map((e=>e.childrenIndex)))))}(o);let l=null;for(let i=0;null==l&&i<o.length;++i){let e=P(a);l=_(o[i],e,r)}return l}function w(e,t,n,r){void 0===t&&(t=[]),void 0===n&&(n=[]),void 0===r&&(r="");let a=(e,a,o)=>{let l={relativePath:void 0===o?e.path||"":o,caseSensitive:!0===e.caseSensitive,childrenIndex:a,route:e};l.relativePath.startsWith("/")&&(c(l.relativePath.startsWith(r),'Absolute route path "'+l.relativePath+'" nested under path "'+r+'" is not valid. An absolute child route path must start with the combined path of all its parent routes.'),l.relativePath=l.relativePath.slice(r.length));let i=z([r,l.relativePath]),s=n.concat(l);e.children&&e.children.length>0&&(c(!0!==e.index,'Index routes must not have child routes. Please remove all child routes from route path "'+i+'".'),w(e.children,t,s,i)),(null!=e.path||e.index)&&t.push({path:i,score:j(i,e.index),routesMeta:s})};return e.forEach(((e,t)=>{var n;if(""!==e.path&&null!=(n=e.path)&&n.includes("?"))for(let r of x(e.path))a(e,t,r);else a(e,t)})),t}function x(e){let t=e.split("/");if(0===t.length)return[];let[n,...r]=t,a=n.endsWith("?"),o=n.replace(/\?$/,"");if(0===r.length)return a?[o,""]:[o];let l=x(r.join("/")),i=[];return i.push(...l.map((e=>""===e?o:[o,e].join("/")))),a&&i.push(...l),i.map((t=>e.startsWith("/")&&""===t?"/":t))}const E=/^:[\w-]+$/,S=3,k=2,C=1,N=10,O=-2,R=e=>"*"===e;function j(e,t){let n=e.split("/"),r=n.length;return n.some(R)&&(r+=O),t&&(r+=k),n.filter((e=>!R(e))).reduce(((e,t)=>e+(E.test(t)?S:""===t?C:N)),r)}function _(e,t,n){void 0===n&&(n=!1);let{routesMeta:r}=e,a={},o="/",l=[];for(let i=0;i<r.length;++i){let e=r[i],s=i===r.length-1,u="/"===o?t:t.slice(o.length)||"/",c=T({path:e.relativePath,caseSensitive:e.caseSensitive,end:s},u),d=e.route;if(!c&&s&&n&&!r[r.length-1].route.index&&(c=T({path:e.relativePath,caseSensitive:e.caseSensitive,end:!1},u)),!c)return null;Object.assign(a,c.params),l.push({params:a,pathname:z([o,c.pathname]),pathnameBase:I(z([o,c.pathnameBase])),route:d}),"/"!==c.pathnameBase&&(o=z([o,c.pathnameBase]))}return l}function T(e,t){"string"===typeof e&&(e={path:e,caseSensitive:!1,end:!0});let[n,r]=function(e,t,n){void 0===t&&(t=!1);void 0===n&&(n=!0);d("*"===e||!e.endsWith("*")||e.endsWith("/*"),'Route path "'+e+'" will be treated as if it were "'+e.replace(/\*$/,"/*")+'" because the `*` character must always follow a `/` in the pattern. To get rid of this warning, please change the route path to "'+e.replace(/\*$/,"/*")+'".');let r=[],a="^"+e.replace(/\/*\*?$/,"").replace(/^\/*/,"/").replace(/[\\.*+^${}|()[\]]/g,"\\$&").replace(/\/:([\w-]+)(\?)?/g,((e,t,n)=>(r.push({paramName:t,isOptional:null!=n}),n?"/?([^\\/]+)?":"/([^\\/]+)")));e.endsWith("*")?(r.push({paramName:"*"}),a+="*"===e||"/*"===e?"(.*)$":"(?:\\/(.+)|\\/*)$"):n?a+="\\/*$":""!==e&&"/"!==e&&(a+="(?:(?=\\/|$))");let o=new RegExp(a,t?void 0:"i");return[o,r]}(e.path,e.caseSensitive,e.end),a=t.match(n);if(!a)return null;let o=a[0],l=o.replace(/(.)\/+$/,"$1"),i=a.slice(1);return{params:r.reduce(((e,t,n)=>{let{paramName:r,isOptional:a}=t;if("*"===r){let e=i[n]||"";l=o.slice(0,o.length-e.length).replace(/(.)\/+$/,"$1")}const s=i[n];return e[r]=a&&!s?void 0:(s||"").replace(/%2F/g,"/"),e}),{}),pathname:o,pathnameBase:l,pattern:e}}function P(e){try{return e.split("/").map((e=>decodeURIComponent(e).replace(/\//g,"%2F"))).join("/")}catch(t){return d(!1,'The URL path "'+e+'" could not be decoded because it is is a malformed URL segment. This is probably due to a bad percent encoding ('+t+")."),e}}function L(e,t){if("/"===t)return e;if(!e.toLowerCase().startsWith(t.toLowerCase()))return null;let n=t.endsWith("/")?t.length-1:t.length,r=e.charAt(n);return r&&"/"!==r?null:e.slice(n)||"/"}function A(e,t,n,r){return"Cannot include a '"+e+"' character in a manually specified `to."+t+"` field ["+JSON.stringify(r)+"].  Please separate it out to the `to."+n+'` field. Alternatively you may provide the full path as a string in <Link to="..."> and the router will parse it for you.'}function D(e){return e.filter(((e,t)=>0===t||e.route.path&&e.route.path.length>0))}function F(e,t){let n=D(e);return t?n.map(((e,t)=>t===n.length-1?e.pathname:e.pathnameBase)):n.map((e=>e.pathnameBase))}function U(e,t,n,r){let a;void 0===r&&(r=!1),"string"===typeof e?a=m(e):(a=s({},e),c(!a.pathname||!a.pathname.includes("?"),A("?","pathname","search",a)),c(!a.pathname||!a.pathname.includes("#"),A("#","pathname","hash",a)),c(!a.search||!a.search.includes("#"),A("#","search","hash",a)));let o,l=""===e||""===a.pathname,i=l?"/":a.pathname;if(null==i)o=n;else{let e=t.length-1;if(!r&&i.startsWith("..")){let t=i.split("/");for(;".."===t[0];)t.shift(),e-=1;a.pathname=t.join("/")}o=e>=0?t[e]:"/"}let u=function(e,t){void 0===t&&(t="/");let{pathname:n,search:r="",hash:a=""}="string"===typeof e?m(e):e,o=n?n.startsWith("/")?n:function(e,t){let n=t.replace(/\/+$/,"").split("/");return e.split("/").forEach((e=>{".."===e?n.length>1&&n.pop():"."!==e&&n.push(e)})),n.length>1?n.join("/"):"/"}(n,t):t;return{pathname:o,search:M(r),hash:B(a)}}(a,o),d=i&&"/"!==i&&i.endsWith("/"),f=(l||"."===i)&&n.endsWith("/");return u.pathname.endsWith("/")||!d&&!f||(u.pathname+="/"),u}const z=e=>e.join("/").replace(/\/\/+/g,"/"),I=e=>e.replace(/\/+$/,"").replace(/^\/*/,"/"),M=e=>e&&"?"!==e?e.startsWith("?")?e:"?"+e:"",B=e=>e&&"#"!==e?e.startsWith("#")?e:"#"+e:"";Error;function $(e){return null!=e&&"number"===typeof e.status&&"string"===typeof e.statusText&&"boolean"===typeof e.internal&&"data"in e}const W=["post","put","patch","delete"],H=(new Set(W),["get",...W]);new Set(H),new Set([301,302,303,307,308]),new Set([307,308]);Symbol("deferred");function V(){return V=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},V.apply(this,arguments)}const q=r.createContext(null);const K=r.createContext(null);const Q=r.createContext(null);const X=r.createContext(null);const G=r.createContext({outlet:null,matches:[],isDataRoute:!1});const J=r.createContext(null);function Y(){return null!=r.useContext(X)}function Z(){return Y()||c(!1),r.useContext(X).location}function ee(e){r.useContext(Q).static||r.useLayoutEffect(e)}function te(){let{isDataRoute:e}=r.useContext(G);return e?function(){let{router:e}=de(ue.UseNavigateStable),t=pe(ce.UseNavigateStable),n=r.useRef(!1);ee((()=>{n.current=!0}));let a=r.useCallback((function(r,a){void 0===a&&(a={}),n.current&&("number"===typeof r?e.navigate(r):e.navigate(r,V({fromRouteId:t},a)))}),[e,t]);return a}():function(){Y()||c(!1);let e=r.useContext(q),{basename:t,future:n,navigator:a}=r.useContext(Q),{matches:o}=r.useContext(G),{pathname:l}=Z(),i=JSON.stringify(F(o,n.v7_relativeSplatPath)),s=r.useRef(!1);ee((()=>{s.current=!0}));let u=r.useCallback((function(n,r){if(void 0===r&&(r={}),!s.current)return;if("number"===typeof n)return void a.go(n);let o=U(n,JSON.parse(i),l,"path"===r.relative);null==e&&"/"!==t&&(o.pathname="/"===o.pathname?t:z([t,o.pathname])),(r.replace?a.replace:a.push)(o,r.state,r)}),[t,a,i,l,e]);return u}()}function ne(e,t){let{relative:n}=void 0===t?{}:t,{future:a}=r.useContext(Q),{matches:o}=r.useContext(G),{pathname:l}=Z(),i=JSON.stringify(F(o,a.v7_relativeSplatPath));return r.useMemo((()=>U(e,JSON.parse(i),l,"path"===n)),[e,i,l,n])}function re(e,n,a,o){Y()||c(!1);let{navigator:l}=r.useContext(Q),{matches:i}=r.useContext(G),s=i[i.length-1],u=s?s.params:{},d=(s&&s.pathname,s?s.pathnameBase:"/");s&&s.route;let f,p=Z();if(n){var h;let e="string"===typeof n?m(n):n;"/"===d||(null==(h=e.pathname)?void 0:h.startsWith(d))||c(!1),f=e}else f=p;let v=f.pathname||"/",g=v;if("/"!==d){let e=d.replace(/^\//,"").split("/");g="/"+v.replace(/^\//,"").split("/").slice(e.length).join("/")}let b=y(e,{pathname:g});let w=se(b&&b.map((e=>Object.assign({},e,{params:Object.assign({},u,e.params),pathname:z([d,l.encodeLocation?l.encodeLocation(e.pathname).pathname:e.pathname]),pathnameBase:"/"===e.pathnameBase?d:z([d,l.encodeLocation?l.encodeLocation(e.pathnameBase).pathname:e.pathnameBase])}))),i,a,o);return n&&w?r.createElement(X.Provider,{value:{location:V({pathname:"/",search:"",hash:"",state:null,key:"default"},f),navigationType:t.Pop}},w):w}function ae(){let e=function(){var e;let t=r.useContext(J),n=fe(ce.UseRouteError),a=pe(ce.UseRouteError);if(void 0!==t)return t;return null==(e=n.errors)?void 0:e[a]}(),t=$(e)?e.status+" "+e.statusText:e instanceof Error?e.message:JSON.stringify(e),n=e instanceof Error?e.stack:null,a="rgba(200,200,200, 0.5)",o={padding:"0.5rem",backgroundColor:a};return r.createElement(r.Fragment,null,r.createElement("h2",null,"Unexpected Application Error!"),r.createElement("h3",{style:{fontStyle:"italic"}},t),n?r.createElement("pre",{style:o},n):null,null)}const oe=r.createElement(ae,null);class le extends r.Component{constructor(e){super(e),this.state={location:e.location,revalidation:e.revalidation,error:e.error}}static getDerivedStateFromError(e){return{error:e}}static getDerivedStateFromProps(e,t){return t.location!==e.location||"idle"!==t.revalidation&&"idle"===e.revalidation?{error:e.error,location:e.location,revalidation:e.revalidation}:{error:void 0!==e.error?e.error:t.error,location:t.location,revalidation:e.revalidation||t.revalidation}}componentDidCatch(e,t){console.error("React Router caught the following error during render",e,t)}render(){return void 0!==this.state.error?r.createElement(G.Provider,{value:this.props.routeContext},r.createElement(J.Provider,{value:this.state.error,children:this.props.component})):this.props.children}}function ie(e){let{routeContext:t,match:n,children:a}=e,o=r.useContext(q);return o&&o.static&&o.staticContext&&(n.route.errorElement||n.route.ErrorBoundary)&&(o.staticContext._deepestRenderedBoundaryId=n.route.id),r.createElement(G.Provider,{value:t},a)}function se(e,t,n,a){var o;if(void 0===t&&(t=[]),void 0===n&&(n=null),void 0===a&&(a=null),null==e){var l;if(!n)return null;if(n.errors)e=n.matches;else{if(!(null!=(l=a)&&l.v7_partialHydration&&0===t.length&&!n.initialized&&n.matches.length>0))return null;e=n.matches}}let i=e,s=null==(o=n)?void 0:o.errors;if(null!=s){let e=i.findIndex((e=>e.route.id&&void 0!==(null==s?void 0:s[e.route.id])));e>=0||c(!1),i=i.slice(0,Math.min(i.length,e+1))}let u=!1,d=-1;if(n&&a&&a.v7_partialHydration)for(let r=0;r<i.length;r++){let e=i[r];if((e.route.HydrateFallback||e.route.hydrateFallbackElement)&&(d=r),e.route.id){let{loaderData:t,errors:r}=n,a=e.route.loader&&void 0===t[e.route.id]&&(!r||void 0===r[e.route.id]);if(e.route.lazy||a){u=!0,i=d>=0?i.slice(0,d+1):[i[0]];break}}}return i.reduceRight(((e,a,o)=>{let l,c=!1,f=null,p=null;var h;n&&(l=s&&a.route.id?s[a.route.id]:void 0,f=a.route.errorElement||oe,u&&(d<0&&0===o?(h="route-fallback",!1||he[h]||(he[h]=!0),c=!0,p=null):d===o&&(c=!0,p=a.route.hydrateFallbackElement||null)));let m=t.concat(i.slice(0,o+1)),v=()=>{let t;return t=l?f:c?p:a.route.Component?r.createElement(a.route.Component,null):a.route.element?a.route.element:e,r.createElement(ie,{match:a,routeContext:{outlet:e,matches:m,isDataRoute:null!=n},children:t})};return n&&(a.route.ErrorBoundary||a.route.errorElement||0===o)?r.createElement(le,{location:n.location,revalidation:n.revalidation,component:f,error:l,children:v(),routeContext:{outlet:null,matches:m,isDataRoute:!0}}):v()}),null)}var ue=function(e){return e.UseBlocker="useBlocker",e.UseRevalidator="useRevalidator",e.UseNavigateStable="useNavigate",e}(ue||{}),ce=function(e){return e.UseBlocker="useBlocker",e.UseLoaderData="useLoaderData",e.UseActionData="useActionData",e.UseRouteError="useRouteError",e.UseNavigation="useNavigation",e.UseRouteLoaderData="useRouteLoaderData",e.UseMatches="useMatches",e.UseRevalidator="useRevalidator",e.UseNavigateStable="useNavigate",e.UseRouteId="useRouteId",e}(ce||{});function de(e){let t=r.useContext(q);return t||c(!1),t}function fe(e){let t=r.useContext(K);return t||c(!1),t}function pe(e){let t=function(){let e=r.useContext(G);return e||c(!1),e}(),n=t.matches[t.matches.length-1];return n.route.id||c(!1),n.route.id}const he={};a.startTransition;function me(e){c(!1)}function ve(e){let{basename:n="/",children:a=null,location:o,navigationType:l=t.Pop,navigator:i,static:s=!1,future:u}=e;Y()&&c(!1);let d=n.replace(/^\/*/,"/"),f=r.useMemo((()=>({basename:d,navigator:i,static:s,future:V({v7_relativeSplatPath:!1},u)})),[d,u,i,s]);"string"===typeof o&&(o=m(o));let{pathname:p="/",search:h="",hash:v="",state:g=null,key:y="default"}=o,b=r.useMemo((()=>{let e=L(p,d);return null==e?null:{location:{pathname:e,search:h,hash:v,state:g,key:y},navigationType:l}}),[d,p,h,v,g,y,l]);return null==b?null:r.createElement(Q.Provider,{value:f},r.createElement(X.Provider,{children:a,value:b}))}function ge(e){let{children:t,location:n}=e;return re(ye(t),n)}new Promise((()=>{}));r.Component;function ye(e,t){void 0===t&&(t=[]);let n=[];return r.Children.forEach(e,((e,a)=>{if(!r.isValidElement(e))return;let o=[...t,a];if(e.type===r.Fragment)return void n.push.apply(n,ye(e.props.children,o));e.type!==me&&c(!1),e.props.index&&e.props.children&&c(!1);let l={id:e.props.id||o.join("-"),caseSensitive:e.props.caseSensitive,element:e.props.element,Component:e.props.Component,index:e.props.index,path:e.props.path,loader:e.props.loader,action:e.props.action,errorElement:e.props.errorElement,ErrorBoundary:e.props.ErrorBoundary,hasErrorBoundary:null!=e.props.ErrorBoundary||null!=e.props.errorElement,shouldRevalidate:e.props.shouldRevalidate,handle:e.props.handle,lazy:e.props.lazy};e.props.children&&(l.children=ye(e.props.children,o)),n.push(l)})),n}function be(){return be=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},be.apply(this,arguments)}function we(e,t){if(null==e)return{};var n,r,a={},o=Object.keys(e);for(r=0;r<o.length;r++)n=o[r],t.indexOf(n)>=0||(a[n]=e[n]);return a}new Set(["application/x-www-form-urlencoded","multipart/form-data","text/plain"]);const xe=["onClick","relative","reloadDocument","replace","state","target","to","preventScrollReset","unstable_viewTransition"];try{window.__reactRouterVersion="6"}catch(il){}new Map;const Ee=a.startTransition;i.flushSync,a.useId;function Se(e){let{basename:t,children:n,future:a,window:o}=e,l=r.useRef();null==l.current&&(l.current=function(e){return void 0===e&&(e={}),v((function(e,t){let{pathname:n,search:r,hash:a}=e.location;return p("",{pathname:n,search:r,hash:a},t.state&&t.state.usr||null,t.state&&t.state.key||"default")}),(function(e,t){return"string"===typeof t?t:h(t)}),null,e)}({window:o,v5Compat:!0}));let i=l.current,[s,u]=r.useState({action:i.action,location:i.location}),{v7_startTransition:c}=a||{},d=r.useCallback((e=>{c&&Ee?Ee((()=>u(e))):u(e)}),[u,c]);return r.useLayoutEffect((()=>i.listen(d)),[i,d]),r.createElement(ve,{basename:t,children:n,location:s.location,navigationType:s.action,navigator:i,future:a})}const ke="undefined"!==typeof window&&"undefined"!==typeof window.document&&"undefined"!==typeof window.document.createElement,Ce=/^(?:[a-z][a-z0-9+.-]*:|\/\/)/i,Ne=r.forwardRef((function(e,t){let n,{onClick:a,relative:o,reloadDocument:l,replace:i,state:s,target:u,to:d,preventScrollReset:f,unstable_viewTransition:p}=e,m=we(e,xe),{basename:v}=r.useContext(Q),g=!1;if("string"===typeof d&&Ce.test(d)&&(n=d,ke))try{let e=new URL(window.location.href),t=d.startsWith("//")?new URL(e.protocol+d):new URL(d),n=L(t.pathname,v);t.origin===e.origin&&null!=n?d=n+t.search+t.hash:g=!0}catch(il){}let y=function(e,t){let{relative:n}=void 0===t?{}:t;Y()||c(!1);let{basename:a,navigator:o}=r.useContext(Q),{hash:l,pathname:i,search:s}=ne(e,{relative:n}),u=i;return"/"!==a&&(u="/"===i?a:z([a,i])),o.createHref({pathname:u,search:s,hash:l})}(d,{relative:o}),b=function(e,t){let{target:n,replace:a,state:o,preventScrollReset:l,relative:i,unstable_viewTransition:s}=void 0===t?{}:t,u=te(),c=Z(),d=ne(e,{relative:i});return r.useCallback((t=>{if(function(e,t){return 0===e.button&&(!t||"_self"===t)&&!function(e){return!!(e.metaKey||e.altKey||e.ctrlKey||e.shiftKey)}(e)}(t,n)){t.preventDefault();let n=void 0!==a?a:h(c)===h(d);u(e,{replace:n,state:o,preventScrollReset:l,relative:i,unstable_viewTransition:s})}}),[c,u,d,a,o,n,e,l,i,s])}(d,{replace:i,state:s,target:u,preventScrollReset:f,relative:o,unstable_viewTransition:p});return r.createElement("a",be({},m,{href:n||y,onClick:g||l?a:function(e){a&&a(e),e.defaultPrevented||b(e)},ref:t,target:u}))}));var Oe,Re;(function(e){e.UseScrollRestoration="useScrollRestoration",e.UseSubmit="useSubmit",e.UseSubmitFetcher="useSubmitFetcher",e.UseFetcher="useFetcher",e.useViewTransitionState="useViewTransitionState"})(Oe||(Oe={})),function(e){e.UseFetcher="useFetcher",e.UseFetchers="useFetchers",e.UseScrollRestoration="useScrollRestoration"}(Re||(Re={}));function je(e,t){return function(){return e.apply(t,arguments)}}const{toString:_e}=Object.prototype,{getPrototypeOf:Te}=Object,Pe=(Le=Object.create(null),e=>{const t=_e.call(e);return Le[t]||(Le[t]=t.slice(8,-1).toLowerCase())});var Le;const Ae=e=>(e=e.toLowerCase(),t=>Pe(t)===e),De=e=>t=>typeof t===e,{isArray:Fe}=Array,Ue=De("undefined");const ze=Ae("ArrayBuffer");const Ie=De("string"),Me=De("function"),Be=De("number"),$e=e=>null!==e&&"object"===typeof e,We=e=>{if("object"!==Pe(e))return!1;const t=Te(e);return(null===t||t===Object.prototype||null===Object.getPrototypeOf(t))&&!(Symbol.toStringTag in e)&&!(Symbol.iterator in e)},He=Ae("Date"),Ve=Ae("File"),qe=Ae("Blob"),Ke=Ae("FileList"),Qe=Ae("URLSearchParams"),[Xe,Ge,Je,Ye]=["ReadableStream","Request","Response","Headers"].map(Ae);function Ze(e,t){let n,r,{allOwnKeys:a=!1}=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{};if(null!==e&&"undefined"!==typeof e)if("object"!==typeof e&&(e=[e]),Fe(e))for(n=0,r=e.length;n<r;n++)t.call(null,e[n],n,e);else{const r=a?Object.getOwnPropertyNames(e):Object.keys(e),o=r.length;let l;for(n=0;n<o;n++)l=r[n],t.call(null,e[l],l,e)}}function et(e,t){t=t.toLowerCase();const n=Object.keys(e);let r,a=n.length;for(;a-- >0;)if(r=n[a],t===r.toLowerCase())return r;return null}const tt="undefined"!==typeof globalThis?globalThis:"undefined"!==typeof self?self:"undefined"!==typeof window?window:global,nt=e=>!Ue(e)&&e!==tt;const rt=(at="undefined"!==typeof Uint8Array&&Te(Uint8Array),e=>at&&e instanceof at);var at;const ot=Ae("HTMLFormElement"),lt=(e=>{let{hasOwnProperty:t}=e;return(e,n)=>t.call(e,n)})(Object.prototype),it=Ae("RegExp"),st=(e,t)=>{const n=Object.getOwnPropertyDescriptors(e),r={};Ze(n,((n,a)=>{let o;!1!==(o=t(n,a,e))&&(r[a]=o||n)})),Object.defineProperties(e,r)},ut="abcdefghijklmnopqrstuvwxyz",ct="0123456789",dt={DIGIT:ct,ALPHA:ut,ALPHA_DIGIT:ut+ut.toUpperCase()+ct};const ft=Ae("AsyncFunction"),pt=((e,t)=>{return e?setImmediate:t?(n=`axios@${Math.random()}`,r=[],tt.addEventListener("message",(e=>{let{source:t,data:a}=e;t===tt&&a===n&&r.length&&r.shift()()}),!1),e=>{r.push(e),tt.postMessage(n,"*")}):e=>setTimeout(e);var n,r})("function"===typeof setImmediate,Me(tt.postMessage)),ht="undefined"!==typeof queueMicrotask?queueMicrotask.bind(tt):"undefined"!==typeof process&&process.nextTick||pt,mt={isArray:Fe,isArrayBuffer:ze,isBuffer:function(e){return null!==e&&!Ue(e)&&null!==e.constructor&&!Ue(e.constructor)&&Me(e.constructor.isBuffer)&&e.constructor.isBuffer(e)},isFormData:e=>{let t;return e&&("function"===typeof FormData&&e instanceof FormData||Me(e.append)&&("formdata"===(t=Pe(e))||"object"===t&&Me(e.toString)&&"[object FormData]"===e.toString()))},isArrayBufferView:function(e){let t;return t="undefined"!==typeof ArrayBuffer&&ArrayBuffer.isView?ArrayBuffer.isView(e):e&&e.buffer&&ze(e.buffer),t},isString:Ie,isNumber:Be,isBoolean:e=>!0===e||!1===e,isObject:$e,isPlainObject:We,isReadableStream:Xe,isRequest:Ge,isResponse:Je,isHeaders:Ye,isUndefined:Ue,isDate:He,isFile:Ve,isBlob:qe,isRegExp:it,isFunction:Me,isStream:e=>$e(e)&&Me(e.pipe),isURLSearchParams:Qe,isTypedArray:rt,isFileList:Ke,forEach:Ze,merge:function e(){const{caseless:t}=nt(this)&&this||{},n={},r=(r,a)=>{const o=t&&et(n,a)||a;We(n[o])&&We(r)?n[o]=e(n[o],r):We(r)?n[o]=e({},r):Fe(r)?n[o]=r.slice():n[o]=r};for(let a=0,o=arguments.length;a<o;a++)arguments[a]&&Ze(arguments[a],r);return n},extend:function(e,t,n){let{allOwnKeys:r}=arguments.length>3&&void 0!==arguments[3]?arguments[3]:{};return Ze(t,((t,r)=>{n&&Me(t)?e[r]=je(t,n):e[r]=t}),{allOwnKeys:r}),e},trim:e=>e.trim?e.trim():e.replace(/^[\s\uFEFF\xA0]+|[\s\uFEFF\xA0]+$/g,""),stripBOM:e=>(65279===e.charCodeAt(0)&&(e=e.slice(1)),e),inherits:(e,t,n,r)=>{e.prototype=Object.create(t.prototype,r),e.prototype.constructor=e,Object.defineProperty(e,"super",{value:t.prototype}),n&&Object.assign(e.prototype,n)},toFlatObject:(e,t,n,r)=>{let a,o,l;const i={};if(t=t||{},null==e)return t;do{for(a=Object.getOwnPropertyNames(e),o=a.length;o-- >0;)l=a[o],r&&!r(l,e,t)||i[l]||(t[l]=e[l],i[l]=!0);e=!1!==n&&Te(e)}while(e&&(!n||n(e,t))&&e!==Object.prototype);return t},kindOf:Pe,kindOfTest:Ae,endsWith:(e,t,n)=>{e=String(e),(void 0===n||n>e.length)&&(n=e.length),n-=t.length;const r=e.indexOf(t,n);return-1!==r&&r===n},toArray:e=>{if(!e)return null;if(Fe(e))return e;let t=e.length;if(!Be(t))return null;const n=new Array(t);for(;t-- >0;)n[t]=e[t];return n},forEachEntry:(e,t)=>{const n=(e&&e[Symbol.iterator]).call(e);let r;for(;(r=n.next())&&!r.done;){const n=r.value;t.call(e,n[0],n[1])}},matchAll:(e,t)=>{let n;const r=[];for(;null!==(n=e.exec(t));)r.push(n);return r},isHTMLForm:ot,hasOwnProperty:lt,hasOwnProp:lt,reduceDescriptors:st,freezeMethods:e=>{st(e,((t,n)=>{if(Me(e)&&-1!==["arguments","caller","callee"].indexOf(n))return!1;const r=e[n];Me(r)&&(t.enumerable=!1,"writable"in t?t.writable=!1:t.set||(t.set=()=>{throw Error("Can not rewrite read-only method '"+n+"'")}))}))},toObjectSet:(e,t)=>{const n={},r=e=>{e.forEach((e=>{n[e]=!0}))};return Fe(e)?r(e):r(String(e).split(t)),n},toCamelCase:e=>e.toLowerCase().replace(/[-_\s]([a-z\d])(\w*)/g,(function(e,t,n){return t.toUpperCase()+n})),noop:()=>{},toFiniteNumber:(e,t)=>null!=e&&Number.isFinite(e=+e)?e:t,findKey:et,global:tt,isContextDefined:nt,ALPHABET:dt,generateString:function(){let e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:16,t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:dt.ALPHA_DIGIT,n="";const{length:r}=t;for(;e--;)n+=t[Math.random()*r|0];return n},isSpecCompliantForm:function(e){return!!(e&&Me(e.append)&&"FormData"===e[Symbol.toStringTag]&&e[Symbol.iterator])},toJSONObject:e=>{const t=new Array(10),n=(e,r)=>{if($e(e)){if(t.indexOf(e)>=0)return;if(!("toJSON"in e)){t[r]=e;const a=Fe(e)?[]:{};return Ze(e,((e,t)=>{const o=n(e,r+1);!Ue(o)&&(a[t]=o)})),t[r]=void 0,a}}return e};return n(e,0)},isAsyncFn:ft,isThenable:e=>e&&($e(e)||Me(e))&&Me(e.then)&&Me(e.catch),setImmediate:pt,asap:ht};function vt(e,t,n,r,a){Error.call(this),Error.captureStackTrace?Error.captureStackTrace(this,this.constructor):this.stack=(new Error).stack,this.message=e,this.name="AxiosError",t&&(this.code=t),n&&(this.config=n),r&&(this.request=r),a&&(this.response=a,this.status=a.status?a.status:null)}mt.inherits(vt,Error,{toJSON:function(){return{message:this.message,name:this.name,description:this.description,number:this.number,fileName:this.fileName,lineNumber:this.lineNumber,columnNumber:this.columnNumber,stack:this.stack,config:mt.toJSONObject(this.config),code:this.code,status:this.status}}});const gt=vt.prototype,yt={};["ERR_BAD_OPTION_VALUE","ERR_BAD_OPTION","ECONNABORTED","ETIMEDOUT","ERR_NETWORK","ERR_FR_TOO_MANY_REDIRECTS","ERR_DEPRECATED","ERR_BAD_RESPONSE","ERR_BAD_REQUEST","ERR_CANCELED","ERR_NOT_SUPPORT","ERR_INVALID_URL"].forEach((e=>{yt[e]={value:e}})),Object.defineProperties(vt,yt),Object.defineProperty(gt,"isAxiosError",{value:!0}),vt.from=(e,t,n,r,a,o)=>{const l=Object.create(gt);return mt.toFlatObject(e,l,(function(e){return e!==Error.prototype}),(e=>"isAxiosError"!==e)),vt.call(l,e.message,t,n,r,a),l.cause=e,l.name=e.name,o&&Object.assign(l,o),l};const bt=vt;function wt(e){return mt.isPlainObject(e)||mt.isArray(e)}function xt(e){return mt.endsWith(e,"[]")?e.slice(0,-2):e}function Et(e,t,n){return e?e.concat(t).map((function(e,t){return e=xt(e),!n&&t?"["+e+"]":e})).join(n?".":""):t}const St=mt.toFlatObject(mt,{},null,(function(e){return/^is[A-Z]/.test(e)}));const kt=function(e,t,n){if(!mt.isObject(e))throw new TypeError("target must be an object");t=t||new FormData;const r=(n=mt.toFlatObject(n,{metaTokens:!0,dots:!1,indexes:!1},!1,(function(e,t){return!mt.isUndefined(t[e])}))).metaTokens,a=n.visitor||u,o=n.dots,l=n.indexes,i=(n.Blob||"undefined"!==typeof Blob&&Blob)&&mt.isSpecCompliantForm(t);if(!mt.isFunction(a))throw new TypeError("visitor must be a function");function s(e){if(null===e)return"";if(mt.isDate(e))return e.toISOString();if(!i&&mt.isBlob(e))throw new bt("Blob is not supported. Use a Buffer instead.");return mt.isArrayBuffer(e)||mt.isTypedArray(e)?i&&"function"===typeof Blob?new Blob([e]):Buffer.from(e):e}function u(e,n,a){let i=e;if(e&&!a&&"object"===typeof e)if(mt.endsWith(n,"{}"))n=r?n:n.slice(0,-2),e=JSON.stringify(e);else if(mt.isArray(e)&&function(e){return mt.isArray(e)&&!e.some(wt)}(e)||(mt.isFileList(e)||mt.endsWith(n,"[]"))&&(i=mt.toArray(e)))return n=xt(n),i.forEach((function(e,r){!mt.isUndefined(e)&&null!==e&&t.append(!0===l?Et([n],r,o):null===l?n:n+"[]",s(e))})),!1;return!!wt(e)||(t.append(Et(a,n,o),s(e)),!1)}const c=[],d=Object.assign(St,{defaultVisitor:u,convertValue:s,isVisitable:wt});if(!mt.isObject(e))throw new TypeError("data must be an object");return function e(n,r){if(!mt.isUndefined(n)){if(-1!==c.indexOf(n))throw Error("Circular reference detected in "+r.join("."));c.push(n),mt.forEach(n,(function(n,o){!0===(!(mt.isUndefined(n)||null===n)&&a.call(t,n,mt.isString(o)?o.trim():o,r,d))&&e(n,r?r.concat(o):[o])})),c.pop()}}(e),t};function Ct(e){const t={"!":"%21","'":"%27","(":"%28",")":"%29","~":"%7E","%20":"+","%00":"\0"};return encodeURIComponent(e).replace(/[!'()~]|%20|%00/g,(function(e){return t[e]}))}function Nt(e,t){this._pairs=[],e&&kt(e,this,t)}const Ot=Nt.prototype;Ot.append=function(e,t){this._pairs.push([e,t])},Ot.toString=function(e){const t=e?function(t){return e.call(this,t,Ct)}:Ct;return this._pairs.map((function(e){return t(e[0])+"="+t(e[1])}),"").join("&")};const Rt=Nt;function jt(e){return encodeURIComponent(e).replace(/%3A/gi,":").replace(/%24/g,"$").replace(/%2C/gi,",").replace(/%20/g,"+").replace(/%5B/gi,"[").replace(/%5D/gi,"]")}function _t(e,t,n){if(!t)return e;const r=n&&n.encode||jt,a=n&&n.serialize;let o;if(o=a?a(t,n):mt.isURLSearchParams(t)?t.toString():new Rt(t,n).toString(r),o){const t=e.indexOf("#");-1!==t&&(e=e.slice(0,t)),e+=(-1===e.indexOf("?")?"?":"&")+o}return e}const Tt=class{constructor(){this.handlers=[]}use(e,t,n){return this.handlers.push({fulfilled:e,rejected:t,synchronous:!!n&&n.synchronous,runWhen:n?n.runWhen:null}),this.handlers.length-1}eject(e){this.handlers[e]&&(this.handlers[e]=null)}clear(){this.handlers&&(this.handlers=[])}forEach(e){mt.forEach(this.handlers,(function(t){null!==t&&e(t)}))}},Pt={silentJSONParsing:!0,forcedJSONParsing:!0,clarifyTimeoutError:!1},Lt={isBrowser:!0,classes:{URLSearchParams:"undefined"!==typeof URLSearchParams?URLSearchParams:Rt,FormData:"undefined"!==typeof FormData?FormData:null,Blob:"undefined"!==typeof Blob?Blob:null},protocols:["http","https","file","blob","url","data"]},At="undefined"!==typeof window&&"undefined"!==typeof document,Dt="object"===typeof navigator&&navigator||void 0,Ft=At&&(!Dt||["ReactNative","NativeScript","NS"].indexOf(Dt.product)<0),Ut="undefined"!==typeof WorkerGlobalScope&&self instanceof WorkerGlobalScope&&"function"===typeof self.importScripts,zt=At&&window.location.href||"http://localhost",It={...e,...Lt};const Mt=function(e){function t(e,n,r,a){let o=e[a++];if("__proto__"===o)return!0;const l=Number.isFinite(+o),i=a>=e.length;if(o=!o&&mt.isArray(r)?r.length:o,i)return mt.hasOwnProp(r,o)?r[o]=[r[o],n]:r[o]=n,!l;r[o]&&mt.isObject(r[o])||(r[o]=[]);return t(e,n,r[o],a)&&mt.isArray(r[o])&&(r[o]=function(e){const t={},n=Object.keys(e);let r;const a=n.length;let o;for(r=0;r<a;r++)o=n[r],t[o]=e[o];return t}(r[o])),!l}if(mt.isFormData(e)&&mt.isFunction(e.entries)){const n={};return mt.forEachEntry(e,((e,r)=>{t(function(e){return mt.matchAll(/\w+|\[(\w*)]/g,e).map((e=>"[]"===e[0]?"":e[1]||e[0]))}(e),r,n,0)})),n}return null};const Bt={transitional:Pt,adapter:["xhr","http","fetch"],transformRequest:[function(e,t){const n=t.getContentType()||"",r=n.indexOf("application/json")>-1,a=mt.isObject(e);a&&mt.isHTMLForm(e)&&(e=new FormData(e));if(mt.isFormData(e))return r?JSON.stringify(Mt(e)):e;if(mt.isArrayBuffer(e)||mt.isBuffer(e)||mt.isStream(e)||mt.isFile(e)||mt.isBlob(e)||mt.isReadableStream(e))return e;if(mt.isArrayBufferView(e))return e.buffer;if(mt.isURLSearchParams(e))return t.setContentType("application/x-www-form-urlencoded;charset=utf-8",!1),e.toString();let o;if(a){if(n.indexOf("application/x-www-form-urlencoded")>-1)return function(e,t){return kt(e,new It.classes.URLSearchParams,Object.assign({visitor:function(e,t,n,r){return It.isNode&&mt.isBuffer(e)?(this.append(t,e.toString("base64")),!1):r.defaultVisitor.apply(this,arguments)}},t))}(e,this.formSerializer).toString();if((o=mt.isFileList(e))||n.indexOf("multipart/form-data")>-1){const t=this.env&&this.env.FormData;return kt(o?{"files[]":e}:e,t&&new t,this.formSerializer)}}return a||r?(t.setContentType("application/json",!1),function(e,t,n){if(mt.isString(e))try{return(t||JSON.parse)(e),mt.trim(e)}catch(il){if("SyntaxError"!==il.name)throw il}return(n||JSON.stringify)(e)}(e)):e}],transformResponse:[function(e){const t=this.transitional||Bt.transitional,n=t&&t.forcedJSONParsing,r="json"===this.responseType;if(mt.isResponse(e)||mt.isReadableStream(e))return e;if(e&&mt.isString(e)&&(n&&!this.responseType||r)){const n=!(t&&t.silentJSONParsing)&&r;try{return JSON.parse(e)}catch(il){if(n){if("SyntaxError"===il.name)throw bt.from(il,bt.ERR_BAD_RESPONSE,this,null,this.response);throw il}}}return e}],timeout:0,xsrfCookieName:"XSRF-TOKEN",xsrfHeaderName:"X-XSRF-TOKEN",maxContentLength:-1,maxBodyLength:-1,env:{FormData:It.classes.FormData,Blob:It.classes.Blob},validateStatus:function(e){return e>=200&&e<300},headers:{common:{Accept:"application/json, text/plain, */*","Content-Type":void 0}}};mt.forEach(["delete","get","head","post","put","patch"],(e=>{Bt.headers[e]={}}));const $t=Bt,Wt=mt.toObjectSet(["age","authorization","content-length","content-type","etag","expires","from","host","if-modified-since","if-unmodified-since","last-modified","location","max-forwards","proxy-authorization","referer","retry-after","user-agent"]),Ht=Symbol("internals");function Vt(e){return e&&String(e).trim().toLowerCase()}function qt(e){return!1===e||null==e?e:mt.isArray(e)?e.map(qt):String(e)}function Kt(e,t,n,r,a){return mt.isFunction(r)?r.call(this,t,n):(a&&(t=n),mt.isString(t)?mt.isString(r)?-1!==t.indexOf(r):mt.isRegExp(r)?r.test(t):void 0:void 0)}class Qt{constructor(e){e&&this.set(e)}set(e,t,n){const r=this;function a(e,t,n){const a=Vt(t);if(!a)throw new Error("header name must be a non-empty string");const o=mt.findKey(r,a);(!o||void 0===r[o]||!0===n||void 0===n&&!1!==r[o])&&(r[o||t]=qt(e))}const o=(e,t)=>mt.forEach(e,((e,n)=>a(e,n,t)));if(mt.isPlainObject(e)||e instanceof this.constructor)o(e,t);else if(mt.isString(e)&&(e=e.trim())&&!/^[-_a-zA-Z0-9^`|~,!#$%&'*+.]+$/.test(e.trim()))o((e=>{const t={};let n,r,a;return e&&e.split("\n").forEach((function(e){a=e.indexOf(":"),n=e.substring(0,a).trim().toLowerCase(),r=e.substring(a+1).trim(),!n||t[n]&&Wt[n]||("set-cookie"===n?t[n]?t[n].push(r):t[n]=[r]:t[n]=t[n]?t[n]+", "+r:r)})),t})(e),t);else if(mt.isHeaders(e))for(const[l,i]of e.entries())a(i,l,n);else null!=e&&a(t,e,n);return this}get(e,t){if(e=Vt(e)){const n=mt.findKey(this,e);if(n){const e=this[n];if(!t)return e;if(!0===t)return function(e){const t=Object.create(null),n=/([^\s,;=]+)\s*(?:=\s*([^,;]+))?/g;let r;for(;r=n.exec(e);)t[r[1]]=r[2];return t}(e);if(mt.isFunction(t))return t.call(this,e,n);if(mt.isRegExp(t))return t.exec(e);throw new TypeError("parser must be boolean|regexp|function")}}}has(e,t){if(e=Vt(e)){const n=mt.findKey(this,e);return!(!n||void 0===this[n]||t&&!Kt(0,this[n],n,t))}return!1}delete(e,t){const n=this;let r=!1;function a(e){if(e=Vt(e)){const a=mt.findKey(n,e);!a||t&&!Kt(0,n[a],a,t)||(delete n[a],r=!0)}}return mt.isArray(e)?e.forEach(a):a(e),r}clear(e){const t=Object.keys(this);let n=t.length,r=!1;for(;n--;){const a=t[n];e&&!Kt(0,this[a],a,e,!0)||(delete this[a],r=!0)}return r}normalize(e){const t=this,n={};return mt.forEach(this,((r,a)=>{const o=mt.findKey(n,a);if(o)return t[o]=qt(r),void delete t[a];const l=e?function(e){return e.trim().toLowerCase().replace(/([a-z\d])(\w*)/g,((e,t,n)=>t.toUpperCase()+n))}(a):String(a).trim();l!==a&&delete t[a],t[l]=qt(r),n[l]=!0})),this}concat(){for(var e=arguments.length,t=new Array(e),n=0;n<e;n++)t[n]=arguments[n];return this.constructor.concat(this,...t)}toJSON(e){const t=Object.create(null);return mt.forEach(this,((n,r)=>{null!=n&&!1!==n&&(t[r]=e&&mt.isArray(n)?n.join(", "):n)})),t}[Symbol.iterator](){return Object.entries(this.toJSON())[Symbol.iterator]()}toString(){return Object.entries(this.toJSON()).map((e=>{let[t,n]=e;return t+": "+n})).join("\n")}get[Symbol.toStringTag](){return"AxiosHeaders"}static from(e){return e instanceof this?e:new this(e)}static concat(e){const t=new this(e);for(var n=arguments.length,r=new Array(n>1?n-1:0),a=1;a<n;a++)r[a-1]=arguments[a];return r.forEach((e=>t.set(e))),t}static accessor(e){const t=(this[Ht]=this[Ht]={accessors:{}}).accessors,n=this.prototype;function r(e){const r=Vt(e);t[r]||(!function(e,t){const n=mt.toCamelCase(" "+t);["get","set","has"].forEach((r=>{Object.defineProperty(e,r+n,{value:function(e,n,a){return this[r].call(this,t,e,n,a)},configurable:!0})}))}(n,e),t[r]=!0)}return mt.isArray(e)?e.forEach(r):r(e),this}}Qt.accessor(["Content-Type","Content-Length","Accept","Accept-Encoding","User-Agent","Authorization"]),mt.reduceDescriptors(Qt.prototype,((e,t)=>{let{value:n}=e,r=t[0].toUpperCase()+t.slice(1);return{get:()=>n,set(e){this[r]=e}}})),mt.freezeMethods(Qt);const Xt=Qt;function Gt(e,t){const n=this||$t,r=t||n,a=Xt.from(r.headers);let o=r.data;return mt.forEach(e,(function(e){o=e.call(n,o,a.normalize(),t?t.status:void 0)})),a.normalize(),o}function Jt(e){return!(!e||!e.__CANCEL__)}function Yt(e,t,n){bt.call(this,null==e?"canceled":e,bt.ERR_CANCELED,t,n),this.name="CanceledError"}mt.inherits(Yt,bt,{__CANCEL__:!0});const Zt=Yt;function en(e,t,n){const r=n.config.validateStatus;n.status&&r&&!r(n.status)?t(new bt("Request failed with status code "+n.status,[bt.ERR_BAD_REQUEST,bt.ERR_BAD_RESPONSE][Math.floor(n.status/100)-4],n.config,n.request,n)):e(n)}const tn=function(e,t){e=e||10;const n=new Array(e),r=new Array(e);let a,o=0,l=0;return t=void 0!==t?t:1e3,function(i){const s=Date.now(),u=r[l];a||(a=s),n[o]=i,r[o]=s;let c=l,d=0;for(;c!==o;)d+=n[c++],c%=e;if(o=(o+1)%e,o===l&&(l=(l+1)%e),s-a<t)return;const f=u&&s-u;return f?Math.round(1e3*d/f):void 0}};const nn=function(e,t){let n,r,a=0,o=1e3/t;const l=function(t){let o=arguments.length>1&&void 0!==arguments[1]?arguments[1]:Date.now();a=o,n=null,r&&(clearTimeout(r),r=null),e.apply(null,t)};return[function(){const e=Date.now(),t=e-a;for(var i=arguments.length,s=new Array(i),u=0;u<i;u++)s[u]=arguments[u];t>=o?l(s,e):(n=s,r||(r=setTimeout((()=>{r=null,l(n)}),o-t)))},()=>n&&l(n)]},rn=function(e,t){let n=arguments.length>2&&void 0!==arguments[2]?arguments[2]:3,r=0;const a=tn(50,250);return nn((n=>{const o=n.loaded,l=n.lengthComputable?n.total:void 0,i=o-r,s=a(i);r=o;e({loaded:o,total:l,progress:l?o/l:void 0,bytes:i,rate:s||void 0,estimated:s&&l&&o<=l?(l-o)/s:void 0,event:n,lengthComputable:null!=l,[t?"download":"upload"]:!0})}),n)},an=(e,t)=>{const n=null!=e;return[r=>t[0]({lengthComputable:n,total:e,loaded:r}),t[1]]},on=e=>function(){for(var t=arguments.length,n=new Array(t),r=0;r<t;r++)n[r]=arguments[r];return mt.asap((()=>e(...n)))},ln=It.hasStandardBrowserEnv?function(){const e=It.navigator&&/(msie|trident)/i.test(It.navigator.userAgent),t=document.createElement("a");let n;function r(n){let r=n;return e&&(t.setAttribute("href",r),r=t.href),t.setAttribute("href",r),{href:t.href,protocol:t.protocol?t.protocol.replace(/:$/,""):"",host:t.host,search:t.search?t.search.replace(/^\?/,""):"",hash:t.hash?t.hash.replace(/^#/,""):"",hostname:t.hostname,port:t.port,pathname:"/"===t.pathname.charAt(0)?t.pathname:"/"+t.pathname}}return n=r(window.location.href),function(e){const t=mt.isString(e)?r(e):e;return t.protocol===n.protocol&&t.host===n.host}}():function(){return!0},sn=It.hasStandardBrowserEnv?{write(e,t,n,r,a,o){const l=[e+"="+encodeURIComponent(t)];mt.isNumber(n)&&l.push("expires="+new Date(n).toGMTString()),mt.isString(r)&&l.push("path="+r),mt.isString(a)&&l.push("domain="+a),!0===o&&l.push("secure"),document.cookie=l.join("; ")},read(e){const t=document.cookie.match(new RegExp("(^|;\\s*)("+e+")=([^;]*)"));return t?decodeURIComponent(t[3]):null},remove(e){this.write(e,"",Date.now()-864e5)}}:{write(){},read:()=>null,remove(){}};function un(e,t){return e&&!/^([a-z][a-z\d+\-.]*:)?\/\//i.test(t)?function(e,t){return t?e.replace(/\/?\/$/,"")+"/"+t.replace(/^\/+/,""):e}(e,t):t}const cn=e=>e instanceof Xt?{...e}:e;function dn(e,t){t=t||{};const n={};function r(e,t,n){return mt.isPlainObject(e)&&mt.isPlainObject(t)?mt.merge.call({caseless:n},e,t):mt.isPlainObject(t)?mt.merge({},t):mt.isArray(t)?t.slice():t}function a(e,t,n){return mt.isUndefined(t)?mt.isUndefined(e)?void 0:r(void 0,e,n):r(e,t,n)}function o(e,t){if(!mt.isUndefined(t))return r(void 0,t)}function l(e,t){return mt.isUndefined(t)?mt.isUndefined(e)?void 0:r(void 0,e):r(void 0,t)}function i(n,a,o){return o in t?r(n,a):o in e?r(void 0,n):void 0}const s={url:o,method:o,data:o,baseURL:l,transformRequest:l,transformResponse:l,paramsSerializer:l,timeout:l,timeoutMessage:l,withCredentials:l,withXSRFToken:l,adapter:l,responseType:l,xsrfCookieName:l,xsrfHeaderName:l,onUploadProgress:l,onDownloadProgress:l,decompress:l,maxContentLength:l,maxBodyLength:l,beforeRedirect:l,transport:l,httpAgent:l,httpsAgent:l,cancelToken:l,socketPath:l,responseEncoding:l,validateStatus:i,headers:(e,t)=>a(cn(e),cn(t),!0)};return mt.forEach(Object.keys(Object.assign({},e,t)),(function(r){const o=s[r]||a,l=o(e[r],t[r],r);mt.isUndefined(l)&&o!==i||(n[r]=l)})),n}const fn=e=>{const t=dn({},e);let n,{data:r,withXSRFToken:a,xsrfHeaderName:o,xsrfCookieName:l,headers:i,auth:s}=t;if(t.headers=i=Xt.from(i),t.url=_t(un(t.baseURL,t.url),e.params,e.paramsSerializer),s&&i.set("Authorization","Basic "+btoa((s.username||"")+":"+(s.password?unescape(encodeURIComponent(s.password)):""))),mt.isFormData(r))if(It.hasStandardBrowserEnv||It.hasStandardBrowserWebWorkerEnv)i.setContentType(void 0);else if(!1!==(n=i.getContentType())){const[e,...t]=n?n.split(";").map((e=>e.trim())).filter(Boolean):[];i.setContentType([e||"multipart/form-data",...t].join("; "))}if(It.hasStandardBrowserEnv&&(a&&mt.isFunction(a)&&(a=a(t)),a||!1!==a&&ln(t.url))){const e=o&&l&&sn.read(l);e&&i.set(o,e)}return t},pn="undefined"!==typeof XMLHttpRequest&&function(e){return new Promise((function(t,n){const r=fn(e);let a=r.data;const o=Xt.from(r.headers).normalize();let l,i,s,u,c,{responseType:d,onUploadProgress:f,onDownloadProgress:p}=r;function h(){u&&u(),c&&c(),r.cancelToken&&r.cancelToken.unsubscribe(l),r.signal&&r.signal.removeEventListener("abort",l)}let m=new XMLHttpRequest;function v(){if(!m)return;const r=Xt.from("getAllResponseHeaders"in m&&m.getAllResponseHeaders());en((function(e){t(e),h()}),(function(e){n(e),h()}),{data:d&&"text"!==d&&"json"!==d?m.response:m.responseText,status:m.status,statusText:m.statusText,headers:r,config:e,request:m}),m=null}m.open(r.method.toUpperCase(),r.url,!0),m.timeout=r.timeout,"onloadend"in m?m.onloadend=v:m.onreadystatechange=function(){m&&4===m.readyState&&(0!==m.status||m.responseURL&&0===m.responseURL.indexOf("file:"))&&setTimeout(v)},m.onabort=function(){m&&(n(new bt("Request aborted",bt.ECONNABORTED,e,m)),m=null)},m.onerror=function(){n(new bt("Network Error",bt.ERR_NETWORK,e,m)),m=null},m.ontimeout=function(){let t=r.timeout?"timeout of "+r.timeout+"ms exceeded":"timeout exceeded";const a=r.transitional||Pt;r.timeoutErrorMessage&&(t=r.timeoutErrorMessage),n(new bt(t,a.clarifyTimeoutError?bt.ETIMEDOUT:bt.ECONNABORTED,e,m)),m=null},void 0===a&&o.setContentType(null),"setRequestHeader"in m&&mt.forEach(o.toJSON(),(function(e,t){m.setRequestHeader(t,e)})),mt.isUndefined(r.withCredentials)||(m.withCredentials=!!r.withCredentials),d&&"json"!==d&&(m.responseType=r.responseType),p&&([s,c]=rn(p,!0),m.addEventListener("progress",s)),f&&m.upload&&([i,u]=rn(f),m.upload.addEventListener("progress",i),m.upload.addEventListener("loadend",u)),(r.cancelToken||r.signal)&&(l=t=>{m&&(n(!t||t.type?new Zt(null,e,m):t),m.abort(),m=null)},r.cancelToken&&r.cancelToken.subscribe(l),r.signal&&(r.signal.aborted?l():r.signal.addEventListener("abort",l)));const g=function(e){const t=/^([-+\w]{1,25})(:?\/\/|:)/.exec(e);return t&&t[1]||""}(r.url);g&&-1===It.protocols.indexOf(g)?n(new bt("Unsupported protocol "+g+":",bt.ERR_BAD_REQUEST,e)):m.send(a||null)}))},hn=(e,t)=>{const{length:n}=e=e?e.filter(Boolean):[];if(t||n){let n,r=new AbortController;const a=function(e){if(!n){n=!0,l();const t=e instanceof Error?e:this.reason;r.abort(t instanceof bt?t:new Zt(t instanceof Error?t.message:t))}};let o=t&&setTimeout((()=>{o=null,a(new bt(`timeout ${t} of ms exceeded`,bt.ETIMEDOUT))}),t);const l=()=>{e&&(o&&clearTimeout(o),o=null,e.forEach((e=>{e.unsubscribe?e.unsubscribe(a):e.removeEventListener("abort",a)})),e=null)};e.forEach((e=>e.addEventListener("abort",a)));const{signal:i}=r;return i.unsubscribe=()=>mt.asap(l),i}},mn=function*(e,t){let n=e.byteLength;if(!t||n<t)return void(yield e);let r,a=0;for(;a<n;)r=a+t,yield e.slice(a,r),a=r},vn=async function*(e){if(e[Symbol.asyncIterator])return void(yield*e);const t=e.getReader();try{for(;;){const{done:e,value:n}=await t.read();if(e)break;yield n}}finally{await t.cancel()}},gn=(e,t,n,r)=>{const a=async function*(e,t){for await(const n of vn(e))yield*mn(n,t)}(e,t);let o,l=0,i=e=>{o||(o=!0,r&&r(e))};return new ReadableStream({async pull(e){try{const{done:t,value:r}=await a.next();if(t)return i(),void e.close();let o=r.byteLength;if(n){let e=l+=o;n(e)}e.enqueue(new Uint8Array(r))}catch(t){throw i(t),t}},cancel:e=>(i(e),a.return())},{highWaterMark:2})},yn="function"===typeof fetch&&"function"===typeof Request&&"function"===typeof Response,bn=yn&&"function"===typeof ReadableStream,wn=yn&&("function"===typeof TextEncoder?(xn=new TextEncoder,e=>xn.encode(e)):async e=>new Uint8Array(await new Response(e).arrayBuffer()));var xn;const En=function(e){try{for(var t=arguments.length,n=new Array(t>1?t-1:0),r=1;r<t;r++)n[r-1]=arguments[r];return!!e(...n)}catch(il){return!1}},Sn=bn&&En((()=>{let e=!1;const t=new Request(It.origin,{body:new ReadableStream,method:"POST",get duplex(){return e=!0,"half"}}).headers.has("Content-Type");return e&&!t})),kn=bn&&En((()=>mt.isReadableStream(new Response("").body))),Cn={stream:kn&&(e=>e.body)};var Nn;yn&&(Nn=new Response,["text","arrayBuffer","blob","formData","stream"].forEach((e=>{!Cn[e]&&(Cn[e]=mt.isFunction(Nn[e])?t=>t[e]():(t,n)=>{throw new bt(`Response type '${e}' is not supported`,bt.ERR_NOT_SUPPORT,n)})})));const On=async(e,t)=>{const n=mt.toFiniteNumber(e.getContentLength());return null==n?(async e=>{if(null==e)return 0;if(mt.isBlob(e))return e.size;if(mt.isSpecCompliantForm(e)){const t=new Request(It.origin,{method:"POST",body:e});return(await t.arrayBuffer()).byteLength}return mt.isArrayBufferView(e)||mt.isArrayBuffer(e)?e.byteLength:(mt.isURLSearchParams(e)&&(e+=""),mt.isString(e)?(await wn(e)).byteLength:void 0)})(t):n},Rn=yn&&(async e=>{let{url:t,method:n,data:r,signal:a,cancelToken:o,timeout:l,onDownloadProgress:i,onUploadProgress:s,responseType:u,headers:c,withCredentials:d="same-origin",fetchOptions:f}=fn(e);u=u?(u+"").toLowerCase():"text";let p,h=hn([a,o&&o.toAbortSignal()],l);const m=h&&h.unsubscribe&&(()=>{h.unsubscribe()});let v;try{if(s&&Sn&&"get"!==n&&"head"!==n&&0!==(v=await On(c,r))){let e,n=new Request(t,{method:"POST",body:r,duplex:"half"});if(mt.isFormData(r)&&(e=n.headers.get("content-type"))&&c.setContentType(e),n.body){const[e,t]=an(v,rn(on(s)));r=gn(n.body,65536,e,t)}}mt.isString(d)||(d=d?"include":"omit");const a="credentials"in Request.prototype;p=new Request(t,{...f,signal:h,method:n.toUpperCase(),headers:c.normalize().toJSON(),body:r,duplex:"half",credentials:a?d:void 0});let o=await fetch(p);const l=kn&&("stream"===u||"response"===u);if(kn&&(i||l&&m)){const e={};["status","statusText","headers"].forEach((t=>{e[t]=o[t]}));const t=mt.toFiniteNumber(o.headers.get("content-length")),[n,r]=i&&an(t,rn(on(i),!0))||[];o=new Response(gn(o.body,65536,n,(()=>{r&&r(),m&&m()})),e)}u=u||"text";let g=await Cn[mt.findKey(Cn,u)||"text"](o,e);return!l&&m&&m(),await new Promise(((t,n)=>{en(t,n,{data:g,headers:Xt.from(o.headers),status:o.status,statusText:o.statusText,config:e,request:p})}))}catch(g){if(m&&m(),g&&"TypeError"===g.name&&/fetch/i.test(g.message))throw Object.assign(new bt("Network Error",bt.ERR_NETWORK,e,p),{cause:g.cause||g});throw bt.from(g,g&&g.code,e,p)}}),jn={http:null,xhr:pn,fetch:Rn};mt.forEach(jn,((e,t)=>{if(e){try{Object.defineProperty(e,"name",{value:t})}catch(il){}Object.defineProperty(e,"adapterName",{value:t})}}));const _n=e=>`- ${e}`,Tn=e=>mt.isFunction(e)||null===e||!1===e,Pn=e=>{e=mt.isArray(e)?e:[e];const{length:t}=e;let n,r;const a={};for(let o=0;o<t;o++){let t;if(n=e[o],r=n,!Tn(n)&&(r=jn[(t=String(n)).toLowerCase()],void 0===r))throw new bt(`Unknown adapter '${t}'`);if(r)break;a[t||"#"+o]=r}if(!r){const e=Object.entries(a).map((e=>{let[t,n]=e;return`adapter ${t} `+(!1===n?"is not supported by the environment":"is not available in the build")}));let n=t?e.length>1?"since :\n"+e.map(_n).join("\n"):" "+_n(e[0]):"as no adapter specified";throw new bt("There is no suitable adapter to dispatch the request "+n,"ERR_NOT_SUPPORT")}return r};function Ln(e){if(e.cancelToken&&e.cancelToken.throwIfRequested(),e.signal&&e.signal.aborted)throw new Zt(null,e)}function An(e){Ln(e),e.headers=Xt.from(e.headers),e.data=Gt.call(e,e.transformRequest),-1!==["post","put","patch"].indexOf(e.method)&&e.headers.setContentType("application/x-www-form-urlencoded",!1);return Pn(e.adapter||$t.adapter)(e).then((function(t){return Ln(e),t.data=Gt.call(e,e.transformResponse,t),t.headers=Xt.from(t.headers),t}),(function(t){return Jt(t)||(Ln(e),t&&t.response&&(t.response.data=Gt.call(e,e.transformResponse,t.response),t.response.headers=Xt.from(t.response.headers))),Promise.reject(t)}))}const Dn="1.7.7",Fn={};["object","boolean","number","function","string","symbol"].forEach(((e,t)=>{Fn[e]=function(n){return typeof n===e||"a"+(t<1?"n ":" ")+e}}));const Un={};Fn.transitional=function(e,t,n){function r(e,t){return"[Axios v1.7.7] Transitional option '"+e+"'"+t+(n?". "+n:"")}return(n,a,o)=>{if(!1===e)throw new bt(r(a," has been removed"+(t?" in "+t:"")),bt.ERR_DEPRECATED);return t&&!Un[a]&&(Un[a]=!0,console.warn(r(a," has been deprecated since v"+t+" and will be removed in the near future"))),!e||e(n,a,o)}};const zn={assertOptions:function(e,t,n){if("object"!==typeof e)throw new bt("options must be an object",bt.ERR_BAD_OPTION_VALUE);const r=Object.keys(e);let a=r.length;for(;a-- >0;){const o=r[a],l=t[o];if(l){const t=e[o],n=void 0===t||l(t,o,e);if(!0!==n)throw new bt("option "+o+" must be "+n,bt.ERR_BAD_OPTION_VALUE)}else if(!0!==n)throw new bt("Unknown option "+o,bt.ERR_BAD_OPTION)}},validators:Fn},In=zn.validators;class Mn{constructor(e){this.defaults=e,this.interceptors={request:new Tt,response:new Tt}}async request(e,t){try{return await this._request(e,t)}catch(n){if(n instanceof Error){let e;Error.captureStackTrace?Error.captureStackTrace(e={}):e=new Error;const t=e.stack?e.stack.replace(/^.+\n/,""):"";try{n.stack?t&&!String(n.stack).endsWith(t.replace(/^.+\n.+\n/,""))&&(n.stack+="\n"+t):n.stack=t}catch(il){}}throw n}}_request(e,t){"string"===typeof e?(t=t||{}).url=e:t=e||{},t=dn(this.defaults,t);const{transitional:n,paramsSerializer:r,headers:a}=t;void 0!==n&&zn.assertOptions(n,{silentJSONParsing:In.transitional(In.boolean),forcedJSONParsing:In.transitional(In.boolean),clarifyTimeoutError:In.transitional(In.boolean)},!1),null!=r&&(mt.isFunction(r)?t.paramsSerializer={serialize:r}:zn.assertOptions(r,{encode:In.function,serialize:In.function},!0)),t.method=(t.method||this.defaults.method||"get").toLowerCase();let o=a&&mt.merge(a.common,a[t.method]);a&&mt.forEach(["delete","get","head","post","put","patch","common"],(e=>{delete a[e]})),t.headers=Xt.concat(o,a);const l=[];let i=!0;this.interceptors.request.forEach((function(e){"function"===typeof e.runWhen&&!1===e.runWhen(t)||(i=i&&e.synchronous,l.unshift(e.fulfilled,e.rejected))}));const s=[];let u;this.interceptors.response.forEach((function(e){s.push(e.fulfilled,e.rejected)}));let c,d=0;if(!i){const e=[An.bind(this),void 0];for(e.unshift.apply(e,l),e.push.apply(e,s),c=e.length,u=Promise.resolve(t);d<c;)u=u.then(e[d++],e[d++]);return u}c=l.length;let f=t;for(d=0;d<c;){const e=l[d++],t=l[d++];try{f=e(f)}catch(p){t.call(this,p);break}}try{u=An.call(this,f)}catch(p){return Promise.reject(p)}for(d=0,c=s.length;d<c;)u=u.then(s[d++],s[d++]);return u}getUri(e){return _t(un((e=dn(this.defaults,e)).baseURL,e.url),e.params,e.paramsSerializer)}}mt.forEach(["delete","get","head","options"],(function(e){Mn.prototype[e]=function(t,n){return this.request(dn(n||{},{method:e,url:t,data:(n||{}).data}))}})),mt.forEach(["post","put","patch"],(function(e){function t(t){return function(n,r,a){return this.request(dn(a||{},{method:e,headers:t?{"Content-Type":"multipart/form-data"}:{},url:n,data:r}))}}Mn.prototype[e]=t(),Mn.prototype[e+"Form"]=t(!0)}));const Bn=Mn;class $n{constructor(e){if("function"!==typeof e)throw new TypeError("executor must be a function.");let t;this.promise=new Promise((function(e){t=e}));const n=this;this.promise.then((e=>{if(!n._listeners)return;let t=n._listeners.length;for(;t-- >0;)n._listeners[t](e);n._listeners=null})),this.promise.then=e=>{let t;const r=new Promise((e=>{n.subscribe(e),t=e})).then(e);return r.cancel=function(){n.unsubscribe(t)},r},e((function(e,r,a){n.reason||(n.reason=new Zt(e,r,a),t(n.reason))}))}throwIfRequested(){if(this.reason)throw this.reason}subscribe(e){this.reason?e(this.reason):this._listeners?this._listeners.push(e):this._listeners=[e]}unsubscribe(e){if(!this._listeners)return;const t=this._listeners.indexOf(e);-1!==t&&this._listeners.splice(t,1)}toAbortSignal(){const e=new AbortController,t=t=>{e.abort(t)};return this.subscribe(t),e.signal.unsubscribe=()=>this.unsubscribe(t),e.signal}static source(){let e;return{token:new $n((function(t){e=t})),cancel:e}}}const Wn=$n;const Hn={Continue:100,SwitchingProtocols:101,Processing:102,EarlyHints:103,Ok:200,Created:201,Accepted:202,NonAuthoritativeInformation:203,NoContent:204,ResetContent:205,PartialContent:206,MultiStatus:207,AlreadyReported:208,ImUsed:226,MultipleChoices:300,MovedPermanently:301,Found:302,SeeOther:303,NotModified:304,UseProxy:305,Unused:306,TemporaryRedirect:307,PermanentRedirect:308,BadRequest:400,Unauthorized:401,PaymentRequired:402,Forbidden:403,NotFound:404,MethodNotAllowed:405,NotAcceptable:406,ProxyAuthenticationRequired:407,RequestTimeout:408,Conflict:409,Gone:410,LengthRequired:411,PreconditionFailed:412,PayloadTooLarge:413,UriTooLong:414,UnsupportedMediaType:415,RangeNotSatisfiable:416,ExpectationFailed:417,ImATeapot:418,MisdirectedRequest:421,UnprocessableEntity:422,Locked:423,FailedDependency:424,TooEarly:425,UpgradeRequired:426,PreconditionRequired:428,TooManyRequests:429,RequestHeaderFieldsTooLarge:431,UnavailableForLegalReasons:451,InternalServerError:500,NotImplemented:501,BadGateway:502,ServiceUnavailable:503,GatewayTimeout:504,HttpVersionNotSupported:505,VariantAlsoNegotiates:506,InsufficientStorage:507,LoopDetected:508,NotExtended:510,NetworkAuthenticationRequired:511};Object.entries(Hn).forEach((e=>{let[t,n]=e;Hn[n]=t}));const Vn=Hn;const qn=function e(t){const n=new Bn(t),r=je(Bn.prototype.request,n);return mt.extend(r,Bn.prototype,n,{allOwnKeys:!0}),mt.extend(r,n,null,{allOwnKeys:!0}),r.create=function(n){return e(dn(t,n))},r}($t);qn.Axios=Bn,qn.CanceledError=Zt,qn.CancelToken=Wn,qn.isCancel=Jt,qn.VERSION=Dn,qn.toFormData=kt,qn.AxiosError=bt,qn.Cancel=qn.CanceledError,qn.all=function(e){return Promise.all(e)},qn.spread=function(e){return function(t){return e.apply(null,t)}},qn.isAxiosError=function(e){return mt.isObject(e)&&!0===e.isAxiosError},qn.mergeConfig=dn,qn.AxiosHeaders=Xt,qn.formToJSON=e=>Mt(mt.isHTMLForm(e)?new FormData(e):e),qn.getAdapter=Pn,qn.HttpStatusCode=Vn,qn.default=qn;const Kn=qn;var Qn=n(450);const Xn={apiBaseUrl:{NODE_ENV:"production",PUBLIC_URL:"",WDS_SOCKET_HOST:void 0,WDS_SOCKET_PATH:void 0,WDS_SOCKET_PORT:void 0,FAST_REFRESH:!0}.REACT_APP_API_BASE_URL||"http://127.0.0.1:5001"};var Gn=n(579);const Jn=()=>{const[e,t]=(0,r.useState)(null),[n,a]=(0,r.useState)(""),[o,l]=(0,r.useState)(""),[i,s]=(0,r.useState)(""),[u,c]=(0,r.useState)(""),[d,f]=(0,r.useState)(""),[p,h]=(0,r.useState)(""),[m,v]=(0,r.useState)(""),g=e=>{(0,Qn.saveAs)(e,"output.docx")};return(0,Gn.jsx)("div",{className:"container py-5 d-flex justify-content-center",children:(0,Gn.jsxs)("div",{className:"text-center",children:[(0,Gn.jsx)("h2",{className:"mb-4",children:"Extract data from ORCA files to Word documents"}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Upload your ORCA data file"}),(0,Gn.jsxs)("div",{className:"input-group",children:[(0,Gn.jsx)("input",{type:"file",className:"form-control",onChange:e=>{t(e.target.files[0])},accept:".txt"}),(0,Gn.jsx)("button",{className:"btn btn-primary",onClick:()=>{if(!e)return void console.error("No file selected");const t=new FormData;t.append("file",e),Kn.post(`${Xn.apiBaseUrl}/upload`,t).then((e=>{console.log("File uploaded successfully:",e),a(e.data.file_path)})).catch((e=>{console.error("Error uploading file:",e)}))},children:"Upload"})]})]}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Enter the terms you wish to search for (txt only):"}),(0,Gn.jsx)("input",{type:"text",className:"form-control",placeholder:"E.g., CARTESIAN COORDINATES",value:o,onChange:e=>l(e.target.value.toUpperCase())})]}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Enter how you want the lines specified:"}),(0,Gn.jsx)("input",{type:"text",className:"form-control",placeholder:"E.g., WHOLE, FIRST X, LAST X",value:i,onChange:e=>s(e.target.value.toUpperCase())})]}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Number of sections?"}),(0,Gn.jsx)("input",{type:"text",className:"form-control",placeholder:"Input as number...",value:u,onChange:e=>c(e.target.value)})]}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Use total lines?"}),(0,Gn.jsx)("input",{type:"text",className:"form-control",placeholder:"TRUE/FALSE",value:d,onChange:e=>f(e.target.value.toUpperCase())})]}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Total number of lines for output doc?"}),(0,Gn.jsx)("input",{type:"text",className:"form-control",placeholder:"Input as number...",value:p,onChange:e=>{const t=e.target.value;h(""===t?"":parseInt(t))}})]}),(0,Gn.jsx)("button",{className:"btn btn-primary",onClick:()=>{if(!e)return void alert("Please select a file.");const t={file_path:n.toString(),search_terms:o.split(","),sections:u.split(","),specify_lines:i.toString()};d&&(t.use_total_lines=d),p&&(t.total_lines=p),Kn.post(`${Xn.apiBaseUrl}/preview`,t).then((e=>{const t=e.data.document_content;v(t)})).catch((e=>{console.error("Error:",e)}))},children:"Preview Output"}),(0,Gn.jsx)("div",{className:"buttonSpacing",children:(0,Gn.jsx)("button",{className:"btn btn-primary",onClick:()=>{if(!e)return void alert("Please select a file.");const t={file_path:n.toString(),search_terms:o.split(","),sections:u.split(","),specify_lines:i.toString()};d&&(t.use_total_lines=d),p&&(t.total_lines=p),Kn.post(`${Xn.apiBaseUrl}/find-sections`,t,{responseType:"blob"}).then((e=>{const t=new Blob([e.data]);g(t)})).catch((e=>{console.error("Error:",e)}))},children:"Download Output"})}),m&&(0,Gn.jsxs)("div",{className:"document-preview",children:[(0,Gn.jsx)("h3",{children:"Document Preview"}),(0,Gn.jsx)("div",{className:"preview-box",children:(0,Gn.jsx)("pre",{children:m})})]})]})})},Yn=()=>{const[e,t]=(0,r.useState)(null),[n,a]=(0,r.useState)(""),[o,l]=(0,r.useState)([]),[i,s]=(0,r.useState)([]),[u,c]=(0,r.useState)([]),[d,f]=(0,r.useState)(!1),[p,h]=(0,r.useState)([]),m=0===p.length,v=0===o.length,g=0===i.length,y=0===u.length,[b,w]=(0,r.useState)(!1),[x,E]=(0,r.useState)(""),[S,k]=(0,r.useState)(!1),C=()=>{const e=i[0];return"WHOLE"===e.value||"SELECT"===e.value?e.value:`${e.value} ${e.lineNumber}`},N=e=>{(0,Qn.saveAs)(e,"output.docx")};return(0,Gn.jsx)("div",{className:"container py-5 d-flex justify-content-center",children:(0,Gn.jsxs)("div",{className:"text-center",children:[(0,Gn.jsx)("h2",{className:"mb-4",children:"Extract data from ORCA files to Word documents"}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Upload your ORCA data file"}),(0,Gn.jsxs)("div",{className:"input-group",children:[(0,Gn.jsx)("input",{type:"file",className:"form-control",onChange:e=>{const n=e.target.files[0];"text/plain"===n.type?t(n):alert("Invalid file type. Please upload a .txt file.")},accept:".txt",value:"","aria-label":"Upload ORCA data file"}),(0,Gn.jsx)("button",{className:"btn btn-primary",onClick:()=>{if(!e)return void console.error("No file selected");const t=new FormData;t.append("file",e),Kn.post(`${Xn.apiBaseUrl}/upload`,t).then((e=>{const t=e.data.file_name;a(e.data.file_path),h((e=>[...e,t]))})).catch((e=>{console.error("Error uploading file:",e)}))},children:"Upload"})]})]}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Uploaded Files:"}),p.map(((e,t)=>(0,Gn.jsxs)("span",{className:"badge bg-secondary me-2 mb-2",children:[e,(0,Gn.jsx)("button",{type:"button",className:"btn-close ms-1","aria-label":"Remove",onClick:()=>(e=>{h((t=>t.filter((t=>t!==e))))})(e)})]},t)))]}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Enter the terms you wish to search for (txt only):"}),(0,Gn.jsxs)("div",{children:[(0,Gn.jsx)("input",{type:"text",className:"form-control",placeholder:"E.g., CARTESIAN COORDINATES",onKeyPress:e=>((e,t)=>{if("Enter"===e.key||","===e.key||"Tab"===e.key){e.preventDefault();const n=e.target.value.trim();if(n){const r=n.split(",");t((e=>[...e,...r.map((e=>e.trim().toUpperCase()))])),e.target.value=""}}})(e,l),onBlur:e=>((e,t)=>{const n=e.target.value.trim();if(n){const r=n.split(",");t((e=>[...e,...r.map((e=>e.trim().toUpperCase()))])),e.target.value=""}})(e,l)}),o.map(((e,t)=>(0,Gn.jsxs)("span",{className:"badge bg-secondary me-2 mb-2",onClick:()=>((e,t)=>{t((t=>{const n=[...t];return n.splice(e,1),n}))})(t,l),children:[e,(0,Gn.jsx)("button",{type:"button",className:"btn-close ms-1","aria-label":"Remove"})]},t)))]}),o.length>1&&(0,Gn.jsxs)("div",{className:"form-check mt-2",children:[(0,Gn.jsx)("input",{className:"form-check-input",type:"checkbox",id:"sameCriteriaCheckbox",checked:b,onChange:e=>{w(e.target.checked)}}),(0,Gn.jsx)("label",{className:"form-check-label",htmlFor:"sameCriteriaCheckbox",children:"Is the search criteria same for all search terms"})]})]}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Enter how you want the lines specified:"}),(()=>{const e=i[0]||{value:"",showInput:!1};return(0,Gn.jsxs)("div",{className:"mb-2 d-flex align-items-center",children:[(0,Gn.jsxs)("select",{className:"form-select me-2",value:e.value,onChange:e=>{return t=e.target.value,void s([{value:t,showInput:"FIRST"===t||"LAST"===t}]);var t},children:[(0,Gn.jsx)("option",{value:"SELECT",children:"SELECT"}),(0,Gn.jsx)("option",{value:"WHOLE",children:"WHOLE"}),(0,Gn.jsx)("option",{value:"FIRST",children:"FIRST"}),(0,Gn.jsx)("option",{value:"LAST",children:"LAST"})]}),e.showInput&&(0,Gn.jsx)("input",{type:"text",className:"form-control",placeholder:"Enter line number",value:e.lineNumber||"",onChange:t=>{const n=[{...e,lineNumber:t.target.value}];s(n)}})]})})()]}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Number of sections?"}),(0,Gn.jsx)("input",{type:"text",className:"form-control",placeholder:"ex: 1-5 or 1,2,5",value:u.join(", "),onChange:e=>c(e.target.value.split(",").map((e=>e.trim())))})]}),(0,Gn.jsx)("div",{className:"button-container",children:(0,Gn.jsx)("button",{className:"btn btn-primary",onClick:()=>(f(!1),void(e?f(!0):alert("Please select a file."))),disabled:!(!m&&o.length&&i.length&&u.length)||m||v||g||y,children:"Submit Search Query"})}),!m&&!v&&!g&&!y&&d&&(0,Gn.jsx)("div",{className:"card mt-3",children:(0,Gn.jsxs)("div",{className:"card-body",children:[(0,Gn.jsx)("h5",{className:"card-title",children:"Search Query"}),(0,Gn.jsxs)("p",{className:"card-text",children:["Search Terms: ",o.join(", ")]}),(0,Gn.jsxs)("p",{className:"card-text",children:["Specify Lines: ","SELECT"!==i[0].value&&i[0].value,i[0].lineNumber?`, ${i[0].lineNumber}`:""]}),(0,Gn.jsxs)("p",{className:"card-text",children:["Sections: ",u.join(", ")]}),(0,Gn.jsxs)("div",{className:"d-flex justify-content-end",children:[(0,Gn.jsx)("button",{className:"btn btn-primary me-2",children:"Edit"}),(0,Gn.jsx)("button",{className:"btn btn-danger",onClick:()=>{l([]),s([]),c([]),f(!1),w(!1)},children:"Delete"})]})]})}),(0,Gn.jsx)("div",{className:"button-container",children:(0,Gn.jsx)("button",{className:"btn btn-primary",onClick:()=>{if(!e)return void alert("Please select a file.");const t={file_path:n.toString(),search_terms:o,sections:u,specify_lines:C()};Kn.post(`${Xn.apiBaseUrl}/preview`,t).then((e=>{E(e.data.document_content),k(!0)})).catch((e=>{console.error("Error fetching preview:",e)}))},disabled:!o.length||!i.length||!u.length||!e||m,children:"Preview"})}),S&&(0,Gn.jsx)("div",{className:"modal",style:{display:"block",backgroundColor:"rgba(0, 0, 0, 0.5)"},children:(0,Gn.jsx)("div",{className:"modal-dialog",style:{maxWidth:"80vw"},children:(0,Gn.jsxs)("div",{className:"modal-content",children:[(0,Gn.jsxs)("div",{className:"modal-header",children:[(0,Gn.jsx)("h5",{className:"modal-title",children:"Document Preview"}),(0,Gn.jsx)("button",{type:"button",className:"btn-close","aria-label":"Close",onClick:()=>k(!1)})]}),(0,Gn.jsx)("div",{className:"modal-body",children:(0,Gn.jsx)("pre",{children:x})}),(0,Gn.jsx)("div",{className:"modal-footer",children:(0,Gn.jsx)("button",{className:"btn btn-secondary",onClick:()=>k(!1),children:"Close"})})]})})}),(0,Gn.jsx)("div",{className:"button-container",children:(0,Gn.jsx)("button",{className:"btn btn-primary",onClick:()=>{if(!e)return void alert("Please select a file.");const t={file_path:n.toString(),search_terms:o,sections:u,specify_lines:C()};Kn.post(`${Xn.apiBaseUrl}/find-sections`,t,{responseType:"blob"}).then((e=>{const t=new Blob([e.data]);N(t)})).catch((e=>{console.error("Error:",e)}))},disabled:!o.length||!i.length||!u.length||!e||m,children:"Download Output"})})]})})},Zn=()=>{const[e,t]=(0,r.useState)(null),[n,a]=(0,r.useState)(""),[o,l]=(0,r.useState)(""),[i,s]=(0,r.useState)(""),[u,c]=(0,r.useState)(""),[d,f]=(0,r.useState)(""),[p,h]=(0,r.useState)(""),[m,v]=(0,r.useState)(""),g=e=>{(0,Qn.saveAs)(e,"output.docx")};return(0,Gn.jsx)("div",{className:"container py-5 d-flex justify-content-center",children:(0,Gn.jsxs)("div",{className:"text-center",children:[(0,Gn.jsx)("h2",{className:"mb-4",children:"Extract data from Gaussian files to Word documents"}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Upload your Gaussian data file"}),(0,Gn.jsxs)("div",{className:"input-group",children:[(0,Gn.jsx)("input",{type:"file",className:"form-control",onChange:e=>{t(e.target.files[0])},accept:".log"}),(0,Gn.jsx)("button",{className:"btn btn-primary",onClick:()=>{if(!e)return void console.error("No file selected");const t=new FormData;t.append("file",e),Kn.post(`${Xn.apiBaseUrl}/upload`,t).then((e=>{console.log("File uploaded successfully:",e),a(e.data.filename)})).catch((e=>{console.error("Error uploading file:",e)}))},children:"Upload"})]})]}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Enter the terms you wish to search for (txt only):"}),(0,Gn.jsx)("input",{type:"text",className:"form-control",placeholder:"E.g., CARTESIAN COORDINATES",value:o,onChange:e=>l(e.target.value.toUpperCase())})]}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Enter how you want the lines specified:"}),(0,Gn.jsx)("input",{type:"text",className:"form-control",placeholder:"E.g., WHOLE, FIRST X, LAST X",value:i,onChange:e=>s(e.target.value.toUpperCase())})]}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Number of sections?"}),(0,Gn.jsx)("input",{type:"text",className:"form-control",placeholder:"Input as number...",value:u,onChange:e=>c(e.target.value)})]}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Use total lines?"}),(0,Gn.jsx)("input",{type:"text",className:"form-control",placeholder:"TRUE/FALSE",value:d,onChange:e=>f(e.target.value.toUpperCase())})]}),(0,Gn.jsxs)("div",{className:"mb-3 text-start",children:[(0,Gn.jsx)("span",{children:"Total number of lines for output doc?"}),(0,Gn.jsx)("input",{type:"text",className:"form-control",placeholder:"Input as number...",value:p,onChange:e=>{const t=e.target.value;h(""===t?"":parseInt(t))}})]}),(0,Gn.jsx)("button",{className:"btn btn-primary",onClick:()=>{if(!e)return void alert("Please select a file.");const t={file_path:n.toString(),search_terms:o.split(","),sections:u.split(","),specify_lines:i.toString()};d&&(t.use_total_lines=d),p&&(t.total_lines=p),Kn.post(`${Xn.apiBaseUrl}/preview`,t).then((e=>{const t=e.data.document_content;v(t)})).catch((e=>{console.error("Error:",e)}))},children:"Preview Output"}),(0,Gn.jsx)("div",{className:"buttonSpacing",children:(0,Gn.jsx)("button",{className:"btn btn-primary",onClick:()=>{if(!e)return void alert("Please select a file.");const t={file_path:n.toString(),search_terms:o.split(","),sections:u.split(","),specify_lines:i.toString()};d&&(t.use_total_lines=d),p&&(t.total_lines=p),Kn.post(`${Xn.apiBaseUrl}/find-sections`,t,{responseType:"blob"}).then((e=>{const t=new Blob([e.data]);g(t)})).catch((e=>{console.error("Error:",e)}))},children:"Download Output"})}),m&&(0,Gn.jsxs)("div",{className:"document-preview",children:[(0,Gn.jsx)("h3",{children:"Document Preview"}),(0,Gn.jsx)("div",{className:"preview-box",children:(0,Gn.jsx)("pre",{children:m})})]})]})})};var er=n(139),tr=n.n(er);const nr=["xxl","xl","lg","md","sm","xs"],rr="xs",ar=r.createContext({prefixes:{},breakpoints:nr,minBreakpoint:rr}),{Consumer:or,Provider:lr}=ar;function ir(e,t){const{prefixes:n}=(0,r.useContext)(ar);return e||n[t]||t}const sr=r.forwardRef(((e,t)=>{let{bsPrefix:n,fluid:r=!1,as:a="div",className:o,...l}=e;const i=ir(n,"container"),s="string"===typeof r?`-${r}`:"-fluid";return(0,Gn.jsx)(a,{ref:t,...l,className:tr()(o,r?`${i}${s}`:i)})}));sr.displayName="Container";const ur=sr;n(197);function cr(){return cr=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)({}).hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},cr.apply(null,arguments)}function dr(e,t){if(null==e)return{};var n={};for(var r in e)if({}.hasOwnProperty.call(e,r)){if(t.includes(r))continue;n[r]=e[r]}return n}n(740);function fr(e){return"default"+e.charAt(0).toUpperCase()+e.substr(1)}function pr(e){var t=function(e,t){if("object"!==typeof e||null===e)return e;var n=e[Symbol.toPrimitive];if(void 0!==n){var r=n.call(e,t||"default");if("object"!==typeof r)return r;throw new TypeError("@@toPrimitive must return a primitive value.")}return("string"===t?String:Number)(e)}(e,"string");return"symbol"===typeof t?t:String(t)}function hr(e,t){return Object.keys(t).reduce((function(n,a){var o,l=n,i=l[fr(a)],s=l[a],u=dr(l,[fr(a),a].map(pr)),c=t[a],d=function(e,t,n){var a=(0,r.useRef)(void 0!==e),o=(0,r.useState)(t),l=o[0],i=o[1],s=void 0!==e,u=a.current;return a.current=s,!s&&u&&l!==t&&i(t),[s?e:l,(0,r.useCallback)((function(e){for(var t=arguments.length,r=new Array(t>1?t-1:0),a=1;a<t;a++)r[a-1]=arguments[a];n&&n.apply(void 0,[e].concat(r)),i(e)}),[n])]}(s,i,e[c]),f=d[0],p=d[1];return cr({},u,((o={})[a]=f,o[c]=p,o))}),e)}function mr(){var e=this.constructor.getDerivedStateFromProps(this.props,this.state);null!==e&&void 0!==e&&this.setState(e)}function vr(e){this.setState(function(t){var n=this.constructor.getDerivedStateFromProps(e,t);return null!==n&&void 0!==n?n:null}.bind(this))}function gr(e,t){try{var n=this.props,r=this.state;this.props=e,this.state=t,this.__reactInternalSnapshotFlag=!0,this.__reactInternalSnapshot=this.getSnapshotBeforeUpdate(n,r)}finally{this.props=n,this.state=r}}mr.__suppressDeprecationWarning=!0,vr.__suppressDeprecationWarning=!0,gr.__suppressDeprecationWarning=!0;var yr=Function.prototype.bind.call(Function.prototype.call,[].slice);function br(e,t){return yr(e.querySelectorAll(t))}const wr=e=>e&&"function"!==typeof e?t=>{e.current=t}:e;const xr=function(e,t){return(0,r.useMemo)((()=>function(e,t){const n=wr(e),r=wr(t);return e=>{n&&n(e),r&&r(e)}}(e,t)),[e,t])},Er=r.createContext(null);Er.displayName="NavContext";const Sr=Er,kr=function(e){let t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:null;return null!=e?String(e):t||null},Cr=r.createContext(null),Nr=r.createContext(null);function Or(e){return`data-rr-ui-${e}`}const Rr=function(e){const t=(0,r.useRef)(e);return(0,r.useEffect)((()=>{t.current=e}),[e]),t};function jr(e){const t=Rr(e);return(0,r.useCallback)((function(){return t.current&&t.current(...arguments)}),[t])}const _r=["as","disabled"];function Tr(e){let{tagName:t,disabled:n,href:r,target:a,rel:o,role:l,onClick:i,tabIndex:s=0,type:u}=e;t||(t=null!=r||null!=a||null!=o?"a":"button");const c={tagName:t};if("button"===t)return[{type:u||"button",disabled:n},c];const d=e=>{(n||"a"===t&&function(e){return!e||"#"===e.trim()}(r))&&e.preventDefault(),n?e.stopPropagation():null==i||i(e)};return"a"===t&&(r||(r="#"),n&&(r=void 0)),[{role:null!=l?l:"button",disabled:void 0,tabIndex:n?void 0:s,href:r,target:"a"===t?a:void 0,"aria-disabled":n||void 0,rel:"a"===t?o:void 0,onClick:d,onKeyDown:e=>{" "===e.key&&(e.preventDefault(),d(e))}},c]}const Pr=r.forwardRef(((e,t)=>{let{as:n,disabled:r}=e,a=function(e,t){if(null==e)return{};var n={};for(var r in e)if({}.hasOwnProperty.call(e,r)){if(t.indexOf(r)>=0)continue;n[r]=e[r]}return n}(e,_r);const[o,{tagName:l}]=Tr(Object.assign({tagName:n,disabled:r},a));return(0,Gn.jsx)(l,Object.assign({},a,o,{ref:t}))}));Pr.displayName="Button";const Lr=Pr,Ar=["as","active","eventKey"];function Dr(e){let{key:t,onClick:n,active:a,id:o,role:l,disabled:i}=e;const s=(0,r.useContext)(Cr),u=(0,r.useContext)(Sr),c=(0,r.useContext)(Nr);let d=a;const f={role:l};if(u){l||"tablist"!==u.role||(f.role="tab");const e=u.getControllerId(null!=t?t:null),n=u.getControlledId(null!=t?t:null);f[Or("event-key")]=t,f.id=e||o,d=null==a&&null!=t?u.activeKey===t:a,!d&&(null!=c&&c.unmountOnExit||null!=c&&c.mountOnEnter)||(f["aria-controls"]=n)}return"tab"===f.role&&(f["aria-selected"]=d,d||(f.tabIndex=-1),i&&(f.tabIndex=-1,f["aria-disabled"]=!0)),f.onClick=jr((e=>{i||(null==n||n(e),null!=t&&s&&!e.isPropagationStopped()&&s(t,e))})),[f,{isActive:d}]}const Fr=r.forwardRef(((e,t)=>{let{as:n=Lr,active:r,eventKey:a}=e,o=function(e,t){if(null==e)return{};var n={};for(var r in e)if({}.hasOwnProperty.call(e,r)){if(t.indexOf(r)>=0)continue;n[r]=e[r]}return n}(e,Ar);const[l,i]=Dr(Object.assign({key:kr(a,o.href),active:r},o));return l[Or("active")]=i.isActive,(0,Gn.jsx)(n,Object.assign({},o,l,{ref:t}))}));Fr.displayName="NavItem";const Ur=Fr,zr=["as","onSelect","activeKey","role","onKeyDown"];const Ir=()=>{},Mr=Or("event-key"),Br=r.forwardRef(((e,t)=>{let{as:n="div",onSelect:a,activeKey:o,role:l,onKeyDown:i}=e,s=function(e,t){if(null==e)return{};var n={};for(var r in e)if({}.hasOwnProperty.call(e,r)){if(t.indexOf(r)>=0)continue;n[r]=e[r]}return n}(e,zr);const u=function(){const[,e]=(0,r.useReducer)((e=>!e),!1);return e}(),c=(0,r.useRef)(!1),d=(0,r.useContext)(Cr),f=(0,r.useContext)(Nr);let p,h;f&&(l=l||"tablist",o=f.activeKey,p=f.getControlledId,h=f.getControllerId);const m=(0,r.useRef)(null),v=e=>{const t=m.current;if(!t)return null;const n=br(t,`[${Mr}]:not([aria-disabled=true])`),r=t.querySelector("[aria-selected=true]");if(!r||r!==document.activeElement)return null;const a=n.indexOf(r);if(-1===a)return null;let o=a+e;return o>=n.length&&(o=0),o<0&&(o=n.length-1),n[o]},g=(e,t)=>{null!=e&&(null==a||a(e,t),null==d||d(e,t))};(0,r.useEffect)((()=>{if(m.current&&c.current){const e=m.current.querySelector(`[${Mr}][aria-selected=true]`);null==e||e.focus()}c.current=!1}));const y=xr(t,m);return(0,Gn.jsx)(Cr.Provider,{value:g,children:(0,Gn.jsx)(Sr.Provider,{value:{role:l,activeKey:kr(o),getControlledId:p||Ir,getControllerId:h||Ir},children:(0,Gn.jsx)(n,Object.assign({},s,{onKeyDown:e=>{if(null==i||i(e),!f)return;let t;switch(e.key){case"ArrowLeft":case"ArrowUp":t=v(-1);break;case"ArrowRight":case"ArrowDown":t=v(1);break;default:return}var n;t&&(e.preventDefault(),g(t.dataset[(n="EventKey",`rrUi${n}`)]||null,e),c.current=!0,u())},ref:y,role:l}))})})}));Br.displayName="Nav";const $r=Object.assign(Br,{Item:Ur}),Wr=r.createContext(null);Wr.displayName="NavbarContext";const Hr=Wr,Vr=r.createContext(null);Vr.displayName="CardHeaderContext";const qr=Vr,Kr=r.forwardRef(((e,t)=>{let{className:n,bsPrefix:r,as:a="div",...o}=e;return r=ir(r,"nav-item"),(0,Gn.jsx)(a,{ref:t,className:tr()(n,r),...o})}));Kr.displayName="NavItem";const Qr=Kr;const Xr="undefined"!==typeof n.g&&n.g.navigator&&"ReactNative"===n.g.navigator.product,Gr="undefined"!==typeof document||Xr?r.useLayoutEffect:r.useEffect;new WeakMap;const Jr=["onKeyDown"];const Yr=r.forwardRef(((e,t)=>{let{onKeyDown:n}=e,r=function(e,t){if(null==e)return{};var n={};for(var r in e)if({}.hasOwnProperty.call(e,r)){if(t.indexOf(r)>=0)continue;n[r]=e[r]}return n}(e,Jr);const[a]=Tr(Object.assign({tagName:"a"},r)),o=jr((e=>{a.onKeyDown(e),null==n||n(e)}));return(l=r.href)&&"#"!==l.trim()&&"button"!==r.role?(0,Gn.jsx)("a",Object.assign({ref:t},r,{onKeyDown:n})):(0,Gn.jsx)("a",Object.assign({ref:t},r,a,{onKeyDown:o}));var l}));Yr.displayName="Anchor";const Zr=Yr,ea=r.forwardRef(((e,t)=>{let{bsPrefix:n,className:r,as:a=Zr,active:o,eventKey:l,disabled:i=!1,...s}=e;n=ir(n,"nav-link");const[u,c]=Dr({key:kr(l,s.href),active:o,disabled:i,...s});return(0,Gn.jsx)(a,{...s,...u,ref:t,disabled:i,className:tr()(r,n,i&&"disabled",c.isActive&&"active")})}));ea.displayName="NavLink";const ta=ea,na=r.forwardRef(((e,t)=>{const{as:n="div",bsPrefix:a,variant:o,fill:l=!1,justify:i=!1,navbar:s,navbarScroll:u,className:c,activeKey:d,...f}=hr(e,{activeKey:"onSelect"}),p=ir(a,"nav");let h,m,v=!1;const g=(0,r.useContext)(Hr),y=(0,r.useContext)(qr);return g?(h=g.bsPrefix,v=null==s||s):y&&({cardHeaderBsPrefix:m}=y),(0,Gn.jsx)($r,{as:n,ref:t,activeKey:d,className:tr()(c,{[p]:!v,[`${h}-nav`]:v,[`${h}-nav-scroll`]:v&&u,[`${m}-${o}`]:!!m,[`${p}-${o}`]:!!o,[`${p}-fill`]:l,[`${p}-justified`]:i}),...f})}));na.displayName="Nav";const ra=Object.assign(na,{Item:Qr,Link:ta}),aa=r.forwardRef(((e,t)=>{let{bsPrefix:n,className:r,as:a,...o}=e;n=ir(n,"navbar-brand");const l=a||(o.href?"a":"span");return(0,Gn.jsx)(l,{...o,ref:t,className:tr()(r,n)})}));aa.displayName="NavbarBrand";const oa=aa;function la(e){return e&&e.ownerDocument||document}function ia(e,t){return function(e){var t=la(e);return t&&t.defaultView||window}(e).getComputedStyle(e,t)}var sa=/([A-Z])/g;var ua=/^ms-/;function ca(e){return function(e){return e.replace(sa,"-$1").toLowerCase()}(e).replace(ua,"-ms-")}var da=/^((translate|rotate|scale)(X|Y|Z|3d)?|matrix(3d)?|perspective|skew(X|Y)?)$/i;const fa=function(e,t){var n="",r="";if("string"===typeof t)return e.style.getPropertyValue(ca(t))||ia(e).getPropertyValue(ca(t));Object.keys(t).forEach((function(a){var o=t[a];o||0===o?!function(e){return!(!e||!da.test(e))}(a)?n+=ca(a)+": "+o+";":r+=a+"("+o+") ":e.style.removeProperty(ca(a))})),r&&(n+="transform: "+r+";"),e.style.cssText+=";"+n};function pa(e,t){return pa=Object.setPrototypeOf?Object.setPrototypeOf.bind():function(e,t){return e.__proto__=t,e},pa(e,t)}const ha=!1,ma=r.createContext(null);var va="unmounted",ga="exited",ya="entering",ba="entered",wa="exiting",xa=function(e){var t,n;function a(t,n){var r;r=e.call(this,t,n)||this;var a,o=n&&!n.isMounting?t.enter:t.appear;return r.appearStatus=null,t.in?o?(a=ga,r.appearStatus=ya):a=ba:a=t.unmountOnExit||t.mountOnEnter?va:ga,r.state={status:a},r.nextCallback=null,r}n=e,(t=a).prototype=Object.create(n.prototype),t.prototype.constructor=t,pa(t,n),a.getDerivedStateFromProps=function(e,t){return e.in&&t.status===va?{status:ga}:null};var o=a.prototype;return o.componentDidMount=function(){this.updateStatus(!0,this.appearStatus)},o.componentDidUpdate=function(e){var t=null;if(e!==this.props){var n=this.state.status;this.props.in?n!==ya&&n!==ba&&(t=ya):n!==ya&&n!==ba||(t=wa)}this.updateStatus(!1,t)},o.componentWillUnmount=function(){this.cancelNextCallback()},o.getTimeouts=function(){var e,t,n,r=this.props.timeout;return e=t=n=r,null!=r&&"number"!==typeof r&&(e=r.exit,t=r.enter,n=void 0!==r.appear?r.appear:t),{exit:e,enter:t,appear:n}},o.updateStatus=function(e,t){if(void 0===e&&(e=!1),null!==t)if(this.cancelNextCallback(),t===ya){if(this.props.unmountOnExit||this.props.mountOnEnter){var n=this.props.nodeRef?this.props.nodeRef.current:l.findDOMNode(this);n&&function(e){e.scrollTop}(n)}this.performEnter(e)}else this.performExit();else this.props.unmountOnExit&&this.state.status===ga&&this.setState({status:va})},o.performEnter=function(e){var t=this,n=this.props.enter,r=this.context?this.context.isMounting:e,a=this.props.nodeRef?[r]:[l.findDOMNode(this),r],o=a[0],i=a[1],s=this.getTimeouts(),u=r?s.appear:s.enter;!e&&!n||ha?this.safeSetState({status:ba},(function(){t.props.onEntered(o)})):(this.props.onEnter(o,i),this.safeSetState({status:ya},(function(){t.props.onEntering(o,i),t.onTransitionEnd(u,(function(){t.safeSetState({status:ba},(function(){t.props.onEntered(o,i)}))}))})))},o.performExit=function(){var e=this,t=this.props.exit,n=this.getTimeouts(),r=this.props.nodeRef?void 0:l.findDOMNode(this);t&&!ha?(this.props.onExit(r),this.safeSetState({status:wa},(function(){e.props.onExiting(r),e.onTransitionEnd(n.exit,(function(){e.safeSetState({status:ga},(function(){e.props.onExited(r)}))}))}))):this.safeSetState({status:ga},(function(){e.props.onExited(r)}))},o.cancelNextCallback=function(){null!==this.nextCallback&&(this.nextCallback.cancel(),this.nextCallback=null)},o.safeSetState=function(e,t){t=this.setNextCallback(t),this.setState(e,t)},o.setNextCallback=function(e){var t=this,n=!0;return this.nextCallback=function(r){n&&(n=!1,t.nextCallback=null,e(r))},this.nextCallback.cancel=function(){n=!1},this.nextCallback},o.onTransitionEnd=function(e,t){this.setNextCallback(t);var n=this.props.nodeRef?this.props.nodeRef.current:l.findDOMNode(this),r=null==e&&!this.props.addEndListener;if(n&&!r){if(this.props.addEndListener){var a=this.props.nodeRef?[this.nextCallback]:[n,this.nextCallback],o=a[0],i=a[1];this.props.addEndListener(o,i)}null!=e&&setTimeout(this.nextCallback,e)}else setTimeout(this.nextCallback,0)},o.render=function(){var e=this.state.status;if(e===va)return null;var t=this.props,n=t.children,a=(t.in,t.mountOnEnter,t.unmountOnExit,t.appear,t.enter,t.exit,t.timeout,t.addEndListener,t.onEnter,t.onEntering,t.onEntered,t.onExit,t.onExiting,t.onExited,t.nodeRef,dr(t,["children","in","mountOnEnter","unmountOnExit","appear","enter","exit","timeout","addEndListener","onEnter","onEntering","onEntered","onExit","onExiting","onExited","nodeRef"]));return r.createElement(ma.Provider,{value:null},"function"===typeof n?n(e,a):r.cloneElement(r.Children.only(n),a))},a}(r.Component);function Ea(){}xa.contextType=ma,xa.propTypes={},xa.defaultProps={in:!1,mountOnEnter:!1,unmountOnExit:!1,appear:!1,enter:!0,exit:!0,onEnter:Ea,onEntering:Ea,onEntered:Ea,onExit:Ea,onExiting:Ea,onExited:Ea},xa.UNMOUNTED=va,xa.EXITED=ga,xa.ENTERING=ya,xa.ENTERED=ba,xa.EXITING=wa;const Sa=xa,ka=!("undefined"===typeof window||!window.document||!window.document.createElement);var Ca=!1,Na=!1;try{var Oa={get passive(){return Ca=!0},get once(){return Na=Ca=!0}};ka&&(window.addEventListener("test",Oa,Oa),window.removeEventListener("test",Oa,!0))}catch(il){}const Ra=function(e,t,n,r){if(r&&"boolean"!==typeof r&&!Na){var a=r.once,o=r.capture,l=n;!Na&&a&&(l=n.__once||function e(r){this.removeEventListener(t,e,o),n.call(this,r)},n.__once=l),e.addEventListener(t,l,Ca?r:o)}e.addEventListener(t,n,r)};const ja=function(e,t,n,r){var a=r&&"boolean"!==typeof r?r.capture:r;e.removeEventListener(t,n,a),n.__once&&e.removeEventListener(t,n.__once,a)};const _a=function(e,t,n,r){return Ra(e,t,n,r),function(){ja(e,t,n,r)}};function Ta(e,t,n){void 0===n&&(n=5);var r=!1,a=setTimeout((function(){r||function(e,t,n,r){if(void 0===n&&(n=!1),void 0===r&&(r=!0),e){var a=document.createEvent("HTMLEvents");a.initEvent(t,n,r),e.dispatchEvent(a)}}(e,"transitionend",!0)}),t+n),o=_a(e,"transitionend",(function(){r=!0}),{once:!0});return function(){clearTimeout(a),o()}}function Pa(e,t,n,r){null==n&&(n=function(e){var t=fa(e,"transitionDuration")||"",n=-1===t.indexOf("ms")?1e3:1;return parseFloat(t)*n}(e)||0);var a=Ta(e,n,r),o=_a(e,"transitionend",t);return function(){a(),o()}}function La(e,t){const n=fa(e,t)||"",r=-1===n.indexOf("ms")?1e3:1;return parseFloat(n)*r}function Aa(e,t){const n=La(e,"transitionDuration"),r=La(e,"transitionDelay"),a=Pa(e,(n=>{n.target===e&&(a(),t(n))}),n+r)}const Da=function(){for(var e=arguments.length,t=new Array(e),n=0;n<e;n++)t[n]=arguments[n];return t.filter((e=>null!=e)).reduce(((e,t)=>{if("function"!==typeof t)throw new Error("Invalid Argument Type, must only provide functions, undefined, or null.");return null===e?t:function(){for(var n=arguments.length,r=new Array(n),a=0;a<n;a++)r[a]=arguments[a];e.apply(this,r),t.apply(this,r)}}),null)};function Fa(e){e.offsetHeight}const Ua=r.forwardRef(((e,t)=>{let{onEnter:n,onEntering:a,onEntered:o,onExit:i,onExiting:s,onExited:u,addEndListener:c,children:d,childRef:f,...p}=e;const h=(0,r.useRef)(null),m=xr(h,f),v=e=>{var t;m((t=e)&&"setState"in t?l.findDOMNode(t):null!=t?t:null)},g=e=>t=>{e&&h.current&&e(h.current,t)},y=(0,r.useCallback)(g(n),[n]),b=(0,r.useCallback)(g(a),[a]),w=(0,r.useCallback)(g(o),[o]),x=(0,r.useCallback)(g(i),[i]),E=(0,r.useCallback)(g(s),[s]),S=(0,r.useCallback)(g(u),[u]),k=(0,r.useCallback)(g(c),[c]);return(0,Gn.jsx)(Sa,{ref:t,...p,onEnter:y,onEntered:w,onEntering:b,onExit:x,onExited:S,onExiting:E,addEndListener:k,nodeRef:h,children:"function"===typeof d?(e,t)=>d(e,{...t,ref:v}):r.cloneElement(d,{ref:v})})})),za={height:["marginTop","marginBottom"],width:["marginLeft","marginRight"]};function Ia(e,t){const n=t[`offset${e[0].toUpperCase()}${e.slice(1)}`],r=za[e];return n+parseInt(fa(t,r[0]),10)+parseInt(fa(t,r[1]),10)}const Ma={[ga]:"collapse",[wa]:"collapsing",[ya]:"collapsing",[ba]:"collapse show"},Ba=r.forwardRef(((e,t)=>{let{onEnter:n,onEntering:a,onEntered:o,onExit:l,onExiting:i,className:s,children:u,dimension:c="height",in:d=!1,timeout:f=300,mountOnEnter:p=!1,unmountOnExit:h=!1,appear:m=!1,getDimensionValue:v=Ia,...g}=e;const y="function"===typeof c?c():c,b=(0,r.useMemo)((()=>Da((e=>{e.style[y]="0"}),n)),[y,n]),w=(0,r.useMemo)((()=>Da((e=>{const t=`scroll${y[0].toUpperCase()}${y.slice(1)}`;e.style[y]=`${e[t]}px`}),a)),[y,a]),x=(0,r.useMemo)((()=>Da((e=>{e.style[y]=null}),o)),[y,o]),E=(0,r.useMemo)((()=>Da((e=>{e.style[y]=`${v(y,e)}px`,Fa(e)}),l)),[l,v,y]),S=(0,r.useMemo)((()=>Da((e=>{e.style[y]=null}),i)),[y,i]);return(0,Gn.jsx)(Ua,{ref:t,addEndListener:Aa,...g,"aria-expanded":g.role?d:null,onEnter:b,onEntering:w,onEntered:x,onExit:E,onExiting:S,childRef:u.ref,in:d,timeout:f,mountOnEnter:p,unmountOnExit:h,appear:m,children:(e,t)=>r.cloneElement(u,{...t,className:tr()(s,u.props.className,Ma[e],"width"===y&&"collapse-horizontal")})})})),$a=Ba,Wa=r.forwardRef(((e,t)=>{let{children:n,bsPrefix:a,...o}=e;a=ir(a,"navbar-collapse");const l=(0,r.useContext)(Hr);return(0,Gn.jsx)($a,{in:!(!l||!l.expanded),...o,children:(0,Gn.jsx)("div",{ref:t,className:a,children:n})})}));Wa.displayName="NavbarCollapse";const Ha=Wa,Va=r.forwardRef(((e,t)=>{let{bsPrefix:n,className:a,children:o,label:l="Toggle navigation",as:i="button",onClick:s,...u}=e;n=ir(n,"navbar-toggler");const{onToggle:c,expanded:d}=(0,r.useContext)(Hr)||{},f=jr((e=>{s&&s(e),c&&c()}));return"button"===i&&(u.type="button"),(0,Gn.jsx)(i,{...u,ref:t,onClick:f,"aria-label":l,className:tr()(a,n,!d&&"collapsed"),children:o||(0,Gn.jsx)("span",{className:`${n}-icon`})})}));Va.displayName="NavbarToggle";const qa=Va,Ka=new WeakMap,Qa=(e,t)=>{if(!e||!t)return;const n=Ka.get(t)||new Map;Ka.set(t,n);let r=n.get(e);return r||(r=t.matchMedia(e),r.refCount=0,n.set(r.media,r)),r};function Xa(e){let t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"undefined"===typeof window?void 0:window;const n=Qa(e,t),[a,o]=(0,r.useState)((()=>!!n&&n.matches));return Gr((()=>{let n=Qa(e,t);if(!n)return o(!1);let r=Ka.get(t);const a=()=>{o(n.matches)};return n.refCount++,n.addListener(a),a(),()=>{n.removeListener(a),n.refCount--,n.refCount<=0&&(null==r||r.delete(n.media)),n=void 0}}),[e]),a}const Ga=function(e){const t=Object.keys(e);function n(e,t){return e===t?t:e?`${e} and ${t}`:t}function a(n){const r=function(e){return t[Math.min(t.indexOf(e)+1,t.length-1)]}(n);let a=e[r];return a="number"===typeof a?a-.2+"px":`calc(${a} - 0.2px)`,`(max-width: ${a})`}return function(t,o,l){let i;return"object"===typeof t?(i=t,l=o,o=!0):(o=o||!0,i={[t]:o}),Xa((0,r.useMemo)((()=>Object.entries(i).reduce(((t,r)=>{let[o,l]=r;return"up"!==l&&!0!==l||(t=n(t,function(t){let n=e[t];return"number"===typeof n&&(n=`${n}px`),`(min-width: ${n})`}(o))),"down"!==l&&!0!==l||(t=n(t,a(o))),t}),"")),[JSON.stringify(i)]),l)}}({xs:0,sm:576,md:768,lg:992,xl:1200,xxl:1400});function Ja(e){void 0===e&&(e=la());try{var t=e.activeElement;return t&&t.nodeName?t:null}catch(il){return e.body}}function Ya(e,t){return e.contains?e.contains(t):e.compareDocumentPosition?e===t||!!(16&e.compareDocumentPosition(t)):void 0}function Za(e){const t=function(e){const t=(0,r.useRef)(e);return t.current=e,t}(e);(0,r.useEffect)((()=>()=>t.current()),[])}const eo=Or("modal-open");const to=class{constructor(){let{ownerDocument:e,handleContainerOverflow:t=!0,isRTL:n=!1}=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};this.handleContainerOverflow=t,this.isRTL=n,this.modals=[],this.ownerDocument=e}getScrollbarWidth(){return function(){let e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:document;const t=e.defaultView;return Math.abs(t.innerWidth-e.documentElement.clientWidth)}(this.ownerDocument)}getElement(){return(this.ownerDocument||document).body}setModalAttributes(e){}removeModalAttributes(e){}setContainerStyle(e){const t={overflow:"hidden"},n=this.isRTL?"paddingLeft":"paddingRight",r=this.getElement();e.style={overflow:r.style.overflow,[n]:r.style[n]},e.scrollBarWidth&&(t[n]=`${parseInt(fa(r,n)||"0",10)+e.scrollBarWidth}px`),r.setAttribute(eo,""),fa(r,t)}reset(){[...this.modals].forEach((e=>this.remove(e)))}removeContainerStyle(e){const t=this.getElement();t.removeAttribute(eo),Object.assign(t.style,e.style)}add(e){let t=this.modals.indexOf(e);return-1!==t?t:(t=this.modals.length,this.modals.push(e),this.setModalAttributes(e),0!==t||(this.state={scrollBarWidth:this.getScrollbarWidth(),style:{}},this.handleContainerOverflow&&this.setContainerStyle(this.state)),t)}remove(e){const t=this.modals.indexOf(e);-1!==t&&(this.modals.splice(t,1),!this.modals.length&&this.handleContainerOverflow&&this.removeContainerStyle(this.state),this.removeModalAttributes(e))}isTopModal(e){return!!this.modals.length&&this.modals[this.modals.length-1]===e}},no=(0,r.createContext)(ka?window:void 0);no.Provider;function ro(){return(0,r.useContext)(no)}const ao=(e,t)=>ka?null==e?(t||la()).body:("function"===typeof e&&(e=e()),e&&"current"in e&&(e=e.current),e&&("nodeType"in e||e.getBoundingClientRect)?e:null):null;const oo=function(e){let{children:t,in:n,onExited:a,mountOnEnter:o,unmountOnExit:l}=e;const i=(0,r.useRef)(null),s=(0,r.useRef)(n),u=jr(a);(0,r.useEffect)((()=>{n?s.current=!0:u(i.current)}),[n,u]);const c=xr(i,t.ref),d=(0,r.cloneElement)(t,{ref:c});return n?d:l||!s.current&&o?null:d};function lo(){const e=r.version.split(".");return{major:+e[0],minor:+e[1],patch:+e[2]}}const io=["onEnter","onEntering","onEntered","onExit","onExiting","onExited","addEndListener","children"];const so=["component"];const uo=r.forwardRef(((e,t)=>{let{component:n}=e;const a=function(e){let{onEnter:t,onEntering:n,onEntered:a,onExit:o,onExiting:l,onExited:i,addEndListener:s,children:u}=e,c=function(e,t){if(null==e)return{};var n={};for(var r in e)if({}.hasOwnProperty.call(e,r)){if(t.indexOf(r)>=0)continue;n[r]=e[r]}return n}(e,io);const{major:d}=lo(),f=d>=19?u.props.ref:u.ref,p=(0,r.useRef)(null),h=xr(p,"function"===typeof u?null:f),m=e=>t=>{e&&p.current&&e(p.current,t)},v=(0,r.useCallback)(m(t),[t]),g=(0,r.useCallback)(m(n),[n]),y=(0,r.useCallback)(m(a),[a]),b=(0,r.useCallback)(m(o),[o]),w=(0,r.useCallback)(m(l),[l]),x=(0,r.useCallback)(m(i),[i]),E=(0,r.useCallback)(m(s),[s]);return Object.assign({},c,{nodeRef:p},t&&{onEnter:v},n&&{onEntering:g},a&&{onEntered:y},o&&{onExit:b},l&&{onExiting:w},i&&{onExited:x},s&&{addEndListener:E},{children:"function"===typeof u?(e,t)=>u(e,Object.assign({},t,{ref:h})):(0,r.cloneElement)(u,{ref:h})})}(function(e,t){if(null==e)return{};var n={};for(var r in e)if({}.hasOwnProperty.call(e,r)){if(t.indexOf(r)>=0)continue;n[r]=e[r]}return n}(e,so));return(0,Gn.jsx)(n,Object.assign({ref:t},a))}));function co(e){let{children:t,in:n,onExited:a,onEntered:o,transition:l}=e;const[i,s]=(0,r.useState)(!n);n&&i&&s(!1);const u=function(e){let{in:t,onTransition:n}=e;const a=(0,r.useRef)(null),o=(0,r.useRef)(!0),l=jr(n);return Gr((()=>{if(!a.current)return;let e=!1;return l({in:t,element:a.current,initial:o.current,isStale:()=>e}),()=>{e=!0}}),[t,l]),Gr((()=>(o.current=!1,()=>{o.current=!0})),[]),a}({in:!!n,onTransition:e=>{Promise.resolve(l(e)).then((()=>{e.isStale()||(e.in?null==o||o(e.element,e.initial):(s(!0),null==a||a(e.element)))}),(t=>{throw e.in||s(!0),t}))}}),c=xr(u,t.ref);return i&&!n?null:(0,r.cloneElement)(t,{ref:c})}function fo(e,t,n){return e?(0,Gn.jsx)(uo,Object.assign({},n,{component:e})):t?(0,Gn.jsx)(co,Object.assign({},n,{transition:t})):(0,Gn.jsx)(oo,Object.assign({},n))}const po=["show","role","className","style","children","backdrop","keyboard","onBackdropClick","onEscapeKeyDown","transition","runTransition","backdropTransition","runBackdropTransition","autoFocus","enforceFocus","restoreFocus","restoreFocusOptions","renderDialog","renderBackdrop","manager","container","onShow","onHide","onExit","onExited","onExiting","onEnter","onEntering","onEntered"];let ho;function mo(e){const t=ro(),n=e||function(e){return ho||(ho=new to({ownerDocument:null==e?void 0:e.document})),ho}(t),a=(0,r.useRef)({dialog:null,backdrop:null});return Object.assign(a.current,{add:()=>n.add(a.current),remove:()=>n.remove(a.current),isTopModal:()=>n.isTopModal(a.current),setDialogRef:(0,r.useCallback)((e=>{a.current.dialog=e}),[]),setBackdropRef:(0,r.useCallback)((e=>{a.current.backdrop=e}),[])})}const vo=(0,r.forwardRef)(((e,t)=>{let{show:n=!1,role:a="dialog",className:o,style:i,children:s,backdrop:u=!0,keyboard:c=!0,onBackdropClick:d,onEscapeKeyDown:f,transition:p,runTransition:h,backdropTransition:m,runBackdropTransition:v,autoFocus:g=!0,enforceFocus:y=!0,restoreFocus:b=!0,restoreFocusOptions:w,renderDialog:x,renderBackdrop:E=e=>(0,Gn.jsx)("div",Object.assign({},e)),manager:S,container:k,onShow:C,onHide:N=()=>{},onExit:O,onExited:R,onExiting:j,onEnter:_,onEntering:T,onEntered:P}=e,L=function(e,t){if(null==e)return{};var n={};for(var r in e)if({}.hasOwnProperty.call(e,r)){if(t.indexOf(r)>=0)continue;n[r]=e[r]}return n}(e,po);const A=ro(),D=function(e,t){const n=ro(),[a,o]=(0,r.useState)((()=>ao(e,null==n?void 0:n.document)));if(!a){const t=ao(e);t&&o(t)}return(0,r.useEffect)((()=>{t&&a&&t(a)}),[t,a]),(0,r.useEffect)((()=>{const t=ao(e);t!==a&&o(t)}),[e,a]),a}(k),F=mo(S),U=function(){const e=(0,r.useRef)(!0),t=(0,r.useRef)((()=>e.current));return(0,r.useEffect)((()=>(e.current=!0,()=>{e.current=!1})),[]),t.current}(),z=function(e){const t=(0,r.useRef)(null);return(0,r.useEffect)((()=>{t.current=e})),t.current}(n),[I,M]=(0,r.useState)(!n),B=(0,r.useRef)(null);(0,r.useImperativeHandle)(t,(()=>F),[F]),ka&&!z&&n&&(B.current=Ja(null==A?void 0:A.document)),n&&I&&M(!1);const $=jr((()=>{if(F.add(),Q.current=_a(document,"keydown",q),K.current=_a(document,"focus",(()=>setTimeout(H)),!0),C&&C(),g){var e,t;const n=Ja(null!=(e=null==(t=F.dialog)?void 0:t.ownerDocument)?e:null==A?void 0:A.document);F.dialog&&n&&!Ya(F.dialog,n)&&(B.current=n,F.dialog.focus())}})),W=jr((()=>{var e;(F.remove(),null==Q.current||Q.current(),null==K.current||K.current(),b)&&(null==(e=B.current)||null==e.focus||e.focus(w),B.current=null)}));(0,r.useEffect)((()=>{n&&D&&$()}),[n,D,$]),(0,r.useEffect)((()=>{I&&W()}),[I,W]),Za((()=>{W()}));const H=jr((()=>{if(!y||!U()||!F.isTopModal())return;const e=Ja(null==A?void 0:A.document);F.dialog&&e&&!Ya(F.dialog,e)&&F.dialog.focus()})),V=jr((e=>{e.target===e.currentTarget&&(null==d||d(e),!0===u&&N())})),q=jr((e=>{c&&function(e){return"Escape"===e.code||27===e.keyCode}(e)&&F.isTopModal()&&(null==f||f(e),e.defaultPrevented||N())})),K=(0,r.useRef)(),Q=(0,r.useRef)();if(!D)return null;const X=Object.assign({role:a,ref:F.setDialogRef,"aria-modal":"dialog"===a||void 0},L,{style:i,className:o,tabIndex:-1});let G=x?x(X):(0,Gn.jsx)("div",Object.assign({},X,{children:r.cloneElement(s,{role:"document"})}));G=fo(p,h,{unmountOnExit:!0,mountOnEnter:!0,appear:!0,in:!!n,onExit:O,onExiting:j,onExited:function(){M(!0),null==R||R(...arguments)},onEnter:_,onEntering:T,onEntered:P,children:G});let J=null;return u&&(J=E({ref:F.setBackdropRef,onClick:V}),J=fo(m,v,{in:!!n,appear:!0,mountOnEnter:!0,unmountOnExit:!0,children:J})),(0,Gn.jsx)(Gn.Fragment,{children:l.createPortal((0,Gn.jsxs)(Gn.Fragment,{children:[J,G]}),D)})}));vo.displayName="Modal";const go=Object.assign(vo,{Manager:to}),yo={[ya]:"show",[ba]:"show"},bo=r.forwardRef(((e,t)=>{let{className:n,children:a,transitionClasses:o={},onEnter:l,...i}=e;const s={in:!1,timeout:300,mountOnEnter:!1,unmountOnExit:!1,appear:!1,...i},u=(0,r.useCallback)(((e,t)=>{Fa(e),null==l||l(e,t)}),[l]),{major:c}=lo(),d=c>=19?a.props.ref:a.ref;return(0,Gn.jsx)(Ua,{ref:t,addEndListener:Aa,...s,onEnter:u,childRef:d,children:(e,t)=>r.cloneElement(a,{...t,className:tr()("fade",n,a.props.className,yo[e],o[e])})})}));bo.displayName="Fade";const wo=bo,xo=r.forwardRef(((e,t)=>{let{className:n,bsPrefix:r,as:a="div",...o}=e;return r=ir(r,"offcanvas-body"),(0,Gn.jsx)(a,{ref:t,className:tr()(n,r),...o})}));xo.displayName="OffcanvasBody";const Eo=xo,So={[ya]:"show",[ba]:"show"},ko=r.forwardRef(((e,t)=>{let{bsPrefix:n,className:a,children:o,in:l=!1,mountOnEnter:i=!1,unmountOnExit:s=!1,appear:u=!1,...c}=e;return n=ir(n,"offcanvas"),(0,Gn.jsx)(Ua,{ref:t,addEndListener:Aa,in:l,mountOnEnter:i,unmountOnExit:s,appear:u,...c,childRef:o.ref,children:(e,t)=>r.cloneElement(o,{...t,className:tr()(a,o.props.className,(e===ya||e===wa)&&`${n}-toggling`,So[e])})})}));ko.displayName="OffcanvasToggling";const Co=ko,No=r.createContext({onHide(){}});var Oo=n(173),Ro=n.n(Oo);const jo={"aria-label":Ro().string,onClick:Ro().func,variant:Ro().oneOf(["white"])},_o=r.forwardRef(((e,t)=>{let{className:n,variant:r,"aria-label":a="Close",...o}=e;return(0,Gn.jsx)("button",{ref:t,type:"button",className:tr()("btn-close",r&&`btn-close-${r}`,n),"aria-label":a,...o})}));_o.displayName="CloseButton",_o.propTypes=jo;const To=_o,Po=r.forwardRef(((e,t)=>{let{closeLabel:n="Close",closeVariant:a,closeButton:o=!1,onHide:l,children:i,...s}=e;const u=(0,r.useContext)(No),c=jr((()=>{null==u||u.onHide(),null==l||l()}));return(0,Gn.jsxs)("div",{ref:t,...s,children:[i,o&&(0,Gn.jsx)(To,{"aria-label":n,variant:a,onClick:c})]})})),Lo=Po,Ao=r.forwardRef(((e,t)=>{let{bsPrefix:n,className:r,closeLabel:a="Close",closeButton:o=!1,...l}=e;return n=ir(n,"offcanvas-header"),(0,Gn.jsx)(Lo,{ref:t,...l,className:tr()(r,n),closeLabel:a,closeButton:o})}));Ao.displayName="OffcanvasHeader";const Do=Ao,Fo=(Uo="h5",r.forwardRef(((e,t)=>(0,Gn.jsx)("div",{...e,ref:t,className:tr()(e.className,Uo)}))));var Uo;const zo=r.forwardRef(((e,t)=>{let{className:n,bsPrefix:r,as:a=Fo,...o}=e;return r=ir(r,"offcanvas-title"),(0,Gn.jsx)(a,{ref:t,className:tr()(n,r),...o})}));zo.displayName="OffcanvasTitle";const Io=zo;function Mo(e,t){return e.replace(new RegExp("(^|\\s)"+t+"(?:\\s|$)","g"),"$1").replace(/\s+/g," ").replace(/^\s*|\s*$/g,"")}const Bo=".fixed-top, .fixed-bottom, .is-fixed, .sticky-top",$o=".sticky-top",Wo=".navbar-toggler";class Ho extends to{adjustAndStore(e,t,n){const r=t.style[e];t.dataset[e]=r,fa(t,{[e]:`${parseFloat(fa(t,e))+n}px`})}restore(e,t){const n=t.dataset[e];void 0!==n&&(delete t.dataset[e],fa(t,{[e]:n}))}setContainerStyle(e){super.setContainerStyle(e);const t=this.getElement();if(function(e,t){e.classList?e.classList.add(t):function(e,t){return e.classList?!!t&&e.classList.contains(t):-1!==(" "+(e.className.baseVal||e.className)+" ").indexOf(" "+t+" ")}(e,t)||("string"===typeof e.className?e.className=e.className+" "+t:e.setAttribute("class",(e.className&&e.className.baseVal||"")+" "+t))}(t,"modal-open"),!e.scrollBarWidth)return;const n=this.isRTL?"paddingLeft":"paddingRight",r=this.isRTL?"marginLeft":"marginRight";br(t,Bo).forEach((t=>this.adjustAndStore(n,t,e.scrollBarWidth))),br(t,$o).forEach((t=>this.adjustAndStore(r,t,-e.scrollBarWidth))),br(t,Wo).forEach((t=>this.adjustAndStore(r,t,e.scrollBarWidth)))}removeContainerStyle(e){super.removeContainerStyle(e);const t=this.getElement();!function(e,t){e.classList?e.classList.remove(t):"string"===typeof e.className?e.className=Mo(e.className,t):e.setAttribute("class",Mo(e.className&&e.className.baseVal||"",t))}(t,"modal-open");const n=this.isRTL?"paddingLeft":"paddingRight",r=this.isRTL?"marginLeft":"marginRight";br(t,Bo).forEach((e=>this.restore(n,e))),br(t,$o).forEach((e=>this.restore(r,e))),br(t,Wo).forEach((e=>this.restore(r,e)))}}let Vo;const qo=Ho;function Ko(e){return(0,Gn.jsx)(Co,{...e})}function Qo(e){return(0,Gn.jsx)(wo,{...e})}const Xo=r.forwardRef(((e,t)=>{let{bsPrefix:n,className:a,children:o,"aria-labelledby":l,placement:i="start",responsive:s,show:u=!1,backdrop:c=!0,keyboard:d=!0,scroll:f=!1,onEscapeKeyDown:p,onShow:h,onHide:m,container:v,autoFocus:g=!0,enforceFocus:y=!0,restoreFocus:b=!0,restoreFocusOptions:w,onEntered:x,onExit:E,onExiting:S,onEnter:k,onEntering:C,onExited:N,backdropClassName:O,manager:R,renderStaticNode:j=!1,..._}=e;const T=(0,r.useRef)();n=ir(n,"offcanvas");const[P,L]=(0,r.useState)(!1),A=jr(m),D=Ga(s||"xs","up");(0,r.useEffect)((()=>{L(s?u&&!D:u)}),[u,s,D]);const F=(0,r.useMemo)((()=>({onHide:A})),[A]);const U=(0,r.useCallback)((e=>(0,Gn.jsx)("div",{...e,className:tr()(`${n}-backdrop`,O)})),[O,n]),z=e=>(0,Gn.jsx)("div",{...e,..._,className:tr()(a,s?`${n}-${s}`:n,`${n}-${i}`),"aria-labelledby":l,children:o});return(0,Gn.jsxs)(Gn.Fragment,{children:[!P&&(s||j)&&z({}),(0,Gn.jsx)(No.Provider,{value:F,children:(0,Gn.jsx)(go,{show:P,ref:t,backdrop:c,container:v,keyboard:d,autoFocus:g,enforceFocus:y&&!f,restoreFocus:b,restoreFocusOptions:w,onEscapeKeyDown:p,onShow:h,onHide:A,onEnter:function(e){e&&(e.style.visibility="visible");for(var t=arguments.length,n=new Array(t>1?t-1:0),r=1;r<t;r++)n[r-1]=arguments[r];null==k||k(e,...n)},onEntering:C,onEntered:x,onExit:E,onExiting:S,onExited:function(e){e&&(e.style.visibility="");for(var t=arguments.length,n=new Array(t>1?t-1:0),r=1;r<t;r++)n[r-1]=arguments[r];null==N||N(...n)},manager:R||(f?(T.current||(T.current=new qo({handleContainerOverflow:!1})),T.current):function(e){return Vo||(Vo=new Ho(e)),Vo}()),transition:Ko,backdropTransition:Qo,renderBackdrop:U,renderDialog:z})})]})}));Xo.displayName="Offcanvas";const Go=Object.assign(Xo,{Body:Eo,Header:Do,Title:Io}),Jo=r.forwardRef(((e,t)=>{let{onHide:n,...a}=e;const o=(0,r.useContext)(Hr),l=jr((()=>{null==o||null==o.onToggle||o.onToggle(),null==n||n()}));return(0,Gn.jsx)(Go,{ref:t,show:!(null==o||!o.expanded),...a,renderStaticNode:!0,onHide:l})}));Jo.displayName="NavbarOffcanvas";const Yo=Jo,Zo=r.forwardRef(((e,t)=>{let{className:n,bsPrefix:r,as:a="span",...o}=e;return r=ir(r,"navbar-text"),(0,Gn.jsx)(a,{ref:t,className:tr()(n,r),...o})}));Zo.displayName="NavbarText";const el=Zo,tl=r.forwardRef(((e,t)=>{const{bsPrefix:n,expand:a=!0,variant:o="light",bg:l,fixed:i,sticky:s,className:u,as:c="nav",expanded:d,onToggle:f,onSelect:p,collapseOnSelect:h=!1,...m}=hr(e,{expanded:"onToggle"}),v=ir(n,"navbar"),g=(0,r.useCallback)((function(){null==p||p(...arguments),h&&d&&(null==f||f(!1))}),[p,h,d,f]);void 0===m.role&&"nav"!==c&&(m.role="navigation");let y=`${v}-expand`;"string"===typeof a&&(y=`${y}-${a}`);const b=(0,r.useMemo)((()=>({onToggle:()=>null==f?void 0:f(!d),bsPrefix:v,expanded:!!d,expand:a})),[v,d,a,f]);return(0,Gn.jsx)(Hr.Provider,{value:b,children:(0,Gn.jsx)(Cr.Provider,{value:g,children:(0,Gn.jsx)(c,{ref:t,...m,className:tr()(u,v,a&&y,o&&`${v}-${o}`,l&&`bg-${l}`,s&&`sticky-${s}`,i&&`fixed-${i}`)})})})}));tl.displayName="Navbar";const nl=Object.assign(tl,{Brand:oa,Collapse:Ha,Offcanvas:Yo,Text:el,Toggle:qa});const rl=function(){const e=Z();return(0,Gn.jsx)(nl,{collapseOnSelect:!0,expand:"lg",sticky:"top","data-bs-theme":"dark",className:"bg-body-tertiary",children:(0,Gn.jsxs)(ur,{children:[(0,Gn.jsxs)(nl.Brand,{as:Ne,to:"/",className:"logo-title",children:[(0,Gn.jsx)("img",{alt:"App Logo",src:"logo_temp.png",width:"50",height:"50",className:"d-inline-block align-center me-2"}),(0,Gn.jsx)("span",{children:"Log Data Extraction"})]}),(0,Gn.jsx)(nl.Toggle,{"aria-controls":"responsive-navbar-nav"}),(0,Gn.jsxs)(nl.Collapse,{id:"responsive-navbar-nav",children:[(0,Gn.jsx)(ra,{className:"me-auto"}),(0,Gn.jsxs)(ra,{activeKey:e.pathname,children:[(0,Gn.jsx)(ra.Link,{href:"/",children:"ORCA Log Extraction"}),(0,Gn.jsx)(ra.Link,{href:"/gaussian",children:"Gaussian Log Extraction"}),(0,Gn.jsx)(ra.Link,{href:"/about",children:"About"})]})]})]})})},al=()=>(0,Gn.jsx)(ur,{className:"d-flex justify-content-center mb-4",children:(0,Gn.jsxs)("div",{className:"text-left mb-3 min-vh-100",children:[(0,Gn.jsx)("h2",{className:"mb-4",children:"About Electronic Structure Parser"}),(0,Gn.jsx)("p",{children:"Our web application specializes in extracting specific details from chemistry log files. Users can effortlessly search for information by entering a search term, enabling efficient data retrieval. Additionally, our application offers the convenience of downloading the extracted data as a Word document, ensuring seamless accessibility and usability."}),(0,Gn.jsxs)("p",{children:["Access the application here:"," ",(0,Gn.jsx)("a",{href:"https://raven-intent-mentally.ngrok-free.app/",target:"_blank",rel:"noopener noreferrer",children:"Electronic Structure Parser"})]}),(0,Gn.jsx)("h3",{children:"Example Inputs and Outputs"}),(0,Gn.jsxs)("ul",{className:"examples-list",children:[(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"Input 1:"}),(0,Gn.jsxs)("ul",{children:[(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"Uploaded File:"})," ORCA log file (`.txt`)"]}),(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"Search Terms:"})," `CARTESIAN COORDINATES`"]}),(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"Line Specification:"})," `FIRST`"]}),(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"Sections:"})," `1`"]})]}),(0,Gn.jsx)("strong",{children:"Output:"})," Extracted data in `output.docx` containing the first occurrence of `CARTESIAN COORDINATES` in section 1."]}),(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"Input 2:"}),(0,Gn.jsxs)("ul",{children:[(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"Uploaded File:"})," ORCA log file with molecular energy data"]}),(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"Search Terms:"})," `TOTAL ENERGY, FINAL GRADIENT`"]}),(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"Line Specification:"})," `WHOLE`"]}),(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"Sections:"})," `1,2`"]})]}),(0,Gn.jsx)("strong",{children:"Output:"})," Extracted data in `output.docx` containing all lines with `TOTAL ENERGY` and `FINAL GRADIENT` from sections 1 and 2."]})]}),(0,Gn.jsx)("h3",{children:"Step-by-Step Guide"}),(0,Gn.jsxs)("ol",{className:"how-to-guide",children:[(0,Gn.jsx)("li",{children:"Upload a `.txt` file containing ORCA log data using the file upload option."}),(0,Gn.jsx)("li",{children:"Enter keywords in the search field (e.g., `CARTESIAN COORDINATES`). Press **Enter** or use commas to add multiple terms."}),(0,Gn.jsx)("li",{children:"Select how to specify lines (e.g., `FIRST`, `LAST`, `WHOLE`) and provide additional details like line numbers if required."}),(0,Gn.jsx)("li",{children:"Define sections to search within (e.g., `1-2` or `1,2`)."}),(0,Gn.jsx)("li",{children:"Submit the search query to review and validate your criteria."}),(0,Gn.jsx)("li",{children:"Preview the extracted data using the **Preview** button."}),(0,Gn.jsx)("li",{children:"Download the results as a Word document by clicking the **Download Output** button."})]}),(0,Gn.jsx)("h3",{children:"Key Features"}),(0,Gn.jsxs)("ul",{children:[(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"File Upload:"})," Accepts `.txt` files for ORCA data processing."]}),(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"Search Functionality:"})," Allows users to search for a single term in the uploaded file for precise data extraction. (Support for multiple search terms is planned for a future update.)"]}),(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"Section-Based Selection:"})," Filters data by user-defined sections for better accuracy."]}),(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"Preview Option:"})," Provides a preview of extracted data before downloading."]}),(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"Output Generation:"})," Creates a Word document (`output.docx`) with the extracted results."]}),(0,Gn.jsxs)("li",{children:[(0,Gn.jsx)("strong",{children:"User-Friendly Interface:"})," Interactive design with badges, dropdowns, and modals for seamless navigation."]})]}),(0,Gn.jsx)("h3",{children:"Tutorial Video"}),(0,Gn.jsxs)("div",{className:"tutorial-video",children:[(0,Gn.jsx)("p",{children:"Watch this video to learn how to use the application:"}),(0,Gn.jsxs)("video",{width:"560",height:"315",controls:!0,children:[(0,Gn.jsx)("source",{src:"/tutorialVideo.mov",type:"video/mp4"}),"Your browser does not support the video tag."]})]})]})}),ol=()=>(0,Gn.jsxs)(Se,{children:[(0,Gn.jsx)(rl,{}),(0,Gn.jsxs)(ge,{children:[(0,Gn.jsx)(me,{path:"/",element:(0,Gn.jsx)(Yn,{})}),(0,Gn.jsx)(me,{path:"/temp",element:(0,Gn.jsx)(Jn,{})}),(0,Gn.jsx)(me,{path:"/gaussian",element:(0,Gn.jsx)(Zn,{})}),(0,Gn.jsx)(me,{path:"/about",element:(0,Gn.jsx)(al,{})})]})]}),ll=e=>{e&&e instanceof Function&&n.e(453).then(n.bind(n,453)).then((t=>{let{getCLS:n,getFID:r,getFCP:a,getLCP:o,getTTFB:l}=t;n(e),r(e),a(e),o(e),l(e)}))};o.createRoot(document.getElementById("root")).render((0,Gn.jsx)(r.StrictMode,{children:(0,Gn.jsx)(ol,{})})),ll()})()})();
+//# sourceMappingURL=main.515ed76f.js.map
